% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_anno_finemap_res.R
\name{add_pos}
\alias{add_pos}
\title{Add SNP and gene positions to cTWAS finemapping result.}
\usage{
add_pos(
  finemap_res,
  snp_map,
  mapping_table,
  map_by = "molecular_id",
  use_gene_pos = c("mid", "start", "end")
)
}
\arguments{
\item{finemap_res}{a data frame of cTWAS finemapping result}

\item{snp_map}{a list of data frames with SNP-to-region map for the reference.}

\item{mapping_table}{a data frame of gene annotations, with required columns:
"gene_name", "chrom", "pos" (or "start" and "end").}

\item{map_by}{column name to be mapped by (default: "gene_name").}

\item{use_gene_pos}{Use mid (midpoint), start, or end positions as gene positions.}
}
\value{
a data frame with chromosomes and positions added to fine-mapping result
}
\description{
Add SNP and gene positions to cTWAS finemapping result.
}
