% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_summarize_finemap_res.R
\name{add_pos_to_finemap_res}
\alias{add_pos_to_finemap_res}
\title{Add SNP and gene positions to cTWAS finemapping result.}
\usage{
add_pos_to_finemap_res(
  finemap_res,
  snp_map,
  gene_annot,
  use_gene_pos = c("mid", "start", "end"),
  map_by = "gene_name"
)
}
\arguments{
\item{finemap_res}{a data frame of cTWAS finemapping result}

\item{snp_map}{a list of data frames with SNP-to-region map for the reference.}

\item{gene_annot}{a data frame of gene annotations, with required columns:
"gene_name", "chrom", "pos" (or "start" and "end").}

\item{use_gene_pos}{if "pos" is not available in gene_annot,
use mid (midpoint), start or end positions to
represent gene positions.}

\item{map_by}{column name to be mapped by (default: "gene_name").}
}
\value{
a data frame with chromosomes and positions added to fine-mapping result
}
\description{
Add SNP and gene positions to cTWAS finemapping result.
}
