% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_assemble_susie_input.R
\name{add_z_to_regionlist}
\alias{add_z_to_regionlist}
\title{add z-scores from z_snp and z_gene to regionlist}
\usage{
add_z_to_regionlist(regionlist, z_snp, z_gene, ncore = 1)
}
\description{
add z-scores from z_snp and z_gene to regionlist
}
