% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_region_data.R
\name{adjust_boundary_genes_old}
\alias{adjust_boundary_genes_old}
\title{adjust region_data for boundary genes}
\usage{
adjust_boundary_genes_old(boundary_genes, region_info, weights, region_data)
}
\description{
adjust region_data for boundary genes
}
