% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_summarize_finemap_res.R
\name{anno_finemap_res}
\alias{anno_finemap_res}
\title{Map finemapping result of molecular traits to genes.}
\usage{
anno_finemap_res(
  finemap_res,
  mapping_table = NULL,
  map_by = "gene_id",
  drop_unmapped_genes = TRUE
)
}
\arguments{
\item{finemap_res}{a data frame of cTWAS finemapping results.}

\item{mapping_table}{a data frame of mapping between molecular traits and genes,
with required columns: "gene_id", "gene_name".}

\item{map_by}{column name to be mapped by (default: "gene_id").}

\item{drop_unmapped_genes}{If TRUE, remove unmapped genes.}
}
\value{
a data frame of annotated finemapping result.
}
\description{
Map finemapping result of molecular traits to genes.
}
