% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_susie_rss.R
\name{anno_susie}
\alias{anno_susie}
\title{annotate susie results with SNP and gene information}
\usage{
anno_susie(susie_res, geneinfo, snpinfo, gid, sid, zdf, region_tag)
}
\description{
annotate susie results with SNP and gene information
}
