% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_anno_finemap_res.R
\name{anno_susie_alpha_res}
\alias{anno_susie_alpha_res}
\title{Map molecular traits to genes in susie alpha result.}
\usage{
anno_susie_alpha_res(
  susie_alpha_res,
  mapping_table,
  map_by = "molecular_id",
  drop_unmapped = TRUE
)
}
\arguments{
\item{susie_alpha_res}{a data frame of susie alpha result.}

\item{mapping_table}{a data frame of mapping between molecular traits and genes,
with required columns: "molecular_id", "gene_name".}

\item{map_by}{column name to be mapped by (default: "molecular_id").}

\item{drop_unmapped}{If TRUE, remove unmapped genes.}

\item{snp_map}{a list of data frames with SNP-to-region map for the reference.}

\item{add_gene_annot}{If TRUE, add annotations}

\item{add_position}{If TRUE, add positions}

\item{use_gene_pos}{Use mid (midpoint), start, or end positions as gene positions.}
}
\value{
a data frame of annotated finemapping result.
}
\description{
Map molecular traits to genes in susie alpha result.
}
