% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_assemble_susie_input.R
\name{assemble_region_susie_input}
\alias{assemble_region_susie_input}
\title{assemble susie input data for a region}
\usage{
assemble_region_susie_input(
  sid,
  gid,
  zdf,
  pi_prior,
  V_prior,
  L,
  use_null_weight = TRUE
)
}
\description{
assemble susie input data for a region
}
