% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_assemble_susie_input.R
\name{assemble_susie_input_list}
\alias{assemble_susie_input_list}
\title{assemble susie input data for all regions}
\usage{
assemble_susie_input_list(
  zdf,
  regionlist,
  L,
  group_prior = NULL,
  group_prior_var = NULL,
  use_null_weight = TRUE,
  ncore = 1
)
}
\description{
assemble susie input data for all regions
}
