% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_combine_pips.R
\name{combine_gene_pips}
\alias{combine_gene_pips}
\title{Combines gene PIPs by context, type or group.}
\usage{
combine_gene_pips(
  susie_alpha_res,
  mapping_table = NULL,
  map_by = "molecular_id",
  drop_unmapped = TRUE,
  group_by = "molecular_id",
  by = c("context", "type", "group"),
  method = c("combine_cs", "sum"),
  filter_cs = FALSE,
  keep_alpha_in_cs_only = FALSE,
  include_cs_id = TRUE,
  include_set_id = FALSE,
  missing_value = NA
)
}
\arguments{
\item{susie_alpha_res}{a data frame of annotated susie alpha result.}

\item{mapping_table}{a data frame of mapping between molecular traits and genes,
with required columns: "molecular_id", "gene_name".}

\item{map_by}{column name to be mapped by (default: "molecular_id").}

\item{drop_unmapped}{If TRUE, remove unmapped genes.}

\item{group_by}{column name to group genes by.}

\item{by}{option to combine PIPs by: "context" (default), "type", or "group".}

\item{method}{method to combine PIPs of molecular traits targeting the same gene.
options:
"combine_cs" (default):
first sums PIPs of molecular traits of a genes in each credible set,
and then combine PIPs using the following formula:
\eqn{1 - \prod_k (1 - \text{PIP}_k)},
where \eqn{\text{PIP}_k} is the summed PIP of the \eqn{k}-th credible set of a gene.
This is the default option for combining PIPs from fine-mapping with LD.
"sum": sum over PIPs of all molecular traits for the same gene.
This summation is the expected number of causal molecular traits in this gene,
and could be higher than 1.
We will use this option for combining PIPs from fine-mapping without LD.}

\item{filter_cs}{If TRUE, limits gene results to credible sets (CS).}

\item{keep_alpha_in_cs_only}{If TRUE, only keep single effects (alpha) in credible sets
when calculating combined PIP. This is similar to \code{prune_by_cs} in susie.}

\item{include_cs_id}{If TRUE, include credible set IDs of the genes in the output}

\item{include_set_id}{If TRUE, include susie set IDs of the genes in the output}

\item{missing_value}{set missing value as (default: NA)}
}
\value{
a data frame of combined gene PIPs for each context, type or group
}
\description{
Combines gene PIPs by context, type or group.
}
