% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_compute_gene_z.R
\name{combine_z}
\alias{combine_z}
\title{Combine SNP and gene z-scores}
\usage{
combine_z(z_snp, z_gene)
}
\arguments{
\item{z_snp}{a data frame with four columns: "id", "A1", "A2", "z".
giving the z scores for SNPs. "A1" is effect allele. "A2" is the other allele.}

\item{z_gene}{a data frame with two columns: "id", "z". giving the z scores for genes.
Optionally, a "type" column can also be supplied; this is for using multiple sets of weights}
}
\description{
Combine SNP and gene z-scores
}
