% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_enrichment_se.R
\name{compute_enrichment_se}
\alias{compute_enrichment_se}
\title{Compute standard error and p-value for enrichment}
\usage{
compute_enrichment_se(
  region_data,
  group_prior,
  group_prior_null,
  group_prior_var
)
}
\arguments{
\item{region_data}{a list of assembled region data.}

\item{group_prior}{a vector of prior inclusion probabilities for different groups.}

\item{group_prior_null}{a vector of prior inclusion probabilities for different groups under the null.}

\item{group_prior_var}{a vector of prior variances for different groups.}
}
\value{
a list of estimated enrichment, S.E. and p-value.
}
\description{
Compute standard error and p-value for enrichment
}
