% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_diagnose_LD_mismatch.R
\name{compute_region_condz}
\alias{compute_region_condz}
\title{Compute expected z-scores based on conditional distribution of z-scores using SuSiE RSS}
\usage{
compute_region_condz(region_id, LD_info, snp_info, z_snp, gwas_n)
}
\description{
Compute expected z-scores based on conditional distribution of z-scores using SuSiE RSS
}
