% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_compute_region_cor.R
\name{compute_region_cor}
\alias{compute_region_cor}
\title{Compute correlation matrices for a region}
\usage{
compute_region_cor(sids, gids, R_snp, weights, snpinfo)
}
\description{
Compute correlation matrices for a region
}
