% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_process_regions.R
\name{compute_region_nonSNP_PIPs}
\alias{compute_region_nonSNP_PIPs}
\title{Select regions with high non-SNP PIPs}
\usage{
compute_region_nonSNP_PIPs(finemap_res)
}
\description{
Select regions with high non-SNP PIPs
}
