% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_est_parameters.R
\name{compute_region_p_single_effect}
\alias{compute_region_p_single_effect}
\title{Select single effect regions}
\usage{
compute_region_p_single_effect(region_data, group_prior)
}
\description{
Select single effect regions
}
