% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_weights.R
\name{compute_weight_LD_from_ref}
\alias{compute_weight_LD_from_ref}
\title{Computes LD for weight variants using reference LD}
\usage{
compute_weight_LD_from_ref(
  weights,
  region_info,
  LD_map,
  snp_map = NULL,
  LD_format = c("rds", "rdata", "mtx", "csv", "txt", "custom"),
  LD_loader_fun = NULL,
  snpinfo_loader_fun = NULL,
  ncore = 1
)
}
\arguments{
\item{weights}{a list of preprocessed weights.}

\item{region_info}{a data frame of region definitions.}

\item{LD_map}{a data frame with filenames of LD matrices and SNP information for the regions.
Required when \code{load_predictdb_LD = FALSE}.}

\item{snp_map}{a list of SNP-to-region map for the reference.
If NUll, it will reads SNP info from the "SNP_file" column of LD_map.}

\item{LD_format}{file format for LD matrix. If "custom", use a user defined
\code{LD_loader_fun()} function to load LD matrix.}

\item{LD_loader_fun}{a user defined function to load LD matrix when \code{LD_format = "custom"}.}

\item{snpinfo_loader_fun}{a user defined function to load SNP information file,
if SNP information files are not in standard cTWAS reference format.}

\item{ncore}{The number of cores used to parallelize computation.}
}
\value{
a list of processed weights, with LD of weight variants included.
}
\description{
Computes LD for weight variants using reference LD
}
