% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_weights.R
\name{convert_fusion_to_predictdb}
\alias{convert_fusion_to_predictdb}
\title{Converts fusion weights to predictDB format}
\usage{
convert_fusion_to_predictdb(
  weight_path,
  fusion_method = c("lasso", "enet", "top1", "blup", "bslmm", "bestR2"),
  fusion_genome_version = c("b38", "b37"),
  make_extra_table = TRUE,
  covar_table = NULL,
  outputdir = getwd(),
  outname
)
}
\arguments{
\item{weight_path}{the directory containing FUSION weights ('.wgt.RDat' files).}

\item{fusion_method}{a string, specifying the method to choose in FUSION models.}

\item{fusion_genome_version}{a string, specifying the genome version of FUSION models}

\item{make_extra_table}{TRUE/FALSE. If TRUE, make an extra table in predictDB format}

\item{covar_table}{a data frame of covariances between variants, with columns:
"GENE","RSID1","RSID2","VALUE".
If NULL, do not create covariance files (.txg.gz).}

\item{outputdir}{output directory}

\item{outname}{name of the output weight file}
}
\description{
Converts fusion weights to predictDB format
}
