% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_convert_geno_to_LD_matrix.R
\name{convert_geno_to_LD_matrix}
\alias{convert_geno_to_LD_matrix}
\title{Convert PLINK genotype data to LD matrices and SNP info files,
save LD matrices as .RDS files and SNP info as .Rvar files}
\usage{
convert_geno_to_LD_matrix(
  region_info,
  genotype_files,
  snpinfo_files,
  chrom = 1:22,
  outputdir = getwd(),
  outname = "LD",
  print_variance = TRUE,
  print_allele_freq = TRUE
)
}
\arguments{
\item{region_info}{a data frame of region information, with columns: chr, start, stop positions}

\item{genotype_files}{Reference genotype files in PLINK binary genotype format (BED or PGEN format)}

\item{snpinfo_files}{Reference variant information files for the genome version in PLINK BIM format}

\item{chrom}{a vector of chromosomes to run}

\item{outputdir}{Output directory}

\item{outname}{Output filestem}

\item{print_variance}{TRUE/FALSE, if TRUE, print variance}

\item{print_allele_freq}{TRUE/FALSE, if TRUE, print allele frequency}
}
\description{
Convert PLINK genotype data to LD matrices and SNP info files,
save LD matrices as .RDS files and SNP info as .Rvar files
}
