% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_weights.R
\name{convert_predictdb_weights_varIDs}
\alias{convert_predictdb_weights_varIDs}
\title{Convert rsIDs and varIDs in predictDB weights to the reference variant format.}
\usage{
convert_predictdb_weights_varIDs(
  weight_file,
  load_predictdb_LD = TRUE,
  varID_converter_fun = NULL,
  rsID_converter_fun = NULL,
  outputdir = getwd(),
  outname
)
}
\arguments{
\item{weight_file}{filename of the PredictDB weights '.db' file.}

\item{load_predictdb_LD}{If TRUE, load pre-computed LD among weight SNPs.}

\item{varID_converter_fun}{a user defined function to convert
varIDs in weights to the reference variant format. If NULL, do not convert varIDs.}

\item{rsID_converter_fun}{a user defined function to convert
rsIDs in weights to the reference variant format. If NULL, do not convert rsIDs.}

\item{outputdir}{output directory.}

\item{outname}{name of the output weight file.}
}
\description{
Convert rsIDs and varIDs in predictDB weights to the reference variant format.
}
