% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_utils.R
\name{count_n_snps}
\alias{count_n_snps}
\title{Count the numbers of SNPs in snp_map, z_snp and weights}
\usage{
count_n_snps(snp_map, z_snp, weights)
}
\arguments{
\item{snp_map}{a list of data frames with SNP-to-region map for the reference.}

\item{z_snp}{A data frame with columns: "id", "z", giving the z-scores for SNPs.}

\item{weights}{a list of pre-processed prediction weights.}
}
\value{
a list of numbers of SNPs in snp_map, z_snp and weights
}
\description{
Count the numbers of SNPs in snp_map, z_snp and weights
}
