% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_weights.R
\name{count_z_snp_weights}
\alias{count_z_snp_weights}
\title{Counts the number of SNPs in z_snp and weights}
\usage{
count_z_snp_weights(z_snp, weights)
}
\arguments{
\item{z_snp}{A data frame with columns: "id", "z", giving the z-scores for SNPs.}

\item{weights}{a list of preprocessed weights.}
}
\value{
number of SNPs in z_snp and weights
}
\description{
Counts the number of SNPs in z_snp and weights
}
