% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_snp_LD_map.R
\name{create_snp_map}
\alias{create_snp_map}
\title{Map SNPs to regions using all the variants in the LD reference.}
\usage{
create_snp_map(region_info, ref_snp_info, ncore = 1)
}
\arguments{
\item{region_info}{a data frame of region definitions, with columns:
"chrom", "start", "stop", "region_id".}

\item{ref_snp_info}{a data frame of all variant info in the reference.}

\item{ncore}{The number of cores used to parallelize over regions}
}
\description{
Map SNPs to regions using all the variants in the LD reference.
}
