% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_EM.R
\name{ctwas_EM}
\alias{ctwas_EM}
\title{Run EM to estimate parameters.
Iteratively run susie and estimate parameters - RSS version}
\usage{
ctwas_EM(
  zdf,
  regionlist,
  region_info,
  gene_info,
  niter = 20,
  group_prior = NULL,
  group_prior_var = NULL,
  group_prior_var_structure = c("independent", "shared_all", "shared_QTLtype"),
  use_null_weight = TRUE,
  coverage = 0.95,
  min_abs_corr = 0.5,
  max_iter = 1,
  ncore = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{zdf}{A data frame with three columns: "id", "z", "type".
This data frame gives the the z scores for SNPs and genes, denoted in "type".
The "type" column can also be use to specify multiple sets of weights}

\item{regionlist}{a list object indexing regions, variants and genes.}

\item{region_info}{a data frame of region definition and associated file names}

\item{gene_info}{a data frame of gene information}

\item{niter}{the number of iterations of the E-M algorithm to perform}

\item{group_prior}{a vector of two prior inclusion probabilities for SNPs and genes.}

\item{group_prior_var}{a vector of two prior variances for SNPs and gene effects.}

\item{group_prior_var_structure}{a string indicating the structure to put on the prior variance parameters.
"independent" is the default and allows all groups to have their own separate variance parameters.
"shared_all" allows all groups to share the same variance parameter.
"shared_type" allows all groups in one molecular QTL type to share the same variance parameter.}

\item{use_null_weight}{TRUE/FALSE. If TRUE, allow for a probability of no effect in susie}

\item{coverage}{A number between 0 and 1 specifying the \dQuote{coverage} of the estimated confidence sets}

\item{min_abs_corr}{Minimum absolute correlation allowed in a
credible set. The default, 0.5, corresponds to a squared
correlation of 0.25, which is a commonly used threshold for
genotype data in genetic studies.}

\item{max_iter}{Maximum number of IBSS iterations to perform.}

\item{ncore}{The number of cores used to parallelize susie over regions}
}
\value{
a list of parameters
}
\description{
Run EM to estimate parameters.
Iteratively run susie and estimate parameters - RSS version
}
