% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_plots.R
\name{ctwas_locus_plot}
\alias{ctwas_locus_plot}
\title{Plot the output of cTWAS at a single locus}
\usage{
ctwas_locus_plot(
  ctwas_res,
  region_tag,
  outputdir,
  outname,
  focus = NULL,
  xlim = NULL,
  ymax_twas = NULL,
  twas_sig_thresh = NULL,
  alt_names = NULL,
  label_panel = c("both", "cTWAS", "TWAS"),
  label_genes = NULL,
  label_pos = NULL,
  plot_eqtl = NULL,
  legend_panel = c("cTWAS", "TWAS"),
  legend_side = c("right", "left"),
  rerun_ctwas = F,
  rerun_load_only = F,
  return_table = F,
  show_correlations = T,
  z_snp = NULL,
  z_gene = NULL,
  empty_gene_track = F
)
}
\arguments{
\item{ctwas_res}{the main output of \code{ctwas_rss} as a data frame, with z-scores included}

\item{region_tag}{a character with format "regiontag1_regiontag2" indicating the reigon to plot; e.g. "16_1"}

\item{outputdir}{a string, the directory where ctwas output is stored}

\item{outname}{a string, the output name}

\item{focus}{the focal gene; points are colored by correlations with this gene. The lead TWAS gene by default}

\item{xlim}{a vector with the start and end positions to plot, in megabases}

\item{ymax_twas}{the ymax parameter of the TWAS panel}

\item{twas_sig_thresh}{the significance threshold for TWAS as a p-value}

\item{alt_names}{the name of the column that contains alternative gene names in \code{ctwas_res}; null by default}

\item{label_panel}{the panel(s) to plot gene gene names; both cTWAS and TWAS panels by default}

\item{label_genes}{the gene names to label; all by default}

\item{label_pos}{a vector of the same length as \code{label_genes} indicating the position of the gene labels; 1=below, 2=left, 3=above, 4=right. Above by default.}

\item{plot_eqtl}{the genes to plot eQTL tracks for; the gene specified by \code{focus} by default}

\item{legend_panel}{the panel to plot the legend in}

\item{legend_side}{the side to plot the legend on}

\item{rerun_ctwas}{if TRUE, cTWAS will be re-run at this locus with \code{thin=1} to ensure all variants are plotted}

\item{rerun_load_only}{if TRUE, the function will look for results generated by \code{rerun_ctwas=T} without running cTWAS again}

\item{return_table}{if TRUE, return the table that underlies the plot}

\item{show_correlations}{if TRUE, genes and variants are colored by their correlations with \code{focus}}

\item{z_snp}{the \code{z_snp} argument supplied to \code{ctwas_rss}; required if \code{rerun_ctwas=T}}

\item{z_gene}{the \code{z_gene} argument supplied to \code{ctwas_rss}; required if \code{rerun_ctwas=T}}

\item{empty_gene_track}{if TRUE, leave space below the plot for the gene track generated by the \code{ctwas_locus_plot_genetrack} function}
}
\description{
Plot the output of cTWAS at a single locus
}
