% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_plots.R
\name{ctwas_locus_plot_genetrack}
\alias{ctwas_locus_plot_genetrack}
\title{Create the gene track for the cTWAS locus plot}
\usage{
ctwas_locus_plot_genetrack(plot_df, label_pos = NULL)
}
\arguments{
\item{plot_df}{the output of \code{ctwas_locus_plot}}

\item{label_pos}{the position of the gene labels. Default is NULL, which places gene names to the left of the gene; can also specify "above"}
}
\description{
Create the gene track for the cTWAS locus plot
}
