% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_plots.R
\name{ctwas_locusplot}
\alias{ctwas_locusplot}
\title{Plot the output of cTWAS at a single locus}
\usage{
ctwas_locusplot(
  finemap_res,
  screened_region_data,
  region_id,
  cor_matrix_path = getwd(),
  locus_range = NULL,
  focus_gene = NULL,
  gene_biotype = "protein_coding",
  genome_build = "hg38"
)
}
\description{
Plot the output of cTWAS at a single locus
}
