% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_plots.R
\name{ctwas_summarize_parameters}
\alias{ctwas_summarize_parameters}
\title{Summarize and plot cTWAS parameter estimates}
\usage{
ctwas_summarize_parameters(
  outputdir,
  outname,
  gwas_n = NA,
  thin = 1,
  plot_estimates = T
)
}
\arguments{
\item{outputdir}{a string, the directory where ctwas output is stored}

\item{outname}{a string, the output name}

\item{gwas_n}{the sample size of the GWAS summary statistics}

\item{thin}{the proportion of SNPs used for parameter estimation, as supplied to \code{ctwas_rss}}

\item{plot_estimates}{if TRUE, return a plot of the estimated parameters}
}
\description{
Summarize and plot cTWAS parameter estimates
}
