% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_sumstats.R
\name{ctwas_sumstats}
\alias{ctwas_sumstats}
\title{cTWAS analysis using summary statistics}
\usage{
ctwas_sumstats(
  z_snp,
  weights,
  region_info,
  LD_map,
  snp_map,
  z_gene = NULL,
  thin = 1,
  niter_prefit = 3,
  niter = 50,
  L = 5,
  init_group_prior = NULL,
  init_group_prior_var = NULL,
  group_prior_var_structure = c("shared_all", "shared_type", "shared_context",
    "shared_nonSNP", "independent"),
  min_nonSNP_PIP = 0.5,
  min_snp_pval = 5e-08,
  min_gene_pval = min_snp_pval,
  min_p_single_effect = 0.8,
  maxSNP = Inf,
  min_var = 2,
  min_gene = 1,
  min_group_size = 100,
  null_method = c("ctwas", "susie", "none"),
  EM_tol = 1e-04,
  coverage = 0.95,
  min_abs_corr = 0.1,
  include_prior = FALSE,
  include_susie_result = FALSE,
  LD_format = c("rds", "rdata", "mtx", "csv", "txt", "custom"),
  LD_loader_fun = NULL,
  snpinfo_loader_fun = NULL,
  force_compute_cor = FALSE,
  save_cor = FALSE,
  cor_dir = NULL,
  outputdir = NULL,
  outname = "ctwas",
  ncore = 1,
  ncore_LD = max(ncore - 1, 1),
  seed = 99,
  logfile = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{z_snp}{A data frame with four columns: "id", "A1", "A2", "z".
giving the z scores for SNPs. "A1" is effect allele. "A2" is the other allele.}

\item{weights}{a list of pre-processed prediction weights.}

\item{region_info}{a data frame of region definitions.}

\item{LD_map}{a data frame with filenames of LD matrices and SNP information for the regions.}

\item{snp_map}{a list of data frames with SNP-to-region map for the reference.}

\item{z_gene}{A data frame with columns: "id", "z", giving the z-scores for genes.}

\item{thin}{The proportion of SNPs to be used for estimating parameters and screening regions.}

\item{niter_prefit}{the number of iterations of the E-M algorithm to perform during the initial parameter estimation step.}

\item{niter}{the maximum number of iterations of the E-M algorithm to perform during the complete parameter estimation step.}

\item{L}{the number of effects for susie during the fine mapping steps.}

\item{init_group_prior}{a vector of initial values of prior inclusion probabilities for different groups.}

\item{init_group_prior_var}{a vector of initial values of prior variances for different groups.}

\item{group_prior_var_structure}{a string indicating the structure to put on the prior variance parameters.
"shared_all" allows all groups to share the same variance parameter.
"shared_type" allows all groups in one molecular QTL type to share the same variance parameter.
"shared_context" allows all groups in one context (tissue, cell type, condition) to share the same variance parameter.
"shared_nonSNP" allows all non-SNP groups to share the same variance parameter.
"independent" allows all groups to have their own separate variance parameters.
"fixed" sets prior variance parameters to values in \code{init_group_prior_var}.}

\item{min_nonSNP_PIP}{Regions with non-SNP PIP >= \code{min_nonSNP_PIP}
will be selected to run finemapping using all SNPs.}

\item{min_snp_pval}{Select regions with minimum SNP p-values < \code{min_snp_pval}.}

\item{min_gene_pval}{Select regions with minimum gene p-values < \code{min_gene_pval}.
By default, it is set to the same value as \code{min_snp_pval}.}

\item{min_p_single_effect}{Regions with probability greater than \code{min_p_single_effect} of
having 1 or fewer effects will be used for parameter estimation.}

\item{maxSNP}{Inf or integer. Maximum number of SNPs in a region. Default is
Inf, no limit. This can be useful if there are many SNPs in a region and you don't
have enough memory to run the program.}

\item{min_var}{minimum number of variables (SNPs and genes) in a region
when estimating paramters and screening regions.}

\item{min_gene}{minimum number of genes in a region
when estimating paramters and screening regions.}

\item{min_group_size}{Minimum number of genes in a group.
Groups with number of genes < \code{min_group_size} will be removed for the analysis.}

\item{null_method}{Method to compute null model, options: "ctwas", "susie" or "none".}

\item{EM_tol}{A small, non-negative number specifying the convergence
tolerance of log-likelihood for the EM iterations.}

\item{coverage}{A number between 0 and 1 specifying the \dQuote{coverage} of
the estimated confidence sets.}

\item{min_abs_corr}{Minimum absolute correlation allowed in a credible set.}

\item{include_prior}{If TRUE, include priors in finemapping results.}

\item{include_susie_result}{If TRUE, include the "susie" result object in finemapping results.}

\item{LD_format}{file format for LD matrix. If "custom", use a user defined
\code{LD_loader_fun()} function to load LD matrix.}

\item{LD_loader_fun}{a user defined function to load LD matrix when \code{LD_format = "custom"}.}

\item{snpinfo_loader_fun}{a user defined function to load SNP information file,
if SNP information files are not in standard cTWAS reference format.}

\item{force_compute_cor}{If TRUE, force computing correlation (R) matrices.}

\item{save_cor}{If TRUE, save correlation (R) matrices to \code{cor_dir}.}

\item{cor_dir}{The directory to store correlation (R) matrices.}

\item{outputdir}{The directory to store output. If specified, save outputs to the directory.}

\item{outname}{The output name.}

\item{ncore}{The number of cores used to parallelize computing over regions.}

\item{ncore_LD}{The number of cores used to parallelize computing correlation matrices,
in screening regions and fine-mapping steps with LD.}

\item{seed}{seed for random sampling when thinning the SNPs in region data.}

\item{logfile}{The log filename. If NULL, print log info on screen.}

\item{verbose}{If TRUE, print detailed messages.}

\item{...}{Additional arguments of \code{susie_rss}.}
}
\value{
a list, including z_gene, estimated parameters, region_data,
screening region results, and fine-mapping results.
}
\description{
cTWAS analysis using summary statistics
}
