% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_finemapping.R
\name{ctwas_susie_rss}
\alias{ctwas_susie_rss}
\title{Run cTWAS version of susie_rss for a single region}
\usage{
ctwas_susie_rss(
  z,
  R,
  prior_weights = NULL,
  prior_variance = NULL,
  L = 5,
  z_ld_weight = 0,
  null_weight = NULL,
  coverage = 0.95,
  min_abs_corr = 0.5,
  max_iter = 100,
  ...
)
}
\description{
Run cTWAS version of susie_rss for a single region
}
