% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_LD_mismatch.R
\name{diagnose_LD_mismatch_susie}
\alias{diagnose_LD_mismatch_susie}
\title{Diagnose LD mismatch using SuSiE RSS}
\usage{
diagnose_LD_mismatch_susie(
  region_ids,
  z_snp,
  LD_map,
  gwas_n,
  p_diff_thresh = 5e-06,
  plot = TRUE,
  LD_format = c("rds", "rdata", "mtx", "csv", "txt", "custom"),
  LD_loader_fun = NULL,
  snpinfo_loader_fun = NULL,
  ncore = 1,
  logfile = NULL
)
}
\arguments{
\item{region_ids}{A vector of region IDs to run diagnosis}

\item{z_snp}{A data frame with two columns: "id", "A1", "A2", "z". giving the z scores for
snps. "A1" is effect allele. "A2" is the other allele.}

\item{LD_map}{a data frame with filenames of LD matrices and SNP information for each of the regions.}

\item{gwas_n}{integer, GWAS sample size.}

\item{p_diff_thresh}{numeric, p-value cutoff for identifying problematic SNPs
with significant difference between observed z-scores and estimated values.}

\item{plot}{If TRUE, plot observed z score vs the expected value.}

\item{LD_format}{file format for LD matrix. If "custom", use a user defined
\code{LD_loader_fun()} function to load LD matrix.}

\item{LD_loader_fun}{a user defined function to load LD matrix when \code{LD_format = "custom"}.}

\item{snpinfo_loader_fun}{a user defined function to load SNP information file,
if SNP information files are not in standard cTWAS reference format.}

\item{ncore}{integer, number of cores for parallel computing.}

\item{logfile}{the log file, if NULL will print log info on screen}
}
\value{
a list of problematic SNPs, flipped SNPs,
and test statistics from susie's `kriging_rss` function
}
\description{
Diagnose LD mismatch using SuSiE RSS
}
