% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_EM.R
\name{EM_est_param}
\alias{EM_est_param}
\title{Run EM to estimate parameters.
Iteratively run susie and estimate parameters - RSS version}
\usage{
EM_est_param(
  region_data,
  niter = 20,
  init_group_prior = NULL,
  init_group_prior_var = NULL,
  group_prior_var_structure = c("independent", "shared_nonSNP", "shared_all",
    "shared_type", "shared_context"),
  use_null_weight = TRUE,
  max_IBSS_iter = 1,
  minvar = 2,
  mingene = 1,
  ncore = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{region_data}{a list object with the susie input data for each region}

\item{niter}{the number of iterations of the E-M algorithm to perform}

\item{init_group_prior}{a vector of initial prior inclusion probabilities for SNPs and genes.}

\item{init_group_prior_var}{a vector of initial prior variances for SNPs and gene effects.}

\item{group_prior_var_structure}{a string indicating the structure to put on the prior variance parameters.
"independent" is the default and allows all groups to have their own separate variance parameters.
"shared_nonSNP" allows all non-SNP groups to share the same variance parameter.
"shared_all" allows all groups to share the same variance parameter.
"shared_type" allows all groups in one molecular QTL type to share the same variance parameter.
"shared_context" allows all groups in one molecular QTL type to share the same variance parameter.}

\item{use_null_weight}{TRUE/FALSE. If TRUE, allow for a probability of no effect in susie}

\item{max_IBSS_iter}{Maximum number of IBSS iterations to perform.}

\item{minvar}{minimum number of variables (snps and genes) in a region}

\item{mingene}{minimum number of genes in a region}

\item{ncore}{The number of cores used to parallelize susie over regions}

\item{verbose}{TRUE/FALSE. If TRUE, print detail messages}
}
\value{
a list of parameters
}
\description{
Run EM to estimate parameters.
Iteratively run susie and estimate parameters - RSS version
}
