% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_est_parameters_ser.R
\name{est_param_ser}
\alias{est_param_ser}
\title{Estimates cTWAS parameters using EM with cTWAS SER model}
\usage{
est_param_ser(
  region_data,
  init_group_prior = NULL,
  init_group_prior_var = NULL,
  group_prior_var_structure = c("shared_type", "shared_context", "shared_nonSNP",
    "shared_all", "independent"),
  niter_prefit = 3,
  niter = 30,
  min_p_single_effect = 0.8,
  use_null_weight = TRUE,
  null_weight_method = c("susie", "ctwas"),
  min_var = 2,
  min_gene = 1,
  min_group_size = 100,
  ncore = 1,
  logfile = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{region_data}{a list object indexing regions, variants and genes.}

\item{init_group_prior}{a vector of initial values of prior inclusion probabilities for SNPs and genes.}

\item{init_group_prior_var}{a vector of initial values of prior variances for SNPs and gene effects.}

\item{group_prior_var_structure}{a string indicating the structure to put on the prior variance parameters.
"shared_type" allows all groups in one molecular QTL type to share the same variance parameter.
"shared_context" allows all groups in one context (tissue, cell type, condition) to share the same variance parameter.
"shared_nonSNP" allows all non-SNP groups to share the same variance parameter.
"shared_all" allows all groups to share the same variance parameter.
"independent" allows all groups to have their own separate variance parameters.}

\item{niter_prefit}{the number of iterations of the E-M algorithm to perform during the initial parameter estimation step.}

\item{niter}{the number of iterations of the E-M algorithm to perform during the complete parameter estimation step.}

\item{min_p_single_effect}{Regions with probability greater than \code{min_p_single_effect} of
having at most one causal effect will be used selected for the complete parameter estimation step.}

\item{use_null_weight}{If TRUE, allow for a probability of no effect}

\item{null_weight_method}{Method to compute null weight, options: "susie" or "ctwas".}

\item{min_var}{minimum number of variables (SNPs and genes) in a region.}

\item{min_gene}{minimum number of genes in a region.}

\item{min_group_size}{Minimum number of variables in a group.}

\item{ncore}{The number of cores used to parallelize computation over regions.}

\item{logfile}{The log filename. If NULL, print log info on screen.}

\item{verbose}{If TRUE, print detail messages.}

\item{...}{Additional arguments of \code{susie_rss}.}
}
\value{
a list with estimated parameters
}
\description{
Estimates cTWAS parameters using EM with cTWAS SER model
}
