% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_screen_regions.R
\name{estimate_region_L}
\alias{estimate_region_L}
\title{Estimate L for each region by running finemapping with uniform prior}
\usage{
estimate_region_L(
  region_data,
  LD_map,
  snp_map,
  weights,
  LD_format,
  LD_loader_fun,
  ncore,
  verbose,
  ...
)
}
\arguments{
\item{region_data}{a list object indexing regions, variants and genes.}

\item{LD_map}{a data frame with filenames of LD matrices for each of the regions. Required when \code{use_LD = TRUE}.}

\item{snp_map}{a list of data frames with SNP-to-region map for the reference. Required when \code{use_LD = TRUE}.}

\item{weights}{a list of weights for each gene. Required when \code{use_LD = TRUE}.}

\item{LD_format}{file format for LD matrix. If "custom", use a user defined
\code{LD_loader_fun()} function to load LD matrix.}

\item{LD_loader_fun}{a user defined function to load LD matrix when \code{LD_format = "custom"}.}

\item{ncore}{The number of cores used to parallelize susie over regions}

\item{verbose}{TRUE/FALSE. If TRUE, print detail messages}

\item{...}{Additional arguments of \code{susie_rss}.}

\item{use_LD}{TRUE/FALSE. If TRUE, use LD for finemapping. Otherwise, use "no-LD" version.}
}
\value{
estimated L for each region
}
\description{
Estimate L for each region by running finemapping with uniform prior
}
