% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_region_data.R
\name{expand_region_data}
\alias{expand_region_data}
\title{Expands region_data with full SNPs}
\usage{
expand_region_data(
  region_data,
  snp_map,
  z_snp,
  z_gene,
  trim_by = c("z", "random"),
  maxSNP = Inf,
  add_z = TRUE,
  ncore = 1,
  seed = 99
)
}
\arguments{
\item{region_data}{a list of region gene IDs and SNP IDs and associated file names}

\item{snp_map}{a list of data frames with SNP-to-region map for the reference.}

\item{z_snp}{A data frame with columns: "id", "z", giving the z-scores for SNPs.}

\item{z_gene}{A data frame with columns: "id", "z", giving the z-scores for genes.}

\item{trim_by}{remove SNPs if the total number of SNPs exceeds limit, options: "random",
or "z" (trim SNPs with lower |z|) See parameter `maxSNP` for more information.}

\item{maxSNP}{Inf or integer. Maximum number of SNPs in a region. Default is
Inf, no limit. This can be useful if there are many SNPs in a region and you don't
have enough memory to run the program.}

\item{add_z}{TRUE/FALSE, if TRUE, add z-scores to the region_data}

\item{ncore}{The number of cores used to parallelize susie over regions}

\item{seed}{seed for random sampling}
}
\value{
updated region_data with full SNPs
}
\description{
Expands region_data with full SNPs
}
