% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_susie_rss.R
\name{extract_susie_alpha}
\alias{extract_susie_alpha}
\title{extract the alpha matrix from susie result
adapted from susie's susie_get_pip() function}
\usage{
extract_susie_alpha(susie_res, prune_by_cs = FALSE)
}
\arguments{
\item{prune_by_cs}{Whether or not to ignore single effects not in
a reported CS when calculating PIP.}
}
\description{
extract the alpha matrix from susie result
adapted from susie's susie_get_pip() function
}
\keyword{internal}
