% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_summarize_finemap_res.R
\name{filter_finemap_res}
\alias{filter_finemap_res}
\title{Filter fine-mapping result}
\usage{
filter_finemap_res(
  annotated_finemap_res,
  filter_protein_coding_genes = TRUE,
  filter_cs_genes = TRUE,
  filter_cs_snps = FALSE
)
}
\arguments{
\item{annotated_finemap_res}{Annotated cTWAS fine-mapping result}

\item{filter_protein_coding_genes}{If TRUE, keep protein coding genes only.}

\item{filter_cs_genes}{If TRUE, limits gene results to credible sets.}

\item{filter_cs_snps}{If TRUE, limits SNP results to credible sets.}
}
\value{
a data frame of filtered cTWAS fine-mapping result
}
\description{
Filter fine-mapping result
}
