% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_process_regions.R
\name{filter_regions}
\alias{filter_regions}
\title{filter regions based on probability of at most 1 causal effect}
\usage{
filter_regions(regionlist, group_prior, prob_single = 0.8, zdf)
}
\description{
filter regions based on probability of at most 1 causal effect
}
