% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_postprocess.R
\name{finemap_regions_L1}
\alias{finemap_regions_L1}
\title{run cTWAS finemapping for multiple regions with L = 1}
\usage{
finemap_regions_L1(
  z_snp,
  z_gene,
  gene_info,
  regionlist,
  region_tags,
  weight_list = NULL,
  group_prior = NULL,
  group_prior_var = NULL,
  use_null_weight = TRUE,
  max_iter = 1,
  logfile = NULL,
  ...
)
}
\arguments{
\item{z_snp}{A data frame with four columns: "id", "A1", "A2", "z".
giving the z scores for snps. "A1" is effect allele. "A2" is the other allele.}

\item{z_gene}{A data frame with two columns: "id", "z". giving the z scores for genes.
Optionally, a "type" column can also be supplied; this is for using multiple sets of weights}

\item{gene_info}{a data frame of gene information obtained from \code{compute_gene_z}}

\item{regionlist}{a list object indexing regions, variants and genes.}

\item{weight_list}{a list of weights for each gene}

\item{group_prior}{a vector of two prior inclusion probabilities for SNPs and genes.}

\item{group_prior_var}{a vector of two prior variances for SNPs and gene effects.}

\item{use_null_weight}{TRUE/FALSE. If TRUE, allow for a probability of no effect in susie}

\item{max_iter}{Maximum number of IBSS iterations to perform.}

\item{region_tag}{a character string of region tag to be finemapped}
}
\value{
finemapping results.
}
\description{
run cTWAS finemapping for multiple regions with L = 1
}
