% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_finemapping.R
\name{finemap_regions_noLD}
\alias{finemap_regions_noLD}
\title{Runs cTWAS fine-mapping for regions without LD (L = 1)}
\usage{
finemap_regions_noLD(
  region_data,
  group_prior = NULL,
  group_prior_var = NULL,
  min_var = 2,
  min_gene = 1,
  null_method = c("ctwas", "susie", "none"),
  coverage = 0.95,
  include_cs = TRUE,
  include_prior = FALSE,
  include_mu2 = FALSE,
  include_susie_alpha = TRUE,
  include_susie_result = FALSE,
  snps_only = FALSE,
  ncore = 1,
  verbose = FALSE,
  logfile = NULL,
  ...
)
}
\arguments{
\item{region_data}{region_data to be finemapped}

\item{group_prior}{a vector of prior inclusion probabilities for different groups.
If NULL, it will use uniform prior inclusion probabilities.}

\item{group_prior_var}{a vector of prior variances for different groups.
If NULL, it will set prior variance = 50 as the default in \code{susie_rss}.}

\item{min_var}{minimum number of variables (SNPs and genes) in a region.}

\item{min_gene}{minimum number of genes in a region.}

\item{null_method}{Method to compute null model, options: "ctwas", "susie" or "none".}

\item{coverage}{A number between 0 and 1 specifying the \dQuote{coverage} of the estimated confidence sets}

\item{include_cs}{If TRUE, include credible sets (CS) to fine-mapping results.}

\item{include_prior}{If TRUE, include priors in fine-mapping results.}

\item{include_mu2}{If TRUE, include estimated effect size variance (mu2) in fine-mapping results.}

\item{include_susie_alpha}{If TRUE, include susie alpha matrix from fine-mapping results.}

\item{include_susie_result}{If TRUE, include the "susie" result object in fine-mapping results.}

\item{snps_only}{If TRUE, use only SNPs in the region data.}

\item{ncore}{The number of cores used to parallelize computation over regions}

\item{verbose}{If TRUE, print detail messages}

\item{logfile}{the log file, if NULL will print log info on screen}

\item{...}{Additional arguments of \code{susie_rss}.}
}
\value{
a list with cTWAS fine-mapping results.
}
\description{
Runs cTWAS fine-mapping for regions without LD (L = 1)
}
