% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_finemapping.R
\name{finemap_regions_ser}
\alias{finemap_regions_ser}
\title{Finemap regions with cTWAS SER model}
\usage{
finemap_regions_ser(
  region_data,
  group_prior = NULL,
  group_prior_var = NULL,
  min_var = 2,
  min_gene = 1,
  null_method = c("ctwas", "susie", "none"),
  null_weight = NULL,
  snps_only = FALSE,
  ncore = 1,
  verbose = FALSE,
  logfile = NULL
)
}
\arguments{
\item{region_data}{a list object with the susie input data for each region.}

\item{group_prior}{a vector of prior inclusion probabilities for SNPs and genes.}

\item{group_prior_var}{a vector of prior variances for SNPs and gene effects.}

\item{min_var}{minimum number of variables (SNPs and genes) in a region.}

\item{min_gene}{minimum number of genes in a region.}

\item{null_method}{Method to compute null model, options: "ctwas", "susie" or "none".}

\item{null_weight}{Prior probability of no effect (a number between
0 and 1, and cannot be exactly 1). Only used when \code{null_method = "susie"}.}

\item{snps_only}{If TRUE, use only SNPs in the region data.}

\item{ncore}{The number of cores used to parallelize computation over regions}

\item{verbose}{If TRUE, print detail messages}

\item{logfile}{the log file, if NULL will print log info on screen}
}
\value{
a data frame of SER result for finemapped regions
}
\description{
Finemap regions with cTWAS SER model
}
