% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_finemapping.R
\name{finemap_regions_ser_rss}
\alias{finemap_regions_ser_rss}
\title{Finemap regions with cTWAS SER model}
\usage{
finemap_regions_ser_rss(
  region_data,
  group_prior = NULL,
  group_prior_var = NULL,
  use_null_weight = TRUE,
  null_weight_method = c("susie", "ctwas"),
  ncore = 1
)
}
\arguments{
\item{region_data}{a list object with the susie input data for each region.}

\item{group_prior}{a vector of prior inclusion probabilities for SNPs and genes.}

\item{group_prior_var}{a vector of prior variances for SNPs and gene effects.}

\item{use_null_weight}{If TRUE, allow for a probability of no effect.}

\item{null_weight_method}{Method to compute null weight, options: "susie" or "ctwas".}

\item{ncore}{The number of cores used to parallelize over regions.}
}
\value{
a data frame of SER result
}
\description{
Finemap regions with cTWAS SER model
}
