% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_finemapping.R
\name{finemap_single_region_noLD}
\alias{finemap_single_region_noLD}
\title{Runs cTWAS finemapping for a single region without LD}
\usage{
finemap_single_region_noLD(
  region_data,
  region_id,
  group_prior = NULL,
  group_prior_var = NULL,
  use_null_weight = TRUE,
  snps_only = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{region_data}{a list object with data for the regions}

\item{region_id}{a character string of region id to be finemapped}

\item{group_prior}{a vector of two prior inclusion probabilities for SNPs and genes.}

\item{group_prior_var}{a vector of two prior variances for SNPs and gene effects.}

\item{use_null_weight}{If TRUE, allow for a probability of no effect in susie}

\item{snps_only}{If TRUE, use only SNPs in the region data.}

\item{verbose}{If TRUE, print detail messages}

\item{...}{Additional arguments of \code{susie_rss}.}
}
\value{
a data frame of finemapping results.
}
\description{
Runs cTWAS finemapping for a single region without LD
}
\keyword{internal}
