% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_EM.R
\name{fit_EM}
\alias{fit_EM}
\title{Runs EM to estimate parameters.}
\usage{
fit_EM(
  region_data,
  niter = 20,
  init_group_prior = NULL,
  init_group_prior_var = NULL,
  use_null_weight = TRUE,
  ncore = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{region_data}{a list object with the susie input data for each region}

\item{niter}{the number of iterations of the E-M algorithm to perform}

\item{init_group_prior}{a vector of initial prior inclusion probabilities for SNPs and genes.}

\item{init_group_prior_var}{a vector of initial prior variances for SNPs and gene effects.}

\item{use_null_weight}{If TRUE, allow for a probability of no effect in susie}

\item{ncore}{The number of cores used to parallelize susie over regions}

\item{verbose}{If TRUE, print detail messages}
}
\value{
a list of estimated parameters
}
\description{
Runs EM to estimate parameters.
}
\keyword{internal}
