% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_est_parameters.R
\name{get_enrichment_se_Gtest_from_ser_res}
\alias{get_enrichment_se_Gtest_from_ser_res}
\title{Compute standard error and p-value using the G test for the SER result}
\usage{
get_enrichment_se_Gtest_from_ser_res(ser_res_df, groups)
}
\description{
Compute standard error and p-value using the G test for the SER result
}
