% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_enrichment_se_test.R
\name{get_enrichment_se_LRT}
\alias{get_enrichment_se_LRT}
\title{Compute standard error and p-value for enrichment using
the likelihood ratio test from EM results}
\usage{
get_enrichment_se_LRT(
  region_data,
  group_prior,
  group_prior_var,
  niter = 20,
  null_method = c("ctwas", "susie", "none"),
  ncore = 1,
  log_enrichment = TRUE
)
}
\arguments{
\item{region_data}{a list of assembled region data.}

\item{group_prior}{a vector of prior inclusion probabilities for different groups.}

\item{group_prior_var}{a vector of prior variances for different groups.}

\item{niter}{the number of iterations of the E-M algorithm to perform}

\item{null_method}{Method to compute null model, options: "ctwas", "susie" or "none".}

\item{ncore}{The number of cores used to parallelize over regions}

\item{log_enrichment}{If TRUE, return enrichment result in log scale.}
}
\value{
Estimated enrichment, S.E. and p-value.
}
\description{
Compute standard error and p-value for enrichment using
the likelihood ratio test from EM results
}
