% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_anno_finemap_res.R
\name{get_gene_annot_from_ens_db}
\alias{get_gene_annot_from_ens_db}
\title{Get gene annotation table from Ensembl database}
\usage{
get_gene_annot_from_ens_db(ens_db, gene_ids)
}
\arguments{
\item{ens_db}{Ensembl gene annotation database}

\item{gene_ids}{Ensembl gene IDs}
}
\description{
Get gene annotation table from Ensembl database
}
