% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_compute_gene_z.R
\name{get_gene_regions}
\alias{get_gene_regions}
\title{Get regions for each gene}
\usage{
get_gene_regions(gene_info, region_info)
}
\arguments{
\item{gene_info}{a data frame of gene info}

\item{region_info}{a data frame of region definitions}
}
\value{
a data frame of gene info with regions overlapping each gene
}
\description{
Get regions for each gene
}
