% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_screen_regions.R
\name{get_L}
\alias{get_L}
\title{get L for each region from finemapping result}
\usage{
get_L(finemap_res)
}
\arguments{
\item{finemap_res}{a data frame of finemapping result}
}
\description{
get L for each region from finemapping result
}
