% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_merge_regions.R
\name{get_merged_region_info}
\alias{get_merged_region_info}
\title{Get merged region info}
\usage{
get_merged_region_info(boundary_genes, region_info)
}
\arguments{
\item{boundary_genes}{a data frame of boundary gene info}

\item{region_info}{a data frame of region definition and associated LD file names}
}
\value{
a data frame of merged region info
}
\description{
Get merged region info
}
