% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_anno_finemap_res.R
\name{get_molecular_ids}
\alias{get_molecular_ids}
\title{Get original molecular IDs}
\usage{
get_molecular_ids(finemap_res)
}
\arguments{
\item{finemap_res}{a data frame of cTWAS finemapping results.}
}
\value{
a vector of original molecular IDs
}
\description{
Get original molecular IDs
}
