% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_weights.R
\name{get_predictdb_genome_build}
\alias{get_predictdb_genome_build}
\title{Get genome build of PredictDB weight}
\usage{
get_predictdb_genome_build(weight_file)
}
\arguments{
\item{weight_file}{a string, pointing path to weights in PredictDB format.}
}
\description{
Get genome build of PredictDB weight
}
