% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_LD_mismatch.R
\name{get_problematic_genes}
\alias{get_problematic_genes}
\title{Gets problematic genes from problematic SNPs}
\usage{
get_problematic_genes(
  problematic_snps,
  weights,
  finemap_res,
  pip_thresh = 0.5,
  z_thresh = NULL
)
}
\arguments{
\item{problematic_snps}{a character vector of problematic SNP rsIDs}

\item{weights}{a list of weights}

\item{finemap_res}{a data frame of cTWAS finemapping results}

\item{pip_thresh}{cutoff of gene PIP to select genes}

\item{z_thresh}{cutoff of abs(z-scores) to select genes}
}
\value{
a vector of problematic genes
}
\description{
Gets problematic genes from problematic SNPs
}
