% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_postprocess.R
\name{get_problematic_regions}
\alias{get_problematic_regions}
\title{Get regions with problematic high PIP SNPs or genes}
\usage{
get_problematic_regions(ctwas_res, weight, problematic_snps, pip_thresh = 0.5)
}
\arguments{
\item{ctwas_res}{a data frame of cTWAS finemapping result}

\item{weight}{a string, weight filename used in cTWAS}

\item{problematic_snps}{a character vector of problematic SNP rsIDs}

\item{pip_thresh}{Minimum PIP value to select regions}
}
\value{
a character vector of region tags with problematic high PIP SNPs or genes
}
\description{
Get regions with problematic high PIP SNPs or genes
}
