% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_compute_region_cor.R
\name{get_region_cor}
\alias{get_region_cor}
\title{Gets correlation matrices for a single region.
Load precomputed correlation matrices if available in \code{cor_dir},
otherwise, computes correlation matrices if no precomputed correlation matrices,
or force_compute_cor = TRUE.}
\usage{
get_region_cor(
  region_id,
  region_data = NULL,
  LD_info = NULL,
  snp_info = NULL,
  force_compute_cor = FALSE,
  save_cor = FALSE,
  cor_dir = NULL,
  LD_format = c("rds", "rdata", "csv", "txt", "custom"),
  LD_loader = NULL,
  verbose = FALSE
)
}
\arguments{
\item{region_id}{a character string of region id to be finemapped}

\item{region_data}{a list object with data for the regions}

\item{LD_info}{a list of paths to LD matrices for each of the regions. Required when \code{use_LD = TRUE}.}

\item{snp_info}{a list of SNP info data frames for LD reference. Required when \code{use_LD = TRUE}.}

\item{force_compute_cor}{TRUE/FALSE. If TRUE, force computing correlation (R) matrices}

\item{save_cor}{TRUE/FALSE. If TRUE, save correlation (R) matrices to \code{cor_dir}}

\item{cor_dir}{a string, the directory to store correlation (R) matrices}

\item{LD_format}{file format for LD matrix. If "custom", use a user defined
\code{LD_loader()} function to load LD matrix.}

\item{LD_loader}{a user defined function to load LD matrix when \code{LD_format = "custom"}.}

\item{verbose}{TRUE/FALSE. If TRUE, print detail messages}
}
\value{
a list of correlation matrices (R_snp, R_snp_gene and R_gene)
}
\description{
Gets correlation matrices for a single region.
Load precomputed correlation matrices if available in \code{cor_dir},
otherwise, computes correlation matrices if no precomputed correlation matrices,
or force_compute_cor = TRUE.
}
