% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_get_regionlist.R
\name{get_regionlist}
\alias{get_regionlist}
\title{Get regionlist for each region in region_info}
\usage{
get_regionlist(
  region_info,
  gene_info,
  weight_list = NULL,
  select = NULL,
  thin = 1,
  maxSNP = Inf,
  minvar = 1,
  adjust_boundary = TRUE
)
}
\arguments{
\item{region_info}{a data frame of region definition and associated file names}

\item{gene_info}{a data frame of gene information obtained from \code{compute_gene_z}}

\item{weight_list}{a list of weights}

\item{select}{Default is NULL, all variants will be selected. Or a vector of variant IDs,
or a data frame with columns id and z (id is for gene or SNP id, z is for z scores).
z will be used for remove SNPs if the total number of SNPs exceeds limit. See
parameter `maxSNP` for more information.}

\item{thin}{The proportion of SNPs to be used for the parameter estimation and
initial screening region steps.
Smaller \code{thin} parameters reduce runtime at the expense of accuracy.
The fine mapping step is rerun using full SNPs for regions with strong gene signals.}

\item{maxSNP}{Inf or integer. Maximum number of SNPs in a region. Default is
Inf, no limit. This can be useful if there are many SNPs in a region and you don't
have enough memory to run the program. This applies to the last rerun step
(using full SNPs and rerun susie for regions with strong gene signals) only.}

\item{minvar}{minimum number of SNPs (and genes) in a region}

\item{adjust_boundary}{identify cross-boundary genes, adjust regionlist and update weigh_list}
}
\value{
a list with regionlist, updated weight_list, and cross-bounary genes
}
\description{
Get regionlist for each region in region_info
}
