% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_harmonize_data.R
\name{harmonize_z_ld}
\alias{harmonize_z_ld}
\title{Harmonize z scores from GWAS to match ld reference genotypes.
Flip signs when reverse complement matches.}
\usage{
harmonize_z_ld(z_snp, ld_snpinfo)
}
\arguments{
\item{z_snp}{a data frame, with columns "id", "A1", "A2" and "z".
Z scores for every SNP. "A1" is the effect allele.}

\item{ld_snpinfo}{A data frame, snp info for LD reference,
with columns "chrom", "id", "pos", "alt", "ref".}
}
\value{
A data frame, z_snp with the "z" columns flipped to match LD ref.
}
\description{
Harmonize z scores from GWAS to match ld reference genotypes.
Flip signs when reverse complement matches.
}
