% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_regions.R
\name{index_regions}
\alias{index_regions}
\title{Get gene and SNP index for each region}
\usage{
index_regions(pvarfs, exprvarfs, regionfile)
}
\arguments{
\item{regionfile}{Has three columns, chr, start, end}
}
\value{
A list. Items correspond to each pvarf/exprvarf. Each Item is
 also a list, the items in this list are for each region.
}
\description{
For each region, get the index for snp and gene (index
is location/column number in .pgen file or .expr file) located within
this region.
}
