% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_initialize.R
\name{init_finalize}
\alias{init_finalize}
\title{a susie fit object in order to initialize susie model. Revised from
susieR::init_finalize}
\usage{
init_finalize(s, X = NULL, Xr = NULL)
}
\description{
a susie fit object in order to initialize susie model. Revised from
susieR::init_finalize
}
