% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_revise.R
\name{init_finalize_rss}
\alias{init_finalize_rss}
\title{a susie fit object in order to initialize susie rss model. Revised from
susieR::init_finalize_rss}
\usage{
init_finalize_rss(s, R = NULL, Rz = NULL)
}
\description{
a susie fit object in order to initialize susie rss model. Revised from
susieR::init_finalize_rss
}
