% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_susie_rss.R
\name{initiate_group_priors}
\alias{initiate_group_priors}
\title{set pi_prior and V_prior based on init_group_prior and init_group_prior_var}
\usage{
initiate_group_priors(group_prior = NULL, group_prior_var = NULL, groups)
}
\description{
set pi_prior and V_prior based on init_group_prior and init_group_prior_var
}
