% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_utils.R
\name{label_merged_regions}
\alias{label_merged_regions}
\title{Function to identify consecutive regions and label them}
\usage{
label_merged_regions(df)
}
\description{
Function to identify consecutive regions and label them
}
