% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_merge_region.R
\name{label_overlapping_regions}
\alias{label_overlapping_regions}
\title{Identify overlapping regions}
\usage{
label_overlapping_regions(boundary_genes)
}
\description{
Identify overlapping regions
}
