% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_weights.R
\name{load_fusion_weights}
\alias{load_fusion_weights}
\title{Loads weights in FUSION format}
\usage{
load_fusion_weights(
  weight_file,
  fusion_method = c("lasso", "enet", "top1", "blup"),
  fusion_genome_version = c("b38", "b37"),
  ncore = 1
)
}
\arguments{
\item{weight_file}{a string or a vector, pointing path to one or multiple
sets of weights in FUSION format.}

\item{fusion_method}{a string, specifying the method to choose in FUSION models.}

\item{fusion_genome_version}{a string, specifying the genome version of FUSION models}

\item{ncore}{integer, number of cores for parallel computing.}
}
\description{
Loads weights in FUSION format
}
