% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_utils.R
\name{load_LD}
\alias{load_LD}
\title{Load LD matrix}
\usage{
load_LD(
  file,
  format = c("rds", "rdata", "mtx", "csv", "txt", "custom"),
  LD_loader
)
}
\arguments{
\item{file}{path to LD matrix}

\item{format}{file format for LD matrix. If "custom", use a user defined
\code{LD_loader()} function to load LD matrix.}

\item{LD_loader}{a user defined function to load LD matrix}
}
\description{
Load LD matrix
}
