% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_weights.R
\name{load_weights}
\alias{load_weights}
\title{Loads weights in PredictDB or FUSION format}
\usage{
load_weights(
  weight_file,
  weight_format = c("PredictDB", "FUSION"),
  filter_protein_coding_genes = FALSE,
  load_predictdb_LD = FALSE,
  fusion_method = c("lasso", "enet", "top1", "blup"),
  fusion_genome_version = c("b38", "b37"),
  ncore = 1
)
}
\arguments{
\item{weight_file}{a string or a vector, pointing path to one or multiple sets of weights in PredictDB or FUSION format.}

\item{weight_format}{a string or a vector, specifying format of each weight file, e.g. PredictDB, FUSION.}

\item{filter_protein_coding_genes}{TRUE/FALSE. If TRUE, keep protein coding genes only. This option is only for PredictDB weights}

\item{load_predictdb_LD}{TRUE/FALSE. If TRUE, load pre-computed LD among weight SNPs. This option is only for PredictDB weights}

\item{fusion_method}{a string, specifying the method to choose in FUSION models}

\item{fusion_genome_version}{a string, specifying the genome version of FUSION models}

\item{ncore}{integer, number of cores for parallel computing.}
}
\description{
Loads weights in PredictDB or FUSION format
}
