% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_plots.R
\name{make_convergence_plots}
\alias{make_convergence_plots}
\title{Make convergence plots for the estimated parameters}
\usage{
make_convergence_plots(
  param,
  gwas_n,
  title.size = 10,
  legend.size = 8,
  colors = c("#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00", "#CC79A7",
    "#999999"),
  ncol = 2
)
}
\arguments{
\item{param}{a list of cTWAS parameter estimation result from \code{est_param}}

\item{gwas_n}{the sample size of the GWAS summary statistics}

\item{title.size}{font size of the plot title}

\item{legend.size}{font size of the plot legend title}

\item{colors}{colors for different groups, passed as the
\code{values} input to \code{\link[ggplot2]{scale_color_manual}}.
If fewer colors than "fits" are given, the colors are recycled.}

\item{ncol}{number of columns in the output plot}
}
\description{
Make convergence plots for the estimated parameters
}
