% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_weights.R
\name{make_predictdb_from_QTLs}
\alias{make_predictdb_from_QTLs}
\title{Makes PredictDB weights from QTL data}
\usage{
make_predictdb_from_QTLs(
  weight_table,
  extra_table = NULL,
  covar_table = NULL,
  use_top_QTL = FALSE,
  outputdir = getwd(),
  outname
)
}
\arguments{
\item{weight_table}{a data frame of the genes, QTLs and weights, with columns:
"gene", "rsid", "varID", "ref_allele", "eff_allele", "weight".}

\item{extra_table}{a data frame (optional) with information of the genes
in \code{weight_table} ("gene","genename","gene_type", etc.).
If NULL, create a simply extra_table based on weight_table}

\item{covar_table}{a data frame of covariances between variants, with columns:
"GENE","RSID1","RSID2", "VALUE".
If NULL, do not create covariance files (.txg.gz), unless \code{use_top_QTL=TRUE}.}

\item{use_top_QTL}{TRUE/FALSE. If TRUE, only keep the top QTL with
the largest abs(weight) for each gene (molecular trait), and
create a simple covar_table with covariance set to 1.}

\item{outputdir}{output directory}

\item{outname}{name of the output weight file}
}
\description{
Makes PredictDB weights from QTL data
}
