% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_weights.R
\name{make_predictdb_weights_from_top_QTLs}
\alias{make_predictdb_weights_from_top_QTLs}
\title{Makes PredictDB weights from top QTLs}
\usage{
make_predictdb_weights_from_top_QTLs(
  weight_table,
  extra_table = NULL,
  outputdir = getwd(),
  outname
)
}
\arguments{
\item{weight_table}{a data frame of the genes, QTLs and weights, with columns:
"gene", "rsid", "varID", "ref_allele", "eff_allele", "weight".}

\item{extra_table}{a data frame (optional) with information of the genes
in \code{weight_table} ("gene","genename","gene_type", etc.)}

\item{outputdir}{output directory}

\item{outname}{name of the output weight file}
}
\description{
Make PredictDB weights from only top QTLs.
Each gene (or molecular trait) has only one top QTL, defined by users,
based on min p-value, max abs(weight), etc.
}
