% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_compute_gene_z.R
\name{map_gene_regions}
\alias{map_gene_regions}
\title{Map regions for each gene.}
\usage{
map_gene_regions(gene_info, region_info, ncore = 1)
}
\arguments{
\item{gene_info}{a data frame of gene info.}

\item{region_info}{a data frame of region definitions.}

\item{ncore}{The number of cores used to parallelize over genes.}
}
\value{
a data frame of gene info with regions overlapping each gene
}
\description{
Map regions for each gene.
}
