% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_preprocess_regions.R
\name{map_snp_info_regions}
\alias{map_snp_info_regions}
\title{Reads SNP info from all regions and map SNPs to regions.}
\usage{
map_snp_info_regions(region_info, ref_snp_info, ncore = 1)
}
\arguments{
\item{region_info}{a data frame of region definition}

\item{ref_snp_info}{a data frame of all variant info in the LD reference.}

\item{ncore}{The number of cores used to parallelize over regions}
}
\description{
Reads SNP info from all regions and map SNPs to regions.
}
