% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_merge_regions.R
\name{merge_finemap_regions}
\alias{merge_finemap_regions}
\title{Title}
\usage{
merge_finemap_regions(
  boundary_genes,
  region_data,
  region_info,
  LD_info,
  snp_info,
  z_snp,
  z_gene,
  weights,
  use_LD = TRUE,
  maxSNP = Inf,
  expand = TRUE,
  group_prior = NULL,
  group_prior_var = NULL,
  L = 5,
  use_null_weight = TRUE,
  coverage = 0.95,
  min_abs_corr = 0.5,
  force_compute_cor = FALSE,
  save_cor = FALSE,
  cor_dir = getwd(),
  ncore = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{boundary_genes}{a data frame of boundary gene info}

\item{region_data}{a list of original region_data}

\item{region_info}{a data frame of region definition and associated LD file names}

\item{z_snp}{A data frame with columns: "id", "z", giving the z-scores for SNPs.}

\item{z_gene}{A data frame with columns: "id", "z", giving the z-scores for genes.}

\item{weights}{a list of weights for each gene}

\item{maxSNP}{Inf or integer. Maximum number of SNPs in a region. Default is
Inf, no limit. This can be useful if there are many SNPs in a region and you don't
have enough memory to run the program.}

\item{expand}{TRUE/FALSE. If TRUE, expand merged region_data with full SNPs}

\item{group_prior}{a vector of two prior inclusion probabilities for SNPs and genes.}

\item{group_prior_var}{a vector of two prior variances for SNPs and gene effects.}

\item{L}{the number of effects for susie during the fine mapping steps}

\item{use_null_weight}{TRUE/FALSE. If TRUE, allow for a probability of no effect in susie}

\item{coverage}{A number between 0 and 1 specifying the \dQuote{coverage} of the estimated confidence sets}

\item{min_abs_corr}{Minimum absolute correlation allowed in a
credible set. The default, 0.5, corresponds to a squared
correlation of 0.25, which is a commonly used threshold for
genotype data in genetic studies.}

\item{force_compute_cor}{TRUE/FALSE. If TRUE, force computing correlation (R) matrices}

\item{save_cor}{TRUE/FALSE. If TRUE, save correlation (R) matrices}

\item{cor_dir}{a string, the directory to store correlation (R) matrices}

\item{ncore}{The number of cores used to parallelize computation over regions}

\item{verbose}{TRUE/FALSE. If TRUE, print detail messages}

\item{logfile}{the log file, if NULL will print log info on screen}
}
\value{
a list of merged region data, merged region info,
  merged region IDs,and finemapping results for merged regions.
}
\description{
Title
}
