% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_postprocess_LD_mismatch.R
\name{postprocess_LD_mismatch}
\alias{postprocess_LD_mismatch}
\title{Runs cTWAS post-processing procedure for merging regions}
\usage{
postprocess_LD_mismatch(
  region_data,
  z_snp,
  weights,
  LD_map,
  snp_map,
  gwas_n,
  finemap_res,
  susie_alpha_res,
  group_prior = NULL,
  group_prior_var = NULL,
  min_nonSNP_PIP = 0.8,
  filter_cs = FALSE,
  p_diff_thresh = 5e-06,
  pip_thresh = 0.5,
  z_thresh = NULL,
  plot = TRUE,
  maxSNP = Inf,
  ncore = 1,
  verbose = FALSE,
  logfile = NULL,
  ...
)
}
\arguments{
\item{region_data}{region_data to be fine-mapped.}

\item{z_snp}{A data frame with columns: "id", "z", giving the z-scores for SNPs.}

\item{weights}{a list of preprocessed weights.}

\item{LD_map}{a data frame with filenames of LD matrices for the regions.}

\item{snp_map}{a list of data frames with SNP-to-region map for the reference.}

\item{gwas_n}{integer, GWAS sample size.}

\item{finemap_res}{a data frame of original finemapping result.}

\item{susie_alpha_res}{a data frame of original susie alpha result.}

\item{group_prior}{a vector of prior inclusion probabilities for different groups.
If NULL, it will use uniform prior inclusion probabilities.}

\item{group_prior_var}{a vector of prior variances for different groups.
If NULL, it will set prior variance = 50 as the default in \code{susie_rss}.}

\item{min_nonSNP_PIP}{regions with total non-SNP PIP >= \code{min_nonSNP_PIP}
will be selected to run LD mismatch diagnosis.}

\item{filter_cs}{If TRUE, computes region non-SNP PIP only in credible sets.}

\item{p_diff_thresh}{p-value cutoff for identifying problematic SNPs
with significant difference between observed z-scores and estimated values.}

\item{pip_thresh}{cutoff of gene PIP to select problematic genes.}

\item{z_thresh}{cutoff of abs(z-scores) to select problematic genes.}

\item{plot}{If TRUE, plot observed z-score vs the expected value.}

\item{maxSNP}{Inf or integer. Maximum number of SNPs in a region. Default is
Inf, no limit. This can be useful if there are many SNPs in a region and you don't
have enough memory to run the program.}

\item{ncore}{The number of cores used to parallelize computation over regions}

\item{verbose}{If TRUE, print detail messages.}

\item{logfile}{the log file, if NULL will print log info on screen}

\item{...}{Additional arguments of \code{finemap_regions}.}
}
\value{
a list with region merge results.
}
\description{
Runs cTWAS post-processing procedure for merging regions
}
