% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_postprocess_merge_regions.R
\name{postprocess_merge_regions}
\alias{postprocess_merge_regions}
\title{Runs cTWAS post-processing procedure for merging regions}
\usage{
postprocess_merge_regions(
  region_info,
  region_data,
  z_snp,
  z_gene,
  weights,
  LD_map,
  snp_map,
  finemap_res,
  susie_alpha_res,
  mapping_table = NULL,
  L = 5,
  group_prior = NULL,
  group_prior_var = NULL,
  pip_thresh = 0.5,
  filter_cs = FALSE,
  maxSNP = Inf,
  save_cor = FALSE,
  cor_dir = NULL,
  ncore = 1,
  verbose = FALSE,
  logfile = NULL,
  ...
)
}
\arguments{
\item{region_info}{a data frame of region definitions.}

\item{region_data}{region_data to be fine-mapped.}

\item{z_snp}{A data frame with columns: "id", "z", giving the z-scores for SNPs.}

\item{z_gene}{A data frame with columns: "id", "z", giving the z-scores for genes.}

\item{weights}{a list of preprocessed weights.}

\item{LD_map}{a data frame with filenames of LD matrices for the regions.}

\item{snp_map}{a list of data frames with SNP-to-region map for the reference.}

\item{finemap_res}{a data frame of original finemapping result.}

\item{susie_alpha_res}{a data frame of original susie alpha result.}

\item{mapping_table}{a data frame of mapping between molecular traits and genes,
with required columns: "molecular_id", "gene_name".}

\item{L}{the number of effects or a vector of number of effects for each region.}

\item{group_prior}{a vector of prior inclusion probabilities for different groups.
If NULL, it will use uniform prior inclusion probabilities.}

\item{group_prior_var}{a vector of prior variances for different groups.
If NULL, it will set prior variance = 50 as the default in \code{susie_rss}.}

\item{pip_thresh}{PIP cutoff for selecting boundary genes to merge regions.}

\item{filter_cs}{If TRUE, only select boundary genes in credible sets for region merge.}

\item{maxSNP}{Inf or integer. Maximum number of SNPs in a region. Default is
Inf, no limit. This can be useful if there are many SNPs in a region and you don't
have enough memory to run the program.}

\item{save_cor}{If TRUE, save correlation (R) matrices to \code{cor_dir}}

\item{cor_dir}{a string, the directory to store correlation (R) matrices}

\item{ncore}{The number of cores used to parallelize computation over regions}

\item{verbose}{If TRUE, print detail messages.}

\item{logfile}{the log file, if NULL will print log info on screen}

\item{...}{Additional arguments of \code{finemap_regions}.}
}
\value{
a list with region merge results.
}
\description{
Runs cTWAS post-processing procedure for merging regions
}
