% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_harmonize_data.R
\name{preharmonize_wgt_ld}
\alias{preharmonize_wgt_ld}
\title{Harmonize PredictDB weights and LD reference}
\usage{
preharmonize_wgt_ld(
  weight,
  ld_R_dir,
  outputdir = getwd(),
  outname,
  strand_ambig_action_wgt = c("drop", "none", "recover")
)
}
\arguments{
\item{weight}{a string, pointing to a directory with the fusion/twas format of weights, or a .db file in predictdb format.
A vector of multiple sets of weights in PredictDB format can also be specified; genes will have their filename appended
to their gene name to ensure IDs are unique.}

\item{outputdir}{a string, the directory to store output}

\item{outname}{a string, the output name}

\item{strand_ambig_action_wgt}{the action to take to harmonize strand ambiguous variants (A/T, G/C) between 
the weights and LD reference. "drop" removes the ambiguous variant from the prediction models. "none" treats the variant 
as unambiguous, flipping the weights to match the LD reference and then taking no additional action. "recover" uses a procedure
to recover strand ambiguous variants. This procedure compares correlations between variants in the 
LD reference and prediction models, and it can only be used with PredictDB format prediction models, which include this
information.}

\item{LD_R_dir}{a string, pointing to a directory containing all LD matrix files and variant information. Expects .RDS files which contain LD correlation matrices for a region/block.
For each RDS file, a file with same base name but ended with .Rvar needs to be present in the same folder. the .Rvar file has 5 required columns: "chrom", "id", "pos", "alt", "ref". 
If using PredictDB format weights and \code{scale_by_ld_variance=T}, a 6th column is also required: "variance", which is the variance of the each SNP.
The order of rows needs to match the order of rows in .RDS file.}
}
\description{
Harmonize PredictDB weights and LD reference
}
