% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_utils.R
\name{prep_exprvar}
\alias{prep_exprvar}
\title{Prepare .exprvar file}
\usage{
prep_exprvar(exprf)
}
\arguments{
\item{exprf}{expression variable info files, the output of \code{impute_expr}}
}
\value{
corresponding exprvar file
}
\description{
Prepare .exprvar file
}
