% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_process_genotype.R
\name{prep_pgen}
\alias{prep_pgen}
\title{Read .pgen file into R}
\usage{
prep_pgen(pgenf, pvarf)
}
\arguments{
\item{pgenf}{.pgen file or .bed file}

\item{pvarf}{.pvar file or .bim file with have proper
header.  Matching `pgenf`.}
}
\value{
A matrix of allele count for each variant (columns) in each sample
 (rows). ALT allele in pvar file is counted (A1 allele in .bim file is the ALT
  allele).
}
\description{
Read .pgen file into R
}
