% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_plink_utils.R
\name{prep_pvar}
\alias{prep_pvar}
\title{Prepare .pvar file}
\usage{
prep_pvar(pgenf, outputdir = getwd())
}
\arguments{
\item{pgenf}{pgen file
.pvar file format: https://www.cog-genomics.org/plink/2.0/formats#pvar}

\item{outputdir}{a string, the directory to store output}
}
\value{
corresponding pvar file
}
\description{
Prepare .pvar file
}
