% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_preprocess_regions.R
\name{preprocess_region_LD_snp_info}
\alias{preprocess_region_LD_snp_info}
\title{Preprocess region info, LD info, and SNP info from all regions and map SNPs to regions.}
\usage{
preprocess_region_LD_snp_info(
  region_info = NULL,
  ref_snp_info = NULL,
  use_LD = TRUE,
  chrom = 1:22,
  ncore = 1
)
}
\arguments{
\item{region_info}{a data frame of region definition}

\item{ref_snp_info}{a data frame of all variant info in the LD reference.}

\item{use_LD}{TRUE/FALSE. If TRUE, use LD for finemapping. Otherwise, use "no-LD" version.}

\item{chrom}{chromosomes to keep}

\item{ncore}{The number of cores used to parallelize over regions}
}
\description{
Preprocess region info, LD info, and SNP info from all regions and map SNPs to regions.
}
