% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_preprocess_weight.R
\name{preprocess_weight}
\alias{preprocess_weight}
\title{Preprocess PredictDB weights and harmonize with LD reference
(adapted from preharmonize_wgt_ld)}
\usage{
preprocess_weight(
  weight,
  region_info,
  drop_strand_ambig = TRUE,
  write_db = FALSE,
  outputdir = getwd(),
  outname = NULL,
  logfile = NULL
)
}
\arguments{
\item{weight}{a string, pointing to a directory with the fusion/twas format of weights, or a .db file in predictdb format.
A vector of multiple sets of weights in PredictDB format can also be specified; genes will have their filename appended
to their gene name to ensure IDs are unique.}

\item{region_info}{a data frame of region definition and associated file names.}

\item{drop_strand_ambig}{TRUE/FALSE, if TRUE remove strand ambiguous variants (A/T, G/C).}

\item{write_db}{TRUE/FALSE, if TRUE, write processed weights as .db file in predictdb format}

\item{outputdir}{a string, the directory to store output}

\item{outname}{a string, the output name.}

\item{logfile}{the log file, if NULL will print log info on screen.}
}
\value{
a list of processed weight table and extra table
}
\description{
Preprocess PredictDB weights and harmonize with LD reference
(adapted from preharmonize_wgt_ld)
}
