% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_preprocess_z.R
\name{preprocess_wgt_ld}
\alias{preprocess_wgt_ld}
\title{Preprocess PredictDB weights and harmonize with LD reference
(adapted from preharmonize_wgt_ld)}
\usage{
preprocess_wgt_ld(
  weight,
  ld_R_dir,
  outputdir = getwd(),
  outname,
  drop_strand_ambig = TRUE
)
}
\arguments{
\item{weight}{a string, pointing to a directory with the fusion/twas format of weights, or a .db file in predictdb format.
A vector of multiple sets of weights in PredictDB format can also be specified; genes will have their filename appended
to their gene name to ensure IDs are unique.}

\item{outputdir}{a string, the directory to store output}

\item{outname}{a string, the output name.}

\item{drop_strand_ambig}{TRUE/FALSE, if TRUE remove strand ambiguous variants (A/T, G/C).}

\item{LD_R_dir}{a string, pointing to a directory containing all LD matrix files and variant information. Expects .RDS files which contain LD correlation matrices for a region/block.
For each RDS file, a file with same base name but ended with .Rvar needs to be present in the same folder. the .Rvar file has 5 required columns: "chrom", "id", "pos", "alt", "ref".
If using PredictDB format weights and \code{scale_by_ld_variance=T}, a 6th column is also required: "variance", which is the variance of the each SNP.
The order of rows needs to match the order of rows in .RDS file.}
}
\description{
Preprocess PredictDB weights and harmonize with LD reference
(adapted from preharmonize_wgt_ld)
}
