% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_preprocess_data.R
\name{preprocess_z_ld}
\alias{preprocess_z_ld}
\title{Preprocess GWAS summary statistics, harmonize GWAS z-scores and LD reference, and filter LD mismatches using susie_rss}
\usage{
preprocess_z_ld(
  z_snp,
  ld_R_dir,
  chr = 1:22,
  ld_regions = c("EUR", "ASN", "AFR"),
  ld_regions_version = c("b37", "b38"),
  ld_regions_custom = NULL,
  filestem,
  gwas_n,
  outputdir = getwd(),
  outname = NULL,
  logfile = NULL,
  drop_multiallelic = T,
  detect_ld_mismatch = F,
  flip_allele = T,
  remove_ld_mismatch = F,
  ncore = 1
)
}
\arguments{
\item{z_snp}{A data frame with two columns: "id", "A1", "A2", "z". giving the z scores for
snps. "A1" is effect allele. "A2" is the other allele.}

\item{chr}{a vector containing the chromosome numbers to process.}

\item{ld_regions}{a string, population of LD reference, "EUR", "ASN", or "AFR".}

\item{ld_regions_version}{a string, version LD reference, "b37", "b38".}

\item{ld_regions_custom}{a string, path of LD regions file.}

\item{filestem}{a string, filestem of reference LD matrix.}

\item{gwas_n}{integer, GWAS sample size}

\item{outputdir}{a string, the directory to store output.}

\item{outname}{a string, the output name.}

\item{logfile}{the log file, if NULL will print log info on screen.}

\item{drop_multiallelic}{TRUE/FALSE. If TRUE, multiallelic variants will be dropped from the summary statistics.}

\item{detect_ld_mismatch}{TRUE/FALSE. If TRUE, detect LD mismatches by susie_rss, and report problematic variants, and
variants with allele flipping.}

\item{flip_allele}{TRUE/FALSE. If TRUE, flip the sign for variants detected with allele flipping.}

\item{remove_ld_mismatch}{TRUE/FALSE. If TRUE, remove problematic variants with LD mismatches.}

\item{ncore}{integer, number of cores for parallel computing.}

\item{LD_R_dir}{a string, pointing to a directory containing all LD matrix files and variant information. Expects .RDS files which contain LD correlation matrices for a region/block.
For each RDS file, a file with same base name but ended with .Rvar needs to be present in the same folder. the .Rvar file has 5 required columns: "chrom", "id", "pos", "alt", "ref".
If using PredictDB format weights and \code{scale_by_ld_variance=T}, a 6th column is also required: "variance", which is the variance of the each SNP.
The order of rows needs to match the order of rows in .RDS file.}
}
\description{
Preprocess GWAS summary statistics, harmonize GWAS z-scores and LD reference, and filter LD mismatches using susie_rss
}
