% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_preprocess_z_snp.R
\name{preprocess_z_snp}
\alias{preprocess_z_snp}
\title{Preprocess GWAS z-scores, harmonize GWAS z-scores with LD reference}
\usage{
preprocess_z_snp(
  z_snp,
  LD_snpinfo,
  drop_multiallelic = TRUE,
  drop_strand_ambig = TRUE,
  logfile = NULL
)
}
\arguments{
\item{z_snp}{A data frame with two columns: "id", "A1", "A2", "z". giving the z scores for
snps. "A1" is effect allele. "A2" is the other allele.}

\item{LD_snpinfo}{a data frame, SNP info for LD reference,
with columns "chrom", "id", "pos", "alt", "ref".}

\item{drop_multiallelic}{TRUE/FALSE. If TRUE, multiallelic variants will be dropped from the summary statistics.}

\item{drop_strand_ambig}{TRUE/FALSE, if TRUE remove strand ambiguous variants (A/T, G/C).}

\item{logfile}{the log file, if NULL will print log info on screen.}
}
\value{
a data frame preprocessed GWAS z-scores
}
\description{
Preprocess GWAS z-scores, harmonize GWAS z-scores with LD reference
}
