% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_preprocess_weight.R
\name{process_weight}
\alias{process_weight}
\title{Preprocess PredictDB weights and harmonize with LD reference}
\usage{
process_weight(
  weight_file,
  region_info,
  z_snp = NULL,
  weight_format = c("PredictDB", "Fusion"),
  filter_protein_coding_genes = TRUE,
  drop_strand_ambig = TRUE,
  scale_by_ld_variance = TRUE
)
}
\arguments{
\item{weight_file}{a string, pointing to a directory with the fusion/twas format of weights, or a .db file in predictdb format.
A vector of multiple sets of weights in PredictDB format can also be specified; genes will have their filename appended
to their gene name to ensure IDs are unique.}

\item{region_info}{a data frame of region definition and associated file names.}

\item{z_snp}{A data frame with columns: "id", "A1", "A2", "z". giving the z scores for
snps. "A1" is effect allele. "A2" is the other allele. For harmonized data, A1 and A2 are not required.}

\item{weight_format}{weight format: PredictDB, or Fusion}

\item{filter_protein_coding_genes}{TRUE/FALSE. If TRUE, keep protein coding genes only}

\item{drop_strand_ambig}{TRUE/FALSE, if TRUE remove strand ambiguous variants (A/T, G/C).}

\item{scale_by_ld_variance}{TRUE/FALSE. If TRUE, PredictDB weights are scaled by genotype variance, which is the default
behavior for PredictDB}
}
\value{
a list of processed weight list and weight info table
}
\description{
Preprocess PredictDB weights and harmonize with LD reference
}
