% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_preprocess_z.R
\name{process_z}
\alias{process_z}
\title{Process GWAS summary statistics, harmonize GWAS z-scores and LD reference, and detect LD mismatches using susie_rss}
\usage{
process_z(
  z_snp,
  region_info,
  gwas_n = NULL,
  drop_multiallelic = TRUE,
  drop_strand_ambig = TRUE,
  detect_ld_mismatch = FALSE,
  flip_ld_mismatch = TRUE,
  drop_ld_mismatch = FALSE,
  ncore = 1,
  logfile = NULL
)
}
\arguments{
\item{z_snp}{A data frame with two columns: "id", "A1", "A2", "z". giving the z scores for
snps. "A1" is effect allele. "A2" is the other allele.}

\item{region_info}{a data frame of region definition and associated file names.}

\item{gwas_n}{integer, GWAS sample size}

\item{drop_multiallelic}{TRUE/FALSE. If TRUE, multiallelic variants will be dropped from the summary statistics.}

\item{drop_strand_ambig}{TRUE/FALSE, if TRUE remove strand ambiguous variants (A/T, G/C).}

\item{detect_ld_mismatch}{TRUE/FALSE. If TRUE, detect LD mismatches by susie_rss,
and report problematic variants, and variants with allele flipping.}

\item{flip_ld_mismatch}{TRUE/FALSE. If TRUE, flip the sign for variants detected with allele flipping.}

\item{drop_ld_mismatch}{TRUE/FALSE. If TRUE, remove problematic variants with LD mismatches.}

\item{ncore}{integer, number of cores for parallel computing when detecting LD mismatches.}

\item{logfile}{the log file, if NULL will print log info on screen.}

\item{chrom}{a vector containing the chromosome numbers to process.}
}
\value{
a list of processed z_snp and LD mismatch results
}
\description{
Process GWAS summary statistics, harmonize GWAS z-scores and LD reference, and detect LD mismatches using susie_rss
}
