% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_read_data.R
\name{read_bim}
\alias{read_bim}
\title{Read .bim file into R}
\usage{
read_bim(bimf)
}
\arguments{
\item{bimf}{.bim file with proper format: https://www.cog-genomics.org/plink/2.0/formats#bim}
}
\value{
A data.table. variant info
}
\description{
Read .bim file into R
}
