% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_simulations.R
\name{read_exprvar}
\alias{read_exprvar}
\title{Read .exprvar file into R}
\usage{
read_exprvar(exprvarf)
}
\arguments{
\item{exprvarf}{expression variable info files, prepared by the \code{prep_exprvar} function}
}
\value{
A data.table. variant info
}
\description{
Read .exprvar file into R
}
