% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_preprocess_z_snp.R
\name{read_gwas}
\alias{read_gwas}
\title{Read GWAS summary statistics}
\usage{
read_gwas(
  gwas,
  id = "rsID",
  A1 = "ALT",
  A2 = "REF",
  z = "Z",
  beta = "ES",
  se = "SE"
)
}
\arguments{
\item{gwas}{A data frame of GWAS summary statistics.}

\item{id}{Column name of the variant IDs.}

\item{A1}{Column name of the alternative alleles.}

\item{A2}{Column name of the reference alleles.}

\item{z}{Column name of z-scores.}

\item{beta}{Column name of effect sizes.}

\item{se}{Column name of the standard errors.}
}
\value{
A data frame of processed GWAS summary statistics.
}
\description{
Read GWAS summary statistics
}
