% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_read_data.R
\name{read_ld_Rvar}
\alias{read_ld_Rvar}
\title{read variant information for all ld matrices in `ld_Rf`.}
\usage{
read_ld_Rvar(ld_Rf)
}
\arguments{
\item{ld_Rf}{a vector of paths to the LD matrices}
}
\value{
a data frame with columns: "chrom", "id", "pos", "alt", "ref"
}
\description{
read variant information for all ld matrices in `ld_Rf`.
}
