% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_read_data.R
\name{read_ld_Rvar_RDS}
\alias{read_ld_Rvar_RDS}
\title{read variant information associated with a LD R matrix .RDS file.}
\usage{
read_ld_Rvar_RDS(ld_RDSf)
}
\value{
a data frame with columns: "chrom", "id", "pos", "alt", "ref". "alt" is
the coded allele
}
\description{
read variant information associated with a LD R matrix .RDS file.
}
