% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_utils.R
\name{read_ld_Rvar_snp_info}
\alias{read_ld_Rvar_snp_info}
\title{read LD Rvar files from all regions and combine SNP info a data frame
adapted from read_ld_Rvar()}
\usage{
read_ld_Rvar_snp_info(Rvar_files)
}
\description{
read LD Rvar files from all regions and combine SNP info a data frame
adapted from read_ld_Rvar()
}
