% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_utils.R
\name{read_LD_snpinfo_regions}
\alias{read_LD_snpinfo_regions}
\title{Read LD SNP info for all regions as a data frame
if region_info has files in 'SNP_info' column, read from the files in the 'SNP_info' column;
otherwise, use all SNP info from the LD reference 'LD_ref_snpinfo'.}
\usage{
read_LD_snpinfo_regions(region_info, LD_ref_snpinfo = NULL, ncore = 1)
}
\arguments{
\item{region_info}{a data frame of region definition and associated file names}

\item{LD_ref_snpinfo}{a data frame of all variant info in the LD reference.}

\item{ncore}{The number of cores used to parallelize susie over regions}
}
\description{
Read LD SNP info for all regions as a data frame
if region_info has files in 'SNP_info' column, read from the files in the 'SNP_info' column;
otherwise, use all SNP info from the LD reference 'LD_ref_snpinfo'.
}
