% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_utils.R
\name{read_pgen}
\alias{read_pgen}
\title{Read pgen file into R}
\usage{
read_pgen(pgen, variantidx = NULL, meanimpute = F)
}
\arguments{
\item{pgen}{.pgen file or .bed file}

\item{variantidx}{variant index. If NULL, all variants will be extracted.}
}
\value{
A matrix, columns are allele count for each SNP, rows are
 for each sample.
}
\description{
Read pgen file into R
}
