% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_utils.R
\name{read_pvar}
\alias{read_pvar}
\title{Read .pvar file into R}
\usage{
read_pvar(pvarf)
}
\arguments{
\item{pvarf}{.pvar file with proper format: https://www.cog-genomics.org/plink/2.0/formats#pvar}
}
\value{
A data.table. variant info
}
\description{
Read .pvar file into R
}
