% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_utils.R
\name{read_region_ld_Rinfo}
\alias{read_region_ld_Rinfo}
\title{read and check Rvar from the region_info table, adapted from write_ld_Rf()}
\usage{
read_region_ld_Rinfo(region_info)
}
\description{
read and check Rvar from the region_info table, adapted from write_ld_Rf()
}
