% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_utils.R
\name{read_region_LD_SNP_info}
\alias{read_region_LD_SNP_info}
\title{read all LD SNP info for all regions as a data frame
if region_info has files in SNP_info available, read from the files in SNP_info;
otherwise, use all SNP info from the LD reference.}
\usage{
read_region_LD_SNP_info(region_info, LD_ref_snpinfo = NULL, ncore = 1)
}
\arguments{
\item{region_info}{a data frame of region definition and associated file names}

\item{LD_ref_snpinfo}{a data frame of all variant info in the LD reference.}

\item{ncore}{The number of cores used to parallelize susie over regions}
}
\description{
read all LD SNP info for all regions as a data frame
if region_info has files in SNP_info available, read from the files in SNP_info;
otherwise, use all SNP info from the LD reference.
}
