% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_utils.R
\name{read_snp_info_file}
\alias{read_snp_info_file}
\title{Read a single SNP info file as a data frame}
\usage{
read_snp_info_file(file, snpinfo_loader_fun = NULL)
}
\arguments{
\item{file}{SNP info file name.}
}
\value{
a data frame of SNP info
}
\description{
Read a single SNP info file as a data frame
}
