% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_utils.R
\name{read_snp_info_files}
\alias{read_snp_info_files}
\title{Read multiple single SNP info files as a data frame}
\usage{
read_snp_info_files(files, snpinfo_loader_fun = NULL)
}
\arguments{
\item{files}{a vector of file names for SNP info in all regions}

\item{snpinfo_loader_fun}{a user defined function to load SNP information file,
if SNP information files are not in standard cTWAS reference format.}
}
\value{
a data frame of SNP info from all files
}
\description{
Read multiple single SNP info files as a data frame
}
