% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_read_data.R
\name{read_var_info}
\alias{read_var_info}
\title{Read variant information from .pvar or .bim file into R}
\usage{
read_var_info(var_info_file)
}
\arguments{
\item{var_info_file}{.pvar or .bim file with proper format:
.pvar: https://www.cog-genomics.org/plink/2.0/formats#pvar
.bim: https://www.cog-genomics.org/plink/2.0/formats#bim}
}
\value{
A data.table. variant info
}
\description{
Read variant information from .pvar or .bim file into R
}
