% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_utils.R
\name{read_weights}
\alias{read_weights}
\title{read weights in PredictDB format, do not perform harmonization on weights}
\usage{
read_weights(
  weight_files,
  chr = 1:22,
  ld_snpinfo,
  z_snp = NULL,
  scale_by_ld_variance = TRUE,
  ncore = 1
)
}
\description{
read weights in PredictDB format, do not perform harmonization on weights
}
