% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_process_regions.R
\name{region2core}
\alias{region2core}
\title{parallel regions}
\usage{
region2core(regionlist, ncore = 1)
}
\arguments{
\item{ncore}{integer, numeber of cores, at least 1
regions allocated to given number of cores
regionlist need to contain at least 1 non-empty}
}
\description{
parallel regions
}
