% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_postprocess.R
\name{rerun_finemap_regions_L1}
\alias{rerun_finemap_regions_L1}
\title{Rerun cTWAS finemapping using summary statistics with L = 1 for problematic regions}
\usage{
rerun_finemap_regions_L1(
  z_snp,
  z_gene,
  weight,
  problematic_snps = NULL,
  rerun_region_tags = NULL,
  pip_thresh = 0.5,
  ctwas_outputdir = NULL,
  ctwas_outname = NULL,
  ld_R_dir = NULL,
  outputdir = NULL,
  outname = NULL,
  group_prior = NULL,
  group_prior_var = NULL,
  use_null_weight = T,
  coverage = 0.95,
  min_abs_corr = 0.5,
  ncore = 1,
  logfile = NULL
)
}
\arguments{
\item{z_snp}{A data frame with four columns: "id", "A1", "A2", "z".
giving the z scores for snps. "A1" is effect allele. "A2" is the other allele.}

\item{z_gene}{A data frame with two columns: "id", "z". giving the z scores for genes.
Optionally, a "type" column can also be supplied; this is for using multiple sets of weights}

\item{weight}{a string, weight filename used in cTWAS}

\item{problematic_snps}{a character vector of problematic SNP rsIDs}

\item{rerun_region_tags}{a character vector of region tags to rerun}

\item{pip_thresh}{Minimum PIP value to select regions}

\item{ctwas_outputdir}{a string, the directory to cTWAS output}

\item{ctwas_outname}{a string, the directory to cTWAS output name}

\item{outputdir}{a string, the directory to store output}

\item{outname}{a string, the output name}

\item{group_prior}{a vector of two prior inclusion probabilities for SNPs and genes. This is ignored
if \code{estimate_group_prior = T}}

\item{group_prior_var}{a vector of two prior variances for SNPs and gene effects. This is ignored
if \code{estimate_group_prior_var = T}}

\item{use_null_weight}{TRUE/FALSE. If TRUE, allow for a probability of no effect in susie}

\item{coverage}{A number between 0 and 1 specifying the \dQuote{coverage} of the estimated confidence sets}

\item{min_abs_corr}{Minimum absolute correlation allowed in a
  credible set. The default, 0.5, corresponds to a squared
  correlation of 0.25, which is a commonly used threshold for
  genotype data in genetic studies.
#'}

\item{ncore}{The number of cores used to parallelize over regions}

\item{logfile}{the log file, if NULL will print log info on screen}

\item{LD_R_dir}{a string, pointing to a directory containing all LD matrix files and variant information.}
}
\description{
Rerun cTWAS finemapping using summary statistics with L = 1 for problematic regions
}
