% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_screen_regions.R
\name{screen_regions_noLD}
\alias{screen_regions_noLD}
\title{Screens regions with strong signals without using LD.}
\usage{
screen_regions_noLD(
  region_data,
  group_prior = NULL,
  group_prior_var = NULL,
  min_var = 2,
  min_gene = 1,
  min_nonSNP_PIP = 0.5,
  null_method = c("ctwas", "susie", "none"),
  null_weight = NULL,
  ncore = 1,
  logfile = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{region_data}{a list object indexing regions, variants and genes.}

\item{group_prior}{a vector of two prior inclusion probabilities for SNPs and genes.}

\item{group_prior_var}{a vector of two prior variances for SNPs and gene effects.}

\item{min_var}{minimum number of variables (SNPs and genes) in a region.}

\item{min_gene}{minimum number of genes in a region.}

\item{min_nonSNP_PIP}{If screening by non-SNP PIPs,
regions with total non-SNP PIP >= \code{min_nonSNP_PIP}
will be selected to run finemapping using full SNPs.}

\item{null_method}{Method to compute null model, options: "ctwas", "susie" or "none".}

\item{null_weight}{Prior probability of no effect (a number between
0 and 1, and cannot be exactly 1). Only used when \code{null_method = "susie"}.}

\item{ncore}{The number of cores used to parallelize susie over regions.}

\item{logfile}{The log filename. If NULL, print log info on screen.}

\item{verbose}{If TRUE, print detail messages.}

\item{...}{Additional arguments of \code{susie_rss}.}
}
\value{
a list, containing a data frame of selected region data,
and a data frame of screening summary for all regions.
}
\description{
Screens regions with strong signals without using LD.
}
