% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_compute_gene_z.R
\name{select_boundary_genes}
\alias{select_boundary_genes}
\title{Gets boundary genes and selects high PIP boundary genes}
\usage{
select_boundary_genes(
  region_info,
  weights,
  gene_ids,
  finemap_res,
  susie_alpha_res,
  combine_PIPs = TRUE,
  mapping_table = NULL,
  pip_thresh = 0.5,
  filter_cs = FALSE,
  ncore = 1
)
}
\arguments{
\item{region_info}{a data frame of region definitions.}

\item{weights}{a list of preprocessed weights.}

\item{gene_ids}{a vector of selected gene IDs (z_gene$id).
If specified, limits to these genes. Default: use all genes in weights.}

\item{finemap_res}{a data frame of original finemapping result.}

\item{susie_alpha_res}{a data frame of original susie alpha result.}

\item{combine_PIPs}{if TRUE, select boundary genes after combining gene PIPs.}

\item{mapping_table}{a data frame of mapping between molecular traits and genes,
with required columns: "molecular_id", "gene_name".}

\item{pip_thresh}{PIP cutoff for selecting boundary genes to merge regions.}

\item{filter_cs}{If TRUE, only select boundary genes in credible sets for region merge.}

\item{ncore}{The number of cores used to parallelize computation over regions}
}
\value{
a list with boundary genes and selected boundary genes
}
\description{
Gets boundary genes and selects high PIP boundary genes
}
