% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_process_regions.R
\name{select_highPIP_regions}
\alias{select_highPIP_regions}
\title{Select regions with high non-SNP PIPs}
\usage{
select_highPIP_regions(finemap_res, min_nonSNP_PIP = 0.5)
}
\description{
Select regions with high non-SNP PIPs
}
