% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_postprocess.R
\name{select_problematic_regions}
\alias{select_problematic_regions}
\title{Get regions with problematic high PIP SNPs or genes}
\usage{
select_problematic_regions(problematic_snps, highPIP_finemap_res, weights)
}
\arguments{
\item{problematic_snps}{a character vector of problematic SNP rsIDs}

\item{highPIP_finemap_res}{a data frame of the cTWAS finemapping result with high PIP SNPs and genes}

\item{weights}{weights}
}
\value{
a character vector of region ids with problematic high PIP SNPs or genes
}
\description{
Get regions with problematic high PIP SNPs or genes
}
