% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_summarize_finemap_res.R
\name{sum_pip_across_contexts}
\alias{sum_pip_across_contexts}
\title{Combines PIPs across contexts (tissues)}
\usage{
sum_pip_across_contexts(finemap_res, contexts)
}
\arguments{
\item{finemap_res}{a data frame of cTWAS finemapping result}

\item{contexts}{a character vector of contexts to sum}
}
\value{
a data frame of combined gene PIPs and PIPs for each context
}
\description{
Combines PIPs across contexts (tissues)
}
