% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_summarize_regions.R
\name{summarize_region_signals}
\alias{summarize_region_signals}
\title{Gets a basic summary of region signals}
\usage{
summarize_region_signals(region_data)
}
\arguments{
\item{region_data}{a list object indexing regions, variants and genes.}
}
\value{
a data frame of basic summary of region signals
}
\description{
Gets a basic summary of region signals
}
