% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_revise.R
\name{susie_get_cs}
\alias{susie_get_cs}
\title{susie get credible set. Revised from susieR::susie_get_cs}
\usage{
susie_get_cs(res, X = NULL, Xcorr = NULL, coverage = 0.95,
  min_abs_corr = 0.5, dedup = TRUE, squared = FALSE)
}
\description{
susie get credible set. Revised from susieR::susie_get_cs
}
