% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_revise.R
\name{susie_get_pip}
\alias{susie_get_pip}
\title{susie get PIP. Revised from susieR::susie_get_pip}
\usage{
susie_get_pip(res, prune_by_cs = FALSE, prior_tol = 1e-09)
}
\description{
susie get PIP. Revised from susieR::susie_get_pip
}
