% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_susieI_rss.R
\name{susieI_rss}
\alias{susieI_rss}
\title{SuSiE iteraitve rss version}
\usage{
susieI_rss(zdf, regionlist, ld_exprfs, ld_pgenfs = NULL, ld_Rfs = NULL,
  niter = 20, L = 1, z_ld_weight = 0, group_prior = NULL,
  group_prior_var = NULL, estimate_group_prior = T,
  estimate_group_prior_var = T, use_null_weight = T, coverage = 0.95,
  ncore = 1, outputdir = getwd(), outname = NULL)
}
\description{
SuSiE iteraitve rss version
}
