% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_regiondata.R
\name{trim_region_data}
\alias{trim_region_data}
\title{Remove SNPs from region_data if the total number of SNPs exceeds limit}
\usage{
trim_region_data(
  region_data,
  z_snp,
  trim_by = c("random", "z"),
  maxSNP = Inf,
  seed = 99
)
}
\description{
Remove SNPs from region_data if the total number of SNPs exceeds limit
}
