% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_get_regionlist.R
\name{trim_regionlist}
\alias{trim_regionlist}
\title{remove SNPs from regionlist if the total number of SNPs exceeds limit}
\usage{
trim_regionlist(
  regionlist,
  z_snp,
  trim_by = c("random", "z"),
  maxSNP = Inf,
  seed = 99
)
}
\description{
remove SNPs from regionlist if the total number of SNPs exceeds limit
}
