% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_weights.R
\name{trim_weights}
\alias{trim_weights}
\title{Trim weights by keeping top number of SNPs and/or filtering by minimum absolute weights.}
\usage{
trim_weights(
  weights,
  snp_map,
  top_n_snps = NULL,
  min_abs_weight = 0,
  ncore = 1
)
}
\arguments{
\item{weights}{a list of pre-processed prediction weights.}

\item{snp_map}{a list of data frames with SNP-to-region map for the reference.}

\item{top_n_snps}{an integer, only keeping the top n SNPs included in weight models.
By default, keep all SNPs in weights.}

\item{min_abs_weight}{a numeric value, only keeping the SNPs with abs(weight) > code{min_abs_weight}.
By default, keep all SNPs with abs(weight) > 0.}

\item{ncore}{The number of cores used to parallelize computation.}
}
\value{
Trimmed weights
}
\description{
Trim weights by keeping top number of SNPs and/or filtering by minimum absolute weights.
}
