% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susie_revise.R
\name{update_each_effect_rss}
\alias{update_each_effect_rss}
\title{update each effect rss (revised from susieR)}
\usage{
update_each_effect_rss(R, z, s_init, Sigma,
  estimate_prior_variance = FALSE, estimate_prior_method = "optim",
  check_null_threshold = 0)
}
\description{
update each effect rss (revised from susieR)
}
