% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_merge_regions.R
\name{update_merged_region_data_noLD}
\alias{update_merged_region_data_noLD}
\title{Updates cTWAS input data without LD with merged region data}
\usage{
update_merged_region_data_noLD(
  region_data,
  merged_region_data,
  region_info,
  merged_region_info,
  snp_map,
  merged_snp_map,
  merged_region_id_map
)
}
\arguments{
\item{region_data}{a list of original region data.}

\item{merged_region_data}{a list of merged region data.}

\item{region_info}{a data frame of original region definitions.}

\item{merged_region_info}{a data frame of original region definitions.}

\item{snp_map}{a list of original SNP info.}

\item{merged_snp_map}{a list of merged SNP info.}

\item{merged_region_id_map}{a data frame of new region IDs and original regions IDs.}
}
\value{
a list with updated region_data, region_info, snp_map.
}
\description{
Updates cTWAS input data without LD with merged region data
}
