% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_merge_regions.R
\name{update_merged_region_finemap_res}
\alias{update_merged_region_finemap_res}
\title{Updates cTWAS finemapping result for merged regions}
\usage{
update_merged_region_finemap_res(
  finemap_res,
  susie_alpha_res,
  merged_region_finemap_res,
  merged_region_susie_alpha_res,
  merged_region_id_map
)
}
\arguments{
\item{finemap_res}{a data frame of original finemapping result.}

\item{susie_alpha_res}{a data frame of original susie alpha result.}

\item{merged_region_finemap_res}{a data frame of finemapping result for merged regions.}

\item{merged_region_susie_alpha_res}{a data frame of susie alpha result for merged regions.}

\item{merged_region_id_map}{a data frame of new region IDs and original regions IDs.}
}
\value{
a list with updated cTWAS finemapping result.
}
\description{
Updates cTWAS finemapping result for merged regions
}
