% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_merge_regions.R
\name{update_merged_regions_finemap_res}
\alias{update_merged_regions_finemap_res}
\title{Update finemapping result for merged regions}
\usage{
update_merged_regions_finemap_res(
  finemap_res,
  finemap_merged_regions_res,
  merged_region_info
)
}
\arguments{
\item{finemap_res}{a data frame of finemapping result}

\item{finemap_merged_regions_res}{a data frame of finemapping result for merged regions}

\item{merged_region_info}{a data frame of merged region info}
}
\description{
Update finemapping result for merged regions
}
