% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_region_data.R
\name{update_region_z}
\alias{update_region_z}
\title{Adds or updates z-scores in region_data based on z_snp and z_gene.
this will also update sid and gid based on z_snp and z_gene.}
\usage{
update_region_z(
  region_data,
  z_snp,
  z_gene,
  update = c("all", "snps", "genes"),
  ncore = 1
)
}
\arguments{
\item{region_data}{a list of region gene IDs and SNP IDs and associated file names}

\item{z_snp}{A data frame with columns: "id", "z", giving the z-scores for SNPs.}

\item{z_gene}{A data frame with columns: "id", "z", giving the z-scores for genes.}

\item{update}{options to update z-scores in region data.
"all": update all data (default),
"snps": updates SNP data only,
"genes": updates gene data only.}

\item{ncore}{The number of cores used to parallelize susie over regions}
}
\description{
Adds or updates z-scores in region_data based on z_snp and z_gene.
this will also update sid and gid based on z_snp and z_gene.
}
