% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_update_regionlist.R
\name{update_regionlist_fullSNPs}
\alias{update_regionlist_fullSNPs}
\title{Update regionlist with full SNPs}
\usage{
update_regionlist_fullSNPs(
  regionlist,
  region_tags = NULL,
  select = NULL,
  maxSNP = Inf
)
}
\arguments{
\item{regionlist}{a list of region gene IDs and SNP IDs and associated file names}

\item{region_tags}{region tags to specify which regions to update. If NULL, update all regions in regionlist}

\item{select}{Default is NULL, all variants will be selected.
Or a data frame with columns id and z (id is for gene or SNP id, z is for z scores).
z will be used for remove SNPs if the total number of SNPs exceeds limit. See
parameter `maxSNP` for more information.}

\item{maxSNP}{Inf or integer. Maximum number of SNPs in a region. Default is
Inf, no limit. This can be useful if there are many SNPs in a region and you don't
have enough memory to run the program.}
}
\value{
updated regionlist with full SNPs
}
\description{
Update regionlist with full SNPs
}
