% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_read_data.R
\name{write_ld_Rf}
\alias{write_ld_Rf}
\title{combined variant information associated with a LD R matrix .RDS file.}
\usage{
write_ld_Rf(ld_R_dir, outname = outname, outputdir = getwd())
}
\arguments{
\item{ld_R_dir}{The directory that contains all ld R matrices.
the ld R matrices should not have overlapping positions.}

\item{outname}{a string, the output name}

\item{outputdir}{a string, the directory to store output}
}
\value{
A vector of the `ld_Rf` file names. The function will write one `ld_Rf` file
for each chromosome, so the vector has length 22. The `ld_Rf` file has the following
columns: chr region_name start stop RDS_file.
}
\description{
combined variant information associated with a LD R matrix .RDS file.
}
