% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_anno_finemap_res.R
\name{z2p}
\alias{z2p}
\title{convert z-scores to pvalues}
\usage{
z2p(z, neg_log10_p = FALSE)
}
\arguments{
\item{z}{a vector of z-scores}

\item{neg_log10_p}{If TRUE, returns -log10(p-values). Otherwise, returns p-values.}
}
\value{
a vector of p-values, or -log10(p-values) if neg_log10_p is TRUE.
}
\description{
convert z-scores to pvalues
}
