% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_region_data.R
\name{adjust_boundary_genes}
\alias{adjust_boundary_genes}
\title{Adjust region_data for boundary genes}
\usage{
adjust_boundary_genes(boundary_genes, weights, region_data, snp_info)
}
\description{
Adjust region_data for boundary genes
}
