% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_regionlist.R
\name{adjust_boundary_genes}
\alias{adjust_boundary_genes}
\title{adjust regionlist for boundary genes}
\usage{
adjust_boundary_genes(boundary_genes, region_info, weights, regionlist)
}
\description{
adjust regionlist for boundary genes
}
