% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_summarize_finemap_res.R
\name{anno_finemap_res}
\alias{anno_finemap_res}
\title{Annotate cTWAS finemapping result with ensembl gene annotation database}
\usage{
anno_finemap_res(
  finemap_res,
  snp_info,
  ens_db,
  use_gene_pos = c("mid", "start", "end")
)
}
\arguments{
\item{finemap_res}{a data frame of cTWAS finemapping result}

\item{snp_info}{a list or data frame of SNP info for LD reference,
with columns "chrom", "id", "pos", "alt", "ref".}

\item{ens_db}{ensembl gene annotation database}

\item{use_gene_pos}{use mid (midpoint), start or end positions to represent gene positions}
}
\value{
a data frame of cTWAS finemapping result including gene
names and updated positions
}
\description{
Annotate cTWAS finemapping result with ensembl gene annotation database
}
