% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_summarize_finemap_res.R
\name{anno_finemap_res}
\alias{anno_finemap_res}
\title{Annotates cTWAS finemapping result with gene annotations}
\usage{
anno_finemap_res(
  finemap_res,
  snp_map,
  gene_annot = NULL,
  use_gene_pos = c("mid", "start", "end"),
  filter_protein_coding_genes = FALSE,
  filter_cs = FALSE
)
}
\arguments{
\item{finemap_res}{a data frame of cTWAS finemapping result}

\item{snp_map}{a list of data frames with SNP-to-region map for the reference.}

\item{gene_annot}{a data frame of gene annotations, with columns:
"chrom", "start", "end", "gene_id", "gene_name", "gene_type.}

\item{use_gene_pos}{use mid (midpoint), start or end positions to
represent gene positions.}

\item{filter_protein_coding_genes}{TRUE/FALSE. If TRUE, keep protein coding genes only.}

\item{filter_cs}{TRUE/FALSE. If TRUE, limits results in credible sets.}
}
\value{
a data frame of cTWAS finemapping result including gene
names, types and positions
}
\description{
Annotates cTWAS finemapping result with gene annotations
}
