% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_susie_rss.R
\name{anno_susie}
\alias{anno_susie}
\title{annotate susie results with SNP and gene information}
\usage{
anno_susie(
  susie_res,
  gid,
  sid,
  g_type,
  g_QTLtype,
  region_tag,
  geneinfo = NULL,
  snpinfo = NULL,
  include_cs_index = TRUE
)
}
\description{
annotate susie results with SNP and gene information
}
