% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_region_data.R
\name{assemble_region_data}
\alias{assemble_region_data}
\title{Assemble data for all the regions}
\usage{
assemble_region_data(
  region_info,
  z_snp,
  z_gene,
  weights,
  LD_snpinfo,
  thin = 0.1,
  maxSNP = Inf,
  trim_by = c("random", "z"),
  adjust_boundary_genes = TRUE,
  thin_gwas_snps = TRUE,
  add_z = TRUE,
  ncore = 1,
  seed = 99
)
}
\arguments{
\item{region_info}{a data frame of region definition and associated file names}

\item{z_snp}{A data frame with columns: "id", "z", giving the z-scores for SNPs.}

\item{z_gene}{A data frame with columns: "id", "z", giving the z-scores for genes.}

\item{weights}{a list of weights}

\item{LD_snpinfo}{a data frame, SNP info for LD reference,
with required columns "chrom", "id", "pos", and "region_id".}

\item{thin}{The proportion of SNPs to be used for the parameter estimation and
initial screening region steps.
Smaller \code{thin} parameters reduce runtime at the expense of accuracy.
The fine mapping step is rerun using full SNPs for regions with strong gene signals.}

\item{maxSNP}{Inf or integer. Maximum number of SNPs in a region. Default is
Inf, no limit. This can be useful if there are many SNPs in a region and you don't
have enough memory to run the program. This applies to the last rerun step
(using full SNPs and rerun susie for regions with strong gene signals) only.}

\item{trim_by}{remove SNPs if the total number of SNPs exceeds limit, options: "random",
or "z" (trim SNPs with lower |z|) See parameter `maxSNP` for more information.}

\item{adjust_boundary_genes}{identify cross-boundary genes, adjust region_data and update weighs}

\item{thin_gwas_snps}{TRUE/FALSE, if TRUE, only apply thin to GWAS SNPs, Otherwise, apply thins to all SNPs.}

\item{add_z}{TRUE/FALSE, if TRUE, add z-scores to the region_data}

\item{ncore}{The number of cores used to parallelize susie over regions}

\item{seed}{seed for random sampling}
}
\value{
a list with region_data, updated weights, and cross-bounary genes
}
\description{
Assemble data for all the regions
}
