% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_region_data.R
\name{assign_region_ids}
\alias{assign_region_ids}
\title{assign gene and SNP IDs for regions in the regioninfo}
\usage{
assign_region_ids(
  regioninfo,
  geneinfo,
  snpinfo,
  thin = 0.1,
  thin_gwas_snps = TRUE,
  seed = 99
)
}
\description{
assign gene and SNP IDs for regions in the regioninfo
}
