% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_regionlist.R
\name{assign_region_ids}
\alias{assign_region_ids}
\title{assign gene and SNP IDs for regions in the regioninfo}
\usage{
assign_region_ids(regioninfo, geneinfo, snpinfo, minvar = 1, mingene = 0)
}
\description{
assign gene and SNP IDs for regions in the regioninfo
}
