% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_summarize_finemap_res.R
\name{combine_gene_pips}
\alias{combine_gene_pips}
\title{Combines gene PIPs by context, type or group.}
\usage{
combine_gene_pips(
  annotated_finemap_res,
  by = c("context", "type", "group"),
  gene_col = "gene_name",
  missing_value = NA
)
}
\arguments{
\item{annotated_finemap_res}{a data frame of annotated cTWAS finemapping result}

\item{by}{sum gene PIPs by "context", "type", or "group".}

\item{gene_col}{column name of target genes to combine gene PIPs.}

\item{missing_value}{set missing value (default: NA)}
}
\value{
a data frame of combined gene PIPs for each context, type or group
}
\description{
Combines gene PIPs by context, type or group.
}
