% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_compute_gene_z.R
\name{combine_z}
\alias{combine_z}
\title{Combine SNP and gene z-scores}
\usage{
combine_z(z_snp, z_gene)
}
\arguments{
\item{z_snp}{a data frame with four columns: "id", "A1", "A2", "z". ("A1" and "A2" are optional)}

\item{z_gene}{a data frame with columns: "id", "z", "type", "context", "group".}
}
\description{
Combine SNP and gene z-scores
}
