% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_est_parameters.R
\name{compute_enrichment_test}
\alias{compute_enrichment_test}
\title{Computes standard error and p-value for enrichment}
\usage{
compute_enrichment_test(
  group_prior,
  group_size,
  enrichment_test = c("G", "fisher")
)
}
\arguments{
\item{group_prior}{a vector of prior inclusion probabilities for different groups.}

\item{group_size}{a vector of number of variables in different groups.}

\item{enrichment_test}{Method to test enrichment,
"G": G-test, "fisher": Fisher's exact test.}
}
\value{
Estimated enrichment, S.E. and p-value from G-test or Fisher's exact test.
}
\description{
Computes standard error and p-value for enrichment
}
