% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_est_parameters.R
\name{compute_enrichment_test}
\alias{compute_enrichment_test}
\title{Computes enrichment (log-scale), standard error and p-value}
\usage{
compute_enrichment_test(
  group_prior,
  group_size,
  enrichment_test = c("G", "fisher"),
  include_test_result = FALSE
)
}
\arguments{
\item{group_prior}{a vector of prior inclusion probabilities for different groups.}

\item{group_size}{a vector of number of variables in different groups.}

\item{enrichment_test}{Method to test enrichment,
"G": G-test, "fisher": Fisher's exact test.}

\item{include_test_result}{If TRUE, return the original test result.}
}
\value{
Estimated enrichment, S.E. and p-value from G-test or Fisher's exact test.
}
\description{
Computes enrichment (log-scale), standard error and p-value
}
