% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_compute_gene_z.R
\name{compute_gene_z}
\alias{compute_gene_z}
\title{Compute gene z-scores}
\usage{
compute_gene_z(
  z_snp,
  region_info,
  weight_list,
  weight_info,
  ncore = 1,
  logfile = NULL
)
}
\arguments{
\item{z_snp}{A data frame with columns: "id", "A1", "A2", "z". giving the z scores for
snps. "A1" is effect allele. "A2" is the other allele. For harmonized data, A1 and A2 are not required.}

\item{region_info}{a data frame of region definition and associated file names.}

\item{weight_list}{a list of weights}

\item{weight_info}{a data frame of weight information}

\item{ncore}{The number of cores used to parallelize imputation over weights}

\item{logfile}{the log file, if NULL will print log info on screen}
}
\value{
a list of gene z-scores and gene info table
}
\description{
Compute gene z-scores
}
