% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_compute_region_cor.R
\name{compute_region_cor}
\alias{compute_region_cor}
\title{Compute correlation matrices for a region}
\usage{
compute_region_cor(regionlist, region_tag, weight_list)
}
\arguments{
\item{regionlist}{regionlist}

\item{region_tag}{region tag}

\item{weight_list}{weight list}
}
\value{
a list of correlation matrices
}
\description{
Compute correlation matrices for a region
}
