% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_compute_region_cor.R
\name{compute_region_cor}
\alias{compute_region_cor}
\title{Compute correlation matrices for a region}
\usage{
compute_region_cor(
  R_snp,
  ld_snpinfo,
  region_idx,
  wgtlist,
  save = FALSE,
  outputdir = NULL,
  outname = NULL,
  verbose = FALSE
)
}
\arguments{
\item{R_snp}{Reference LD matrix}

\item{ld_snpinfo}{SNP information of the SNPs in \code{R_snp} matrix}

\item{region_idx}{region indices of the region from regionlist}

\item{wgtlist}{a list of weights for genes in the region}

\item{save}{TRUE/FALSE, if TRUE, save correlation matrices in \code{outputdir}}

\item{outputdir}{directory of correlation matrix result}

\item{outname}{filename of correlation result}

\item{verbose}{TRUE/FALSE, if TRUE,}
}
\value{
a list of correlation matrices (R_snp, R_snp_gene, R_gene), SNP indices and gene id
}
\description{
Compute correlation matrices for a region
}
