% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_region_cor.R
\name{compute_region_cor}
\alias{compute_region_cor}
\title{Computes correlation matrices for a single region}
\usage{
compute_region_cor(sids, gids, R_snp, LD_sids, weights)
}
\arguments{
\item{sids}{SNP IDs in the region}

\item{gids}{gene IDs in the region}

\item{R_snp}{LD (R) matrix}

\item{LD_sids}{SNP IDs for the rows and columns of the LD matrix}

\item{weights}{a list of weights for all the genes}
}
\value{
a list of correlation matrices (R_snp, R_snp_gene and R_gene)
}
\description{
Computes correlation matrices for a single region
}
