% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_screen_regions.R
\name{compute_region_nonSNP_PIPs}
\alias{compute_region_nonSNP_PIPs}
\title{Computes non-SNP PIPs for each region from finemapping result}
\usage{
compute_region_nonSNP_PIPs(finemap_res)
}
\arguments{
\item{finemap_res}{a data frame of finemapping result}
}
\description{
Computes non-SNP PIPs for each region from finemapping result
}
