% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_screen_regions.R
\name{compute_region_nonSNP_PIPs}
\alias{compute_region_nonSNP_PIPs}
\title{Compute non-SNP PIPs for regions}
\usage{
compute_region_nonSNP_PIPs(finemap_res)
}
\arguments{
\item{finemap_res}{a data frame of finemapping result}
}
\description{
Compute non-SNP PIPs for regions
}
