% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_weights.R
\name{convert_fusion_to_predictdb}
\alias{convert_fusion_to_predictdb}
\title{Converts fusion weights to predictDB format}
\usage{
convert_fusion_to_predictdb(
  weight_dir,
  fusion_method = c("lasso", "enet", "top1", "blup", "bslmm", "best.cv"),
  fusion_genome_version = NA,
  make_extra_table = TRUE,
  cov_table = NULL,
  outputdir = getwd(),
  outname
)
}
\arguments{
\item{weight_dir}{the directory containing FUSION weights ('.wgt.RDat' files).}

\item{fusion_method}{a string, specifying the method to choose in FUSION models.
"best.cv" option will use the best model (smallest p-value) under cross-validation.}

\item{fusion_genome_version}{a string, specifying the genome version of FUSION models}

\item{make_extra_table}{If TRUE, make an extra table in predictDB format}

\item{cov_table}{a data frame of covariances between variants, with columns:
"GENE","RSID1","RSID2","VALUE".
If NULL, do not create covariance files (.txg.gz).}

\item{outputdir}{output directory}

\item{outname}{name of the output weight file}
}
\description{
Converts fusion weights to predictDB format
}
