% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_utils.R
\name{convert_to_ukb_varIDs}
\alias{convert_to_ukb_varIDs}
\title{Convert variant IDs from Open GWAS format or PredictDB weight format}
\usage{
convert_to_ukb_varIDs(varIDs)
}
\arguments{
\item{varIDs}{a vector of variant IDs from Open GWAS format
("chr_pos_ref_alt") or PredictDB weight format
(“chr_pos_ref_alt_build” or “chr:pos_ref_alt_build”)}
}
\value{
a vector of variant IDs in the format of "chr:pos_ref_alt"
}
\description{
Convert variant IDs from Open GWAS format or PredictDB weight format
}
