% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas.R
\name{ctwas}
\alias{ctwas}
\title{Causal inference for TWAS}
\usage{
ctwas(pgenfs, exprfs, Y, ld_regions = c("EUR", "ASN", "AFR"),
  ld_regions_custom = NULL, thin = 1, prob_single = 0.8,
  max_snp_region = Inf, rerun_gene_PIP = 0.8, niter1 = 3,
  niter2 = 30, L = 5, group_prior = NULL, group_prior_var = NULL,
  estimate_group_prior = T, estimate_group_prior_var = T,
  use_null_weight = T, coverage = 0.95, stardardize = T, ncore = 1,
  outputdir = getwd(), outname = NULL, logfile = NULL)
}
\arguments{
\item{pgenfs}{A character vector of .pgen or .bed files. One file for one
chromosome, in the order of 1 to 22. Therefore, the length of this vector
needs to be 22. If .pgen files are given, then .pvar and .psam are assumed
to present in the same directory. If .bed files are given, then .bim and
.fam files are assumed to present in the same directory.}

\item{exprfs}{A character vector of .`expr` or `.expr.gz` files. One file for
one chromosome, in the order of 1 to 22. Therefore, the length of this vector
needs to be 22.  `.expr.gz` file is gzip compressed `.expr` files. `.expr` is
a matrix of imputed expression values, row is for each sample, column is for
each gene. Its sample order is same as in files provided by `.pgenfs`. We also
assume corresponding `.exprvar` files are present in the same directory.
`.exprvar` files are just tab delimited text files, with columns:
\describe{
  \item{chrom}{chromosome number, numeric}
  \item{p0}{gene boundary position, the smaller value}
  \item{p1}{gene boundary position, the larger value}
  \item{id}{gene id}
}
Its rows should be in the same order as the columns for corresponding `.expr`
files.}

\item{Y}{a vector of length n, phenotype, the same order as provided
by `.pgenfs` (defined in .psam or .fam files).}

\item{ld_regions}{A string representing the population to use for defining
LD regions. These LD regions are defined by ldetect. The user can also
provide custom LD regions matching genotype data, see
\code{ld_regions_custom}.}

\item{ld_regions_custom}{A bed format file defining LD regions. The default
is \code{NULL}; when specified, \code{ld_regions} will be ignored.}

\item{prob_single}{blocks with probility more than this value will be
used for parameter estimation}

\item{max_snp_region}{Inf or integer. Maximum number of SNPs in a region. Default is
Inf, no limit. This can be useful if there are many SNPs in a region and you don't
have enough memory to run the program. This applies to the last rerun step
 (using full SNPs and rerun susie for regions with strong gene signals) only.}

\item{rerun_gene_PIP}{if thin <1, will rerun blocks with the max gene PIP
> \code{rerun_gene_PIP} using full SNPs. if \code{rerun_gene_PIP} is 0, then
all blocks will rerun with full SNPs}

\item{outname}{a string, the output name}

\item{logfile}{the log file, if NULL will print log info on screen}

\item{harmonize}{TRUE/FALSE. If TRUE, will harmonize GWAS, LD reference
and weight internally (flip alleles and remove strand ambiguous SNPs).}
}
\description{
Causal inference for TWAS
}
