% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_sumstats.R
\name{ctwas_sumstats}
\alias{ctwas_sumstats}
\title{run cTWAS analysis using summary statistics}
\usage{
ctwas_sumstats(
  z_snp,
  weights,
  region_info,
  LD_map,
  snp_map,
  z_gene,
  thin = 0.1,
  niter_prefit = 3,
  niter = 30,
  L = 5,
  init_group_prior = NULL,
  init_group_prior_var = NULL,
  filter_L = TRUE,
  filter_nonSNP_PIP = FALSE,
  min_nonSNP_PIP = 0.5,
  min_p_single_effect = 0.8,
  maxSNP = Inf,
  use_null_weight = TRUE,
  coverage = 0.95,
  min_abs_corr = 0.5,
  LD_format = c("rds", "rdata", "mtx", "csv", "txt", "custom"),
  LD_loader_fun,
  ncore = 1,
  ncore_LD = max(ncore - 1, 1),
  force_compute_cor = FALSE,
  save_cor = FALSE,
  cor_dir = NULL,
  outputdir = NULL,
  outname = "ctwas",
  logfile = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{z_snp}{A data frame with four columns: "id", "A1", "A2", "z".
giving the z scores for snps. "A1" is effect allele. "A2" is the other allele.}

\item{weights}{a list of pre-processed prediction weights}

\item{region_info}{a data frame of region definitions.}

\item{LD_map}{a data frame with filenames of LD matrices for each of the regions.}

\item{snp_map}{a list of data frames with SNP-to-region map for the reference.}

\item{z_gene}{A data frame with columns: "id", "z", giving the z-scores for genes.}

\item{thin}{The proportion of SNPs to be used for estimating parameters and screening regions.}

\item{niter_prefit}{the number of iterations of the E-M algorithm to perform during the initial parameter estimation step}

\item{niter}{the number of iterations of the E-M algorithm to perform during the complete parameter estimation step}

\item{L}{the number of effects for susie during the fine mapping steps}

\item{init_group_prior}{a vector of initial values of prior inclusion probabilities for SNPs and genes.}

\item{init_group_prior_var}{a vector of initial values of prior variances for SNPs and gene effects.}

\item{filter_L}{If TRUE, screening regions with L >= 1}

\item{filter_nonSNP_PIP}{If TRUE, screening regions with total non-SNP PIP >= \code{min_nonSNP_PIP}}

\item{min_nonSNP_PIP}{Regions with non-SNP PIP >= \code{min_nonSNP_PIP}
will be selected to run finemapping using full SNPs.}

\item{min_p_single_effect}{Regions with probability >= \code{min_p_single_effect}
of having at most one causal effect will be selected for the final EM step.}

\item{maxSNP}{Inf or integer. Maximum number of SNPs in a region. Default is
Inf, no limit. This can be useful if there are many SNPs in a region and you don't
have enough memory to run the program.}

\item{use_null_weight}{TRUE/FALSE. If TRUE, allow for a probability of no effect in susie}

\item{coverage}{A number between 0 and 1 specifying the \dQuote{coverage} of the estimated confidence sets}

\item{min_abs_corr}{Minimum absolute correlation allowed in a
.  credible set. The default, 0.5, corresponds to a squared
  correlation of 0.25, which is a commonly used threshold for
  genotype data in genetic studies.}

\item{LD_format}{file format for LD matrix. If "custom", use a user defined
\code{LD_loader_fun()} function to load LD matrix.}

\item{LD_loader_fun}{a user defined function to load LD matrix when \code{LD_format = "custom"}.}

\item{ncore}{The number of cores used to parallelize computing over regions.}

\item{ncore_LD}{The number of cores used to parallelize computing correlation matrices,
in screening regions and fine-mapping steps with LD.}

\item{force_compute_cor}{If TRUE, force computing correlation (R) matrices}

\item{save_cor}{TRUE/FALSE. If TRUE, save correlation (R) matrices to \code{cor_dir}}

\item{cor_dir}{The directory to store correlation (R) matrices}

\item{outputdir}{The directory to store output. If specified, save outputs to the directory.}

\item{outname}{The output name.}

\item{logfile}{path to the log file, if NULL will print log info on screen.}

\item{verbose}{TRUE/FALSE. If TRUE, print detailed messages}

\item{...}{Additional arguments of \code{susie_rss}.}
}
\value{
a list, include z_gene, estimated parameters, region_data,
cross-boundary genes, screening region results, and fine-mapping results.
}
\description{
run cTWAS analysis using summary statistics
}
