% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_sumstats.R
\name{ctwas_sumstats}
\alias{ctwas_sumstats}
\title{cTWAS analysis using summary statistics}
\usage{
ctwas_sumstats(
  z_snp,
  weights,
  region_info,
  snp_info,
  LD_info,
  z_gene = NULL,
  thin = 0.1,
  niter_prefit = 3,
  niter = 30,
  L = 5,
  init_group_prior = NULL,
  init_group_prior_var = NULL,
  group_prior_var_structure = c("shared_type", "shared_context", "shared_nonSNP",
    "shared_all", "independent"),
  filter_L = TRUE,
  filter_nonSNP_PIP = FALSE,
  min_nonSNP_PIP = 0.5,
  p_single_effect = 0.8,
  maxSNP = 20000,
  use_null_weight = TRUE,
  coverage = 0.95,
  min_abs_corr = 0.5,
  LD_format = c("rds", "rdata", "mtx", "csv", "txt", "custom"),
  LD_loader_fun,
  ncore = 1,
  save_cor = FALSE,
  cor_dir = NULL,
  outputdir = NULL,
  outname = "ctwas",
  logfile = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{z_snp}{A data frame with four columns: "id", "A1", "A2", "z".
giving the z scores for snps. "A1" is effect allele. "A2" is the other allele.}

\item{weights}{a list of prediction weights}

\item{region_info}{a data frame of region definitions.}

\item{snp_info}{a list of SNP info data frames for LD reference.}

\item{LD_info}{a list of paths to LD matrices for each of the regions.}

\item{z_gene}{A data frame with columns: "id", "z", giving the z-scores for genes.}

\item{thin}{The proportion of SNPs to be used for estimating parameters and screening regions.}

\item{niter_prefit}{the number of iterations of the E-M algorithm to perform during the initial parameter estimation step}

\item{niter}{the number of iterations of the E-M algorithm to perform during the complete parameter estimation step}

\item{L}{the number of effects for susie during the fine mapping steps}

\item{init_group_prior}{a vector of initial values of prior inclusion probabilities for SNPs and genes.}

\item{init_group_prior_var}{a vector of initial values of prior variances for SNPs and gene effects.}

\item{group_prior_var_structure}{a string indicating the structure to put on the prior variance parameters.
"shared_type" allows all groups in one molecular QTL type to share the same variance parameter.
"shared_context" allows all groups in one context (tissue, cell type, condition) to share the same variance parameter.
"shared_nonSNP" allows all non-SNP groups to share the same variance parameter.
"shared_all" allows all groups to share the same variance parameter.
"independent" allows all groups to have their own separate variance parameters.}

\item{filter_L}{If TRUE, screening regions with L >= 1}

\item{filter_nonSNP_PIP}{If TRUE, screening regions with total non-SNP PIP >= \code{min_nonSNP_PIP}}

\item{min_nonSNP_PIP}{Regions with non-SNP PIP >= \code{min_nonSNP_PIP}
will be selected to run finemapping using full SNPs.}

\item{p_single_effect}{Regions with probability greater than \code{p_single_effect} of
having 1 or fewer effects will be used for parameter estimation}

\item{maxSNP}{Inf or integer. Maximum number of SNPs in a region. Default is
Inf, no limit. This can be useful if there are many SNPs in a region and you don't
have enough memory to run the program. This applies to the finemapping step only.}

\item{use_null_weight}{TRUE/FALSE. If TRUE, allow for a probability of no effect in susie}

\item{coverage}{A number between 0 and 1 specifying the \dQuote{coverage} of the estimated confidence sets}

\item{min_abs_corr}{Minimum absolute correlation allowed in a
.  credible set. The default, 0.5, corresponds to a squared
  correlation of 0.25, which is a commonly used threshold for
  genotype data in genetic studies.}

\item{LD_format}{file format for LD matrix. If "custom", use a user defined
\code{LD_loader_fun()} function to load LD matrix.}

\item{LD_loader_fun}{a user defined function to load LD matrix when \code{LD_format = "custom"}.}

\item{ncore}{The number of cores used to parallelize susie over regions}

\item{save_cor}{TRUE/FALSE. If TRUE, save correlation (R) matrices to \code{cor_dir}}

\item{cor_dir}{The directory to store correlation (R) matrices}

\item{outputdir}{The directory to store output. If specified, save outputs to the directory.}

\item{outname}{The output name.}

\item{logfile}{path to the log file, if NULL will print log info on screen.}

\item{verbose}{TRUE/FALSE. If TRUE, print detailed messages}

\item{...}{Additional arguments of \code{susie_rss}.}
}
\value{
a list of estimated parameters, fine-mapping results, boundary genes,
z_gene, region_data, and screened_region_data
}
\description{
cTWAS analysis using summary statistics
}
