% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_sumstats_noLD.R
\name{ctwas_sumstats_noLD}
\alias{ctwas_sumstats_noLD}
\title{cTWAS analysis using summary statistics with "no LD" version}
\usage{
ctwas_sumstats_noLD(
  z_snp,
  weights,
  region_info,
  snp_map,
  z_gene = NULL,
  thin = 0.1,
  niter_prefit = 3,
  niter = 30,
  init_group_prior = NULL,
  init_group_prior_var = NULL,
  group_prior_var_structure = c("shared_all", "shared_type", "shared_context",
    "shared_nonSNP", "independent"),
  maxSNP = Inf,
  min_var = 2,
  min_gene = 1,
  min_group_size = 100,
  min_nonSNP_PIP = 0.5,
  min_pval = 5e-08,
  min_p_single_effect = 0.8,
  null_method = c("ctwas", "susie", "none"),
  null_weight = NULL,
  include_enrichment_test = TRUE,
  enrichment_test = c("G", "fisher"),
  include_loglik = FALSE,
  outputdir = NULL,
  outname = "ctwas_noLD",
  ncore = 1,
  seed = 99,
  logfile = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{z_snp}{A data frame with four columns: "id", "A1", "A2", "z".
giving the z scores for SNPs. "A1" is effect allele. "A2" is the other allele.}

\item{weights}{a list of pre-processed prediction weights.}

\item{region_info}{a data frame of region definitions.}

\item{snp_map}{a list of data frames with SNP-to-region map for the reference.}

\item{z_gene}{A data frame with columns: "id", "z", giving the z-scores for genes.}

\item{thin}{The proportion of SNPs to be used for estimating parameters and screening regions.}

\item{niter_prefit}{the number of iterations of the E-M algorithm to perform during the initial parameter estimation step.}

\item{niter}{the number of iterations of the E-M algorithm to perform during the complete parameter estimation step.}

\item{init_group_prior}{a vector of initial values of prior inclusion probabilities for SNPs and genes.}

\item{init_group_prior_var}{a vector of initial values of prior variances for SNPs and gene effects.}

\item{group_prior_var_structure}{a string indicating the structure to put on the prior variance parameters.
"shared_all" allows all groups to share the same variance parameter.
"shared_type" allows all groups in one molecular QTL type to share the same variance parameter.
"shared_context" allows all groups in one context (tissue, cell type, condition) to share the same variance parameter.
"shared_nonSNP" allows all non-SNP groups to share the same variance parameter.
"independent" allows all groups to have their own separate variance parameters.}

\item{maxSNP}{Inf or integer. Maximum number of SNPs in a region. Default is
Inf, no limit. This can be useful if there are many SNPs in a region and you don't
have enough memory to run the program.}

\item{min_var}{minimum number of variables (SNPs and genes) in a region
when estimating paramters and screening regions.}

\item{min_gene}{minimum number of genes in a region
when estimating paramters and screening regions.}

\item{min_group_size}{Minimum number of genes in a group.
Groups with number of genes < \code{min_group_size} will be removed for the analysis.}

\item{min_nonSNP_PIP}{Regions with non-SNP PIP >= \code{min_nonSNP_PIP}
will be selected to run finemapping using full SNPs.}

\item{min_pval}{Keep regions with minimum p-values from z_snp and z_gene < \code{min_pval}.}

\item{min_p_single_effect}{Regions with probability greater than \code{min_p_single_effect} of
having 1 or fewer effects will be used for parameter estimation.}

\item{null_method}{Method to compute null model, options: "ctwas", "susie" or "none".}

\item{null_weight}{Prior probability of no effect (a number between
0 and 1, and cannot be exactly 1). Only used when \code{null_method = "susie"}.}

\item{outputdir}{The directory to store output. If specified, save outputs to the directory.}

\item{outname}{The output name.}

\item{ncore}{The number of cores used to parallelize susie over regions.}

\item{seed}{seed for random sampling when thinning the SNPs in region data.}

\item{logfile}{The log filename. If NULL, print log info on screen.}

\item{verbose}{If TRUE, print detailed messages.}

\item{...}{Additional arguments of \code{susie_rss}.}
}
\value{
a list, including z_gene, estimated parameters, region_data,
cross-boundary genes, screening region results, and fine-mapping results.
}
\description{
cTWAS analysis using summary statistics with "no LD" version
}
