% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_preprocess_z_snp.R
\name{detect_ld_mismatch_susie}
\alias{detect_ld_mismatch_susie}
\title{Detect LD mismatches using SuSiE RSS}
\usage{
detect_ld_mismatch_susie(
  z_snp,
  region_info,
  gwas_n = NULL,
  ncore = 1,
  p_diff_thresh = 5e-08
)
}
\arguments{
\item{z_snp}{A data frame with two columns: "id", "A1", "A2", "z". giving the z scores for
snps. "A1" is effect allele. "A2" is the other allele.}

\item{gwas_n}{integer, GWAS sample size}

\item{ncore}{integer, number of cores for parallel computing.}

\item{p_diff_thresh}{numeric, p-value threshold for identifying problematic SNPs
with significant difference between observed z-scores and estimated values}

\item{ld_Rinfo}{a vector of paths to the variant information for all LD matrices}
}
\description{
Detect LD mismatches using SuSiE RSS
}
