% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_est_parameters.R
\name{est_param}
\alias{est_param}
\title{Estimate cTWAS parameters}
\usage{
est_param(
  region_data,
  init_group_prior = NULL,
  init_group_prior_var = NULL,
  group_prior_var_structure = c("independent", "shared_nonSNP", "shared_all",
    "shared_type", "shared_context"),
  niter_prefit = 3,
  niter = 30,
  p_single_effect = 0.8,
  ncore = 1,
  logfile = NULL,
  verbose = FALSE
)
}
\arguments{
\item{region_data}{a list object indexing regions, variants and genes.}

\item{init_group_prior}{a vector of initial values of prior inclusion probabilities for SNPs and genes.}

\item{init_group_prior_var}{a vector of initial values of prior variances for SNPs and gene effects.}

\item{group_prior_var_structure}{a string indicating the structure to put on the prior variance parameters.
"independent" is the default and allows all groups to have their own separate variance parameters.
"shared_nonSNP" allows all non-SNP groups to share the same variance parameter.
"shared_all" allows all groups to share the same variance parameter.
"shared_type" allows all groups in one molecular QTL type to share the same variance parameter.
"shared_context" allows all groups in one molecular QTL type to share the same variance parameter.}

\item{niter_prefit}{the number of iterations of the E-M algorithm to perform during the initial parameter estimation step}

\item{niter}{the number of iterations of the E-M algorithm to perform during the complete parameter estimation step}

\item{p_single_effect}{Regions with probability greater than \code{p_single_effect} of
having at most one causal effect will be used selected for the complete parameter estimation step}

\item{ncore}{The number of cores used to parallelize computation over regions}

\item{logfile}{the log file, if NULL will print log info on screen}

\item{verbose}{TRUE/FALSE. If TRUE, print detail messages}
}
\value{
a list with estimated parameters
}
\description{
Estimate cTWAS parameters
}
