% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_regionlist.R
\name{expand_regionlist}
\alias{expand_regionlist}
\title{Expand regionlist with full SNPs}
\usage{
expand_regionlist(
  regionlist,
  region_info,
  z_snp,
  z_gene,
  trim_by = c("z", "random"),
  maxSNP = Inf,
  ncore = 1,
  seed = 99
)
}
\arguments{
\item{regionlist}{a list of region gene IDs and SNP IDs and associated file names}

\item{region_info}{a data frame of region definition and associated LD file names}

\item{z_snp}{A data frame with columns: "id", "z", giving the z-scores for SNPs.}

\item{z_gene}{A data frame with columns: "id", "z", giving the z-scores for genes.}

\item{trim_by}{remove SNPs if the total number of SNPs exceeds limit, options: "random",
or "z" (trim SNPs with lower |z|) See parameter `maxSNP` for more information.}

\item{maxSNP}{Inf or integer. Maximum number of SNPs in a region. Default is
Inf, no limit. This can be useful if there are many SNPs in a region and you don't
have enough memory to run the program.}

\item{ncore}{The number of cores used to parallelize susie over regions}

\item{seed}{seed for random sampling}
}
\value{
updated regionlist with full SNPs
}
\description{
Expand regionlist with full SNPs
}
