% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_finemapping.R
\name{finemap_region}
\alias{finemap_region}
\title{Runs cTWAS finemapping for a single region}
\usage{
finemap_region(
  region_data,
  region_id,
  use_LD = TRUE,
  LD_info = NULL,
  snp_info = NULL,
  weights = NULL,
  L = 5,
  group_prior = NULL,
  group_prior_var = NULL,
  groups = NULL,
  use_null_weight = TRUE,
  coverage = 0.95,
  min_abs_corr = 0.5,
  force_compute_cor = FALSE,
  save_cor = FALSE,
  cor_dir = NULL,
  include_cs_index = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{region_data}{a list object with data for the regions}

\item{region_id}{a character string of region id to be finemapped}

\item{use_LD}{TRUE/FALSE. If TRUE, use LD for finemapping. Otherwise, use "no-LD" version.}

\item{LD_info}{a list of paths to LD matrices for each of the regions. Required when \code{use_LD = TRUE}.}

\item{snp_info}{a list of SNP info data frames for LD reference. Required when \code{use_LD = TRUE}.}

\item{weights}{a list of preprocessed weights}

\item{L}{the number of effects for susie during the fine mapping steps}

\item{group_prior}{a vector of two prior inclusion probabilities for SNPs and genes.}

\item{group_prior_var}{a vector of two prior variances for SNPs and gene effects.}

\item{groups}{a vector of group names for group_prior and group_prior_var.}

\item{use_null_weight}{TRUE/FALSE. If TRUE, allow for a probability of no effect in susie}

\item{coverage}{A number between 0 and 1 specifying the \dQuote{coverage} of the estimated confidence sets}

\item{min_abs_corr}{Minimum absolute correlation allowed in a
credible set. The default, 0.5, corresponds to a squared
correlation of 0.25, which is a commonly used threshold for
genotype data in genetic studies.}

\item{force_compute_cor}{TRUE/FALSE. If TRUE, force computing correlation (R) matrices}

\item{save_cor}{TRUE/FALSE. If TRUE, save correlation (R) matrices to \code{cor_dir}}

\item{cor_dir}{a string, the directory to store correlation (R) matrices}

\item{verbose}{TRUE/FALSE. If TRUE, print detail messages}

\item{...}{Additional arguments of \code{susie_rss}.}

\item{annotate_susie_result}{TRUE/FALSE. If TRUE, add gene and SNP information and cs_index to
the data frame of finemapping results.}
}
\value{
a data frame of finemapping results.
}
\description{
Runs cTWAS finemapping for a single region
}
