% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_finemapping.R
\name{finemap_region}
\alias{finemap_region}
\title{run cTWAS finemapping for a single region}
\usage{
finemap_region(
  z_snp,
  z_gene,
  sid,
  gid,
  region_tag,
  region_info,
  weights,
  L = 5,
  group_prior = NULL,
  group_prior_var = NULL,
  use_null_weight = TRUE,
  coverage = 0.95,
  min_abs_corr = 0.5,
  max_iter = 100,
  force_compute_cor = FALSE,
  save_cor = FALSE,
  cor_dir = getwd(),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{z_snp}{A data frame with four columns: "id", "A1", "A2", "z".
giving the z scores for snps. "A1" is effect allele. "A2" is the other allele.}

\item{z_gene}{A data frame with two columns: "id", "z". giving the z scores for genes.
Optionally, a "type" column can also be supplied; this is for using multiple sets of weights}

\item{sid}{SNP IDs in the region}

\item{gid}{gene IDs in the region}

\item{region_tag}{a character string of region tags to be finemapped}

\item{region_info}{a data frame of region definition and associated LD file names}

\item{weights}{a list of weights for each gene}

\item{L}{the number of effects for susie during the fine mapping steps}

\item{group_prior}{a vector of two prior inclusion probabilities for SNPs and genes.}

\item{group_prior_var}{a vector of two prior variances for SNPs and gene effects.}

\item{use_null_weight}{TRUE/FALSE. If TRUE, allow for a probability of no effect in susie}

\item{coverage}{A number between 0 and 1 specifying the \dQuote{coverage} of the estimated confidence sets}

\item{min_abs_corr}{Minimum absolute correlation allowed in a
credible set. The default, 0.5, corresponds to a squared
correlation of 0.25, which is a commonly used threshold for
genotype data in genetic studies.}

\item{max_iter}{Maximum number of IBSS iterations to perform.}

\item{force_compute_cor}{TRUE/FALSE. If TRUE, force computing correlation (R) matrices}

\item{save_cor}{TRUE/FALSE. If TRUE, save correlation (R) matrices}

\item{cor_dir}{a string, the directory to store correlation (R) matrices}

\item{verbose}{TRUE/FALSE. If TRUE, print detail messages}
}
\value{
finemapping results.
}
\description{
run cTWAS finemapping for a single region
}
