% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_finemapping.R
\name{finemap_regions}
\alias{finemap_regions}
\title{run cTWAS finemapping for multiple regions}
\usage{
finemap_regions(
  region_data,
  region_info,
  weights,
  snp_info,
  L = 5,
  group_prior = NULL,
  group_prior_var = NULL,
  groups = NULL,
  use_null_weight = TRUE,
  coverage = 0.95,
  min_abs_corr = 0.5,
  ncore = 1,
  force_compute_cor = FALSE,
  save_cor = FALSE,
  cor_dir = NULL,
  annotate_susie_result = TRUE,
  verbose = FALSE,
  logfile = NULL,
  ...
)
}
\arguments{
\item{region_data}{region_data to be finemapped}

\item{region_info}{a data frame of region definition and associated LD file names}

\item{weights}{a list of weights for each gene}

\item{snp_info}{a data frame, SNP info for LD reference,
with columns "chrom", "id", "pos", and "region_id".}

\item{L}{the number of effects for susie during the fine mapping steps}

\item{group_prior}{a vector of two prior inclusion probabilities for SNPs and genes.}

\item{group_prior_var}{a vector of two prior variances for SNPs and gene effects.}

\item{groups}{a vector of group names for group_prior and group_prior_var.}

\item{use_null_weight}{TRUE/FALSE. If TRUE, allow for a probability of no effect in susie}

\item{coverage}{A number between 0 and 1 specifying the \dQuote{coverage} of the estimated confidence sets}

\item{min_abs_corr}{Minimum absolute correlation allowed in a
credible set. The default, 0.5, corresponds to a squared
correlation of 0.25, which is a commonly used threshold for
genotype data in genetic studies.}

\item{ncore}{The number of cores used to parallelize computation over regions}

\item{force_compute_cor}{TRUE/FALSE. If TRUE, force computing correlation (R) matrices}

\item{save_cor}{TRUE/FALSE. If TRUE, save correlation (R) matrices to \code{cor_dir}}

\item{cor_dir}{a string, the directory to store correlation (R) matrices}

\item{verbose}{TRUE/FALSE. If TRUE, print detail messages}

\item{logfile}{the log file, if NULL will print log info on screen}

\item{...}{Additional arguments of \code{susie_rss}.}
}
\value{
finemapping results.
}
\description{
run cTWAS finemapping for multiple regions
}
