% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_finemapping.R
\name{finemap_regions_noLD}
\alias{finemap_regions_noLD}
\title{Runs cTWAS fine-mapping for regions without LD (L = 1)}
\usage{
finemap_regions_noLD(
  region_data,
  group_prior = NULL,
  group_prior_var = NULL,
  min_var = 2,
  min_gene = 1,
  null_method = c("ctwas", "susie", "none"),
  null_weight = NULL,
  coverage = 0.95,
  include_cs = TRUE,
  include_prior = FALSE,
  include_susie_alpha = TRUE,
  snps_only = FALSE,
  ncore = 1,
  verbose = FALSE,
  logfile = NULL,
  ...
)
}
\arguments{
\item{region_data}{region_data to be finemapped}

\item{group_prior}{a vector of two prior inclusion probabilities for SNPs and genes.
If NULL, it will use uniform prior inclusion probabilities.}

\item{group_prior_var}{a vector of two prior variances for SNPs and gene effects.
If NULL, it will set prior variance = 50 as the default in \code{susie_rss}.}

\item{null_method}{Method to compute null model, options: "ctwas", "susie" or "none".}

\item{null_weight}{Prior probability of no effect (a number between
0 and 1, and cannot be exactly 1). Only used when \code{null_method = "susie"}.}

\item{coverage}{A number between 0 and 1 specifying the \dQuote{coverage} of the estimated confidence sets}

\item{include_cs}{If TRUE, include credible sets (CS) to finemapping results.}

\item{include_prior}{If TRUE, include priors in finemapping results.}

\item{include_susie_alpha}{If TRUE, include susie alpha matrix in the result.}

\item{snps_only}{If TRUE, use only SNPs in the region data.}

\item{ncore}{The number of cores used to parallelize computation over regions}

\item{verbose}{If TRUE, print detail messages}

\item{logfile}{the log file, if NULL will print log info on screen}

\item{...}{Additional arguments of \code{susie_rss}.}
}
\value{
a list with cTWAS finemapping results.
}
\description{
Runs cTWAS fine-mapping for regions without LD (L = 1)
}
