% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_finemapping.R
\name{finemap_regions_noLD}
\alias{finemap_regions_noLD}
\title{Runs cTWAS fine-mapping for regions without LD (L = 1)}
\usage{
finemap_regions_noLD(
  region_data,
  group_prior = NULL,
  group_prior_var = NULL,
  use_null_weight = TRUE,
  ncore = 1,
  verbose = FALSE,
  logfile = NULL,
  ...
)
}
\arguments{
\item{region_data}{region_data to be finemapped}

\item{group_prior}{a vector of two prior inclusion probabilities for SNPs and genes.}

\item{group_prior_var}{a vector of two prior variances for SNPs and gene effects.}

\item{use_null_weight}{If TRUE, allow for a probability of no effect in susie}

\item{ncore}{The number of cores used to parallelize computation over regions}

\item{verbose}{If TRUE, print detail messages}

\item{logfile}{the log file, if NULL will print log info on screen}

\item{...}{Additional arguments of \code{susie_rss}.}
}
\value{
a data frame of cTWAS finemapping results.
}
\description{
Runs cTWAS fine-mapping for regions without LD (L = 1)
}
