% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_finemapping.R
\name{finemap_regions_ser_rss}
\alias{finemap_regions_ser_rss}
\title{Finemap regions with cTWAS SER model}
\usage{
finemap_regions_ser_rss(
  region_data,
  group_prior = NULL,
  group_prior_var = NULL,
  min_var = 2,
  min_gene = 1,
  null_method = c("ctwas", "susie", "none"),
  ncore = 1
)
}
\arguments{
\item{region_data}{a list object with the susie input data for each region.}

\item{group_prior}{a vector of prior inclusion probabilities for SNPs and genes.}

\item{group_prior_var}{a vector of prior variances for SNPs and gene effects.}

\item{min_var}{minimum number of variables (SNPs and genes) in a region.}

\item{min_gene}{minimum number of genes in a region.}

\item{null_method}{Method to compute null model, options: "ctwas", "susie" or "none".}

\item{ncore}{The number of cores used to parallelize over regions.}
}
\value{
a data frame of SER result for finemapped regions
}
\description{
Finemap regions with cTWAS SER model
}
