% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_compute_gene_z.R
\name{get_boundary_genes}
\alias{get_boundary_genes}
\title{Get cross-boundary genes.}
\usage{
get_boundary_genes(
  region_info,
  weights,
  gene_ids = NULL,
  mapping_table = NULL,
  map_by = "molecular_id",
  ncore = 1
)
}
\arguments{
\item{region_info}{a data frame of region definitions.}

\item{weights}{a list of preprocessed weights.}

\item{gene_ids}{a vector of gene IDs (z_gene$id). If available, limits to these genes.}

\item{mapping_table}{a data frame of mapping between molecular traits and genes,
with required columns: "molecular_id", "gene_name".}

\item{map_by}{column name to be mapped by (default: "molecular_id").}

\item{ncore}{The number of cores used to parallelize over genes.}
}
\value{
a data frame of cross-boundary genes.
}
\description{
Get cross-boundary genes.
}
