% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_merge_regions.R
\name{get_merged_region_info}
\alias{get_merged_region_info}
\title{Get merged region info}
\usage{
get_merged_region_info(
  boundary_genes,
  region_info,
  use_LD = TRUE,
  LD_info,
  snp_info
)
}
\description{
Get merged region info
}
