% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_diagnose_LD_mismatch.R
\name{get_problematic_genes}
\alias{get_problematic_genes}
\title{Gets problematic genes from problematic SNPs}
\usage{
get_problematic_genes(problematic_snps, weights, z_gene, z_thresh = 3)
}
\arguments{
\item{problematic_snps}{a character vector of problematic SNP rsIDs.}

\item{weights}{a list of weights}

\item{z_gene}{A data frame with columns: "id", "z", giving the z-scores for genes.}

\item{z_thresh}{cutoff of abs(z-scores) to select genes with large effect sizes}
}
\value{
a vector of problematic genes
}
\description{
Gets problematic genes from problematic SNPs
}
