% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_region_cor.R
\name{get_region_cor}
\alias{get_region_cor}
\title{Gets correlation matrices for a single region.}
\usage{
get_region_cor(
  region_id,
  region_data,
  LD_map,
  snp_map = NULL,
  weights,
  force_compute_cor = FALSE,
  save_cor = FALSE,
  cor_dir = NULL,
  LD_format = c("rds", "rdata", "mtx", "csv", "txt", "custom"),
  LD_loader_fun,
  verbose = FALSE
)
}
\arguments{
\item{region_id}{a character string of region id to be finemapped}

\item{region_data}{a list object with data for the regions}

\item{LD_map}{a data frame with filenames of LD matrices for each of the regions.}

\item{snp_map}{list of data frames with SNP-to-region map for the reference.}

\item{weights}{a list of preprocessed weights}

\item{force_compute_cor}{TRUE/FALSE. If TRUE, force computing correlation (R) matrices}

\item{save_cor}{TRUE/FALSE. If TRUE, save correlation (R) matrices to \code{cor_dir}}

\item{cor_dir}{a string, the directory to store correlation (R) matrices}

\item{LD_format}{file format for LD matrix. If "custom", use a user defined
\code{LD_loader_fun()} function to load LD matrix.}

\item{LD_loader_fun}{a user defined function to load LD matrix when \code{LD_format = "custom"}.}

\item{verbose}{TRUE/FALSE. If TRUE, print detail messages}
}
\value{
a list of correlation matrices (R_snp, R_snp_gene and R_gene)
}
\description{
Loads precomputed correlation matrices if available in \code{cor_dir},
otherwise, computes correlation matrices if no precomputed correlation matrices,
or force_compute_cor = TRUE.
}
