% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_region_cor.R
\name{get_region_cor}
\alias{get_region_cor}
\title{Gets correlation matrices for a single region.}
\usage{
get_region_cor(
  region_id,
  sids,
  gids,
  LD_map,
  weights,
  force_compute_cor = FALSE,
  save_cor = FALSE,
  cor_dir = NULL,
  LD_format = c("rds", "rdata", "mtx", "csv", "txt", "custom"),
  LD_loader_fun = NULL,
  snpinfo_loader_fun = NULL,
  verbose = FALSE
)
}
\arguments{
\item{region_id}{a character string of region id.}

\item{sids}{SNP IDs in the region}

\item{gids}{gene IDs in the region}

\item{LD_map}{a data frame with filenames of LD matrices and SNP information for the regions.}

\item{weights}{a list of preprocessed weights}

\item{force_compute_cor}{If TRUE, force computing correlation (R) matrices}

\item{save_cor}{If TRUE, save correlation (R) matrices to \code{cor_dir}}

\item{cor_dir}{a string, the directory to store correlation (R) matrices}

\item{LD_format}{file format for LD matrix. If "custom", use a user defined
\code{LD_loader_fun()} function to load LD matrix.}

\item{LD_loader_fun}{a user defined function to load LD matrix when \code{LD_format = "custom"}.}

\item{snpinfo_loader_fun}{a user defined function to load SNP information file,
if SNP information files are not in standard cTWAS reference format.}

\item{verbose}{If TRUE, print detail messages}
}
\value{
correlation matrices (R_snp, R_snp_gene and R_gene)
}
\description{
Loads precomputed correlation matrices if available in \code{cor_dir},
otherwise, computes correlation matrices if no precomputed correlation matrices,
or force_compute_cor = TRUE.
}
