% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_regionlist.R
\name{get_regionlist}
\alias{get_regionlist}
\title{Get regionlist for each region in region_info}
\usage{
get_regionlist(
  region_info,
  z_snp,
  z_gene,
  weights,
  thin = 1,
  maxSNP = Inf,
  trim_by = c("random", "z"),
  minvar = 1,
  mingene = 0,
  adjust_boundary_genes = TRUE,
  ncore = 1,
  seed = 99
)
}
\arguments{
\item{region_info}{a data frame of region definition and associated file names}

\item{z_snp}{A data frame with columns: "id", "z", giving the z-scores for SNPs.}

\item{z_gene}{A data frame with columns: "id", "z", giving the z-scores for genes.}

\item{weights}{a list of weights}

\item{thin}{The proportion of SNPs to be used for the parameter estimation and
initial screening region steps.
Smaller \code{thin} parameters reduce runtime at the expense of accuracy.
The fine mapping step is rerun using full SNPs for regions with strong gene signals.}

\item{maxSNP}{Inf or integer. Maximum number of SNPs in a region. Default is
Inf, no limit. This can be useful if there are many SNPs in a region and you don't
have enough memory to run the program. This applies to the last rerun step
(using full SNPs and rerun susie for regions with strong gene signals) only.}

\item{trim_by}{remove SNPs if the total number of SNPs exceeds limit, options: "random",
or "z" (trim SNPs with lower |z|) See parameter `maxSNP` for more information.}

\item{minvar}{minimum number of SNPs (and genes) in a region}

\item{mingene}{minimum number of genes in a region}

\item{adjust_boundary_genes}{identify cross-boundary genes, adjust regionlist and update weighs}

\item{ncore}{The number of cores used to parallelize susie over regions}

\item{seed}{seed for random sampling}

\item{gene_info}{a data frame of gene information obtained from \code{get_gene_info}}
}
\value{
a list with regionlist, updated weights, and cross-bounary genes
}
\description{
Get regionlist for each region in region_info
}
