% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_harmonize_data.R
\name{harmonize_weights}
\alias{harmonize_weights}
\title{Harmonize weight to match LD reference genotypes.
Flip signs when reverse complement matches, remove ambiguous variants from the prediction models}
\usage{
harmonize_weights(wgt.matrix, wgt.snpinfo, snp_info, drop_strand_ambig = TRUE)
}
\arguments{
\item{wgt.matrix}{a matrix of the weights}

\item{wgt.snpinfo}{a data frame of the weight variants
with columns "chrom", "id", "cm", "pos", "alt", "ref". "alt" is the effect allele.}

\item{snp_info}{a data frame of SNP info for the reference,
with columns "chrom", "id", "pos", "alt", "ref".}

\item{drop_strand_ambig}{If TRUE, remove strand ambiguous variants (A/T, G/C).}
}
\value{
wgt.matrix and wgt.snpinfo with alleles flipped to match
}
\description{
Harmonize weight to match LD reference genotypes.
Flip signs when reverse complement matches, remove ambiguous variants from the prediction models
}
\keyword{internal}
