% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_harmonize_data.R
\name{harmonize_wgt_ld}
\alias{harmonize_wgt_ld}
\title{Harmonize weight to match ld reference genotypes.
Flip signs when reverse complement matches, remove ambiguous variants from the prediction models}
\usage{
harmonize_wgt_ld(wgt.matrix, snps, LD_snpinfo, drop_strand_ambig = TRUE)
}
\arguments{
\item{wgt.matrix}{from FUSION weight .Rdat file}

\item{snps}{from FUSION weight .Rdat file
with columns "chrom", "id", "pos", "alt", "ref". The effect allele
for FUSION is alt.}

\item{LD_snpinfo}{a data frame, snp info for LD reference,
with columns "chrom", "id", "pos", "alt", "ref".}

\item{drop_strand_ambig}{TRUE/FALSE, if TRUE remove strand ambiguous variants (A/T, G/C).}
}
\value{
wgt.matrix and snps with alleles flipped to match
}
\description{
Harmonize weight to match ld reference genotypes.
Flip signs when reverse complement matches, remove ambiguous variants from the prediction models
}
