% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_harmonize_data.R
\name{harmonize_z}
\alias{harmonize_z}
\title{Harmonizes z scores from GWAS to match LD reference genotypes.
Flips signs when reverse complement matches, and removes strand ambiguous SNPs}
\usage{
harmonize_z(z_snp, snp_info, drop_strand_ambig = TRUE)
}
\arguments{
\item{z_snp}{a data frame, with columns "id", "A1", "A2" and "z".
Z scores for every SNP. "A1" is the effect allele.}

\item{snp_info}{a data frame of SNP info for the reference,
with columns "chrom", "id", "pos", "alt", "ref".}

\item{drop_strand_ambig}{If TRUE, remove strand ambiguous variants (A/T, G/C).}
}
\value{
a data frame, z_snp with the "z" columns flipped to match LD reference.
}
\description{
Harmonizes z scores from GWAS to match LD reference genotypes.
Flips signs when reverse complement matches, and removes strand ambiguous SNPs
}
\keyword{internal}
