% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_impute_expr.R
\name{impute_expr}
\alias{impute_expr}
\title{Impute expression}
\usage{
impute_expr(
  pgenfs,
  weight,
  method = "lasso",
  outputdir = getwd(),
  outname = NULL,
  logfile = NULL,
  compress = T,
  harmonize_wgt = T,
  ncore = 1,
  scale_by_ld_variance = F,
  strand_ambig_action_wgt = c("drop", "none", "recover")
)
}
\arguments{
\item{pgenfs}{A character vector of .pgen or .bed files. One file for one
chromosome, in the order of 1 to 22. Therefore, the length of this vector
needs to be 22. If .pgen files are given, then .pvar and .psam are assumed
to present in the same directory. If .bed files are given, then .bim and
.fam files are assumed to present in the same directory.}

\item{weight}{a string, pointing to a directory with the fusion/twas format of weights, 
or a vector of one or more .db files in PredictDB format.}

\item{method}{a string, blup/bslmm/lasso/top1/enet/best. This option is only used for fusion weights. 
"best" means the method giving the best cross #' validation R^2. Note that top1 uses only the weight 
with largest effect.}

\item{outputdir}{a string, the directory to store output}

\item{outname}{a string, the output name}

\item{logfile}{the log file, if NULL will print log info on screen}

\item{compress}{TRUE/FALSE. If TRUE, the imputed expression files are compressed}

\item{harmonize_wgt}{TRUE/FALSE. If TRUE, GWAS and eQTL genotype alleles are harmonized}

\item{ncore}{The number of cores used to parallelize imputation over weights}

\item{scale_by_ld_variance}{TRUE/FALSE. If TRUE, PredictDB weights are scaled by genotype variance, which is the default 
behavior for PredictDB}

\item{strand_ambig_action_wgt}{the action to take to harmonize strand ambiguous variants (A/T, G/C) between 
the weights and LD reference. "drop" removes the ambiguous variant from the prediction models. "none" treats the variant 
as unambiguous, flipping the weights to match the LD reference and then taking no additional action. "recover" uses a procedure
to recover strand ambiguous variants. This procedure compares correlations between variants in the 
LD reference and prediction models, and it can only be used with PredictDB format prediction models, which include this
information.}
}
\description{
Impute expression
}
