% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_impute_expr.R
\name{impute_expr}
\alias{impute_expr}
\title{Impute expression}
\usage{
impute_expr(pgenf, weight, method = "lasso", outputdir = getwd(),
  outname = NULL, logfile = NULL, compress = T, harmonize = T)
}
\arguments{
\item{weight}{a string, pointing to the fusion/twas format of weights.
Note the effect size are obtained with standardized genotype and phenotype.}

\item{method}{a string,  blup/bslmm/lasso/top1/enet/best
"best" means the method giving the best cross validation R^2}

\item{harmonize}{T/F, if need to harmonize SNP data, if T, will harmonize GWAS and eQTL genotype alleles.}
}
\description{
Impute expression
}
