% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_impute_expr.R
\name{impute_expr}
\alias{impute_expr}
\title{Impute expression}
\usage{
impute_expr(pgenf, weight, method = "lasso", outputdir = getwd(),
  outname = NULL, logfile = NULL, compress = T, checksnps = F)
}
\arguments{
\item{weight}{a string, pointing to the fusion/twas format of weights.
Note the effect size are obtained with standardized genotype and phenotype.}

\item{method}{a string,  blup/bslmm/lasso/top1/enet/best
"best" means the method giving the best cross validation R^2}

\item{checksnps}{T/F, if need to check SNP consistency between weights
file and genotype. fusion format of weights gives snp information from plink .bim file. checking include chr, pos, A1, A2.}
}
\description{
Impute expression
}
