% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_process_regions.R
\name{index_regions}
\alias{index_regions}
\title{Get gene and SNP index for each region}
\usage{
index_regions(pvarfs, exprvarfs, regionfile, select = NULL, thin = 1)
}
\arguments{
\item{regionfile}{Has three columns: chr, start, end}

\item{select}{variant ids to include. If NULL, all variants will be selected}

\item{thin}{A scalar in (0,1]. The proportion of SNPs
left after down sampling. Only applied on SNPs after selecting variants.}
}
\value{
A list. Items correspond to each pvarf/exprvarf. Each Item is
 also a list, the items in this list are for each region.
}
\description{
For each region, get the index for snp and gene (index
is location/column number in .pgen file or .expr file) located within
this region.
}
