% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_weights.R
\name{load_fusion_weights}
\alias{load_fusion_weights}
\title{Loads weights in FUSION format}
\usage{
load_fusion_weights(
  weight_dir,
  fusion_method = c("lasso", "enet", "top1", "blup", "bslmm", "best.cv"),
  fusion_genome_version = NA,
  make_extra_table = FALSE,
  ncore = 1
)
}
\arguments{
\item{weight_dir}{the directory containing FUSION weights ('.wgt.RDat' files).}

\item{fusion_method}{a string, specifying the method to choose in FUSION models.
"best.cv" option will use the best model (smallest p-value) under cross-validation.}

\item{fusion_genome_version}{a string, specifying the genome version of FUSION models}

\item{make_extra_table}{If TRUE, make an extra table in predictDB format}

\item{ncore}{integer, number of cores for parallel computing.}
}
\description{
Loads weights in FUSION format
}
