% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_weights.R
\name{load_predictdb_weights}
\alias{load_predictdb_weights}
\title{Loads weights in PredictDB format}
\usage{
load_predictdb_weights(
  weight_file,
  filter_protein_coding_genes = TRUE,
  load_predictdb_LD = TRUE
)
}
\arguments{
\item{weight_file}{a string, pointing path to weights in PredictDB format.}

\item{filter_protein_coding_genes}{TRUE/FALSE. If TRUE, keep protein coding
genes only.}

\item{load_predictdb_LD}{TRUE/FALSE. If TRUE, load pre-computed LD
among weight variants.}
}
\description{
Loads weights in PredictDB format
}
