% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_utils.R
\name{load_weights}
\alias{load_weights}
\title{Load weight}
\usage{
load_weights(
  weight_file,
  weight_format = c("PredictDB", "Fusion"),
  filter_protein_coding_genes = FALSE,
  load_predictdb_LD = FALSE
)
}
\description{
Load weight
}
