% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_weights.R
\name{load_weights}
\alias{load_weights}
\title{Loads weights in PredictDB or FUSION format}
\usage{
load_weights(
  weight_file,
  weight_format = c("PredictDB", "FUSION"),
  filter_protein_coding_genes = TRUE,
  load_predictdb_LD = TRUE,
  fusion_method = c("lasso", "enet", "top1", "blup", "bslmm", "best.cv"),
  fusion_genome_version = NA,
  ncore = 1
)
}
\arguments{
\item{weight_file}{filename of the '.db' file for PredictDB weights;
or the directory containing '.wgt.RDat' files for FUSION weights.}

\item{weight_format}{a string or a vector, specifying format of each weight file, e.g. PredictDB, FUSION.}

\item{filter_protein_coding_genes}{If TRUE, keep protein coding genes only. This option is only for PredictDB weights}

\item{load_predictdb_LD}{If TRUE, load pre-computed LD among weight SNPs. This option is only for PredictDB weights}

\item{fusion_method}{a string, specifying the method to choose in FUSION models.
"best.cv" option will use the best model (smallest p-value) under cross-validation.}

\item{fusion_genome_version}{a string, specifying the genome version of FUSION models}

\item{ncore}{integer, number of cores for parallel computing.}
}
\description{
Loads weights in PredictDB or FUSION format
}
