% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_plots.R
\name{make_locusplot}
\alias{make_locusplot}
\title{Plot the cTWAS result for a single locus}
\usage{
make_locusplot(
  finemap_res,
  screened_region_data,
  region_id,
  cor_matrix_path = getwd(),
  locus_range = NULL,
  focus_gene = NULL,
  gene_biotype = "protein_coding",
  genome_build = "hg38"
)
}
\arguments{
\item{finemap_res}{a data frame of cTWAS finemapping result}

\item{screened_region_data}{a list of screened region_data for regions with strong signals}

\item{region_id}{region ID to be plotted}

\item{cor_matrix_path}{path of correlation matrices}

\item{locus_range}{a vector of start and end positions to define the locus region to be plotted
If NULL, the entire region will be plotted.}

\item{focus_gene}{the gene to focus the plot. If NULL, use the lead TWAS feature}

\item{gene_biotype}{specified biotypes to be displayed in gene tracks or NULL to display all possible ones}

\item{genome_build}{genome build version: hg38 or hg19}
}
\description{
Plot the cTWAS result for a single locus
}
