% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_plots.R
\name{make_locusplot}
\alias{make_locusplot}
\title{Plots the cTWAS result for a single locus}
\usage{
make_locusplot(
  finemap_res,
  region_id,
  weights,
  ens_db,
  R_snp_gene = NULL,
  R_gene = NULL,
  locus_range = NULL,
  focus_gene = NULL,
  filter_gene_biotype = "protein_coding",
  highlight_pval = NULL,
  highlight_pip = 0.8,
  highlight_pos = NULL,
  point.sizes = c(1, 3.5),
  point.alpha = c(0.4, 0.6),
  point.shapes = c(21:25),
  genelabel.size = 2.5,
  legend.text.size = 10,
  legend.position = c("none", "top", "left", "right", "bottom"),
  panel.heights = c(4, 4, 0.3, 4)
)
}
\arguments{
\item{finemap_res}{a data frame of cTWAS finemapping result}

\item{region_id}{region ID to be plotted}

\item{weights}{a list of weights}

\item{ens_db}{Ensembl database}

\item{R_snp_gene}{SNP-gene correlation matrix of the region.
If both R_snp_gene and R_gene are available,
color data points with correlations with the focus gene.}

\item{R_gene}{gene-gene correlation matrix of the region}

\item{locus_range}{a vector of start and end positions to define the region boundary to be plotted.
If NULL, the entire region will be plotted (with 100 bp flanks on both sides).}

\item{focus_gene}{the gene name to focus the plot.
If NULL, choose the gene with the highest PIP.}

\item{filter_gene_biotype}{biotype to be displayed in gene tracks.
By default, limits to protein coding genes. If NULL, display all possible ones.}

\item{highlight_pval}{pvalue to highlight with a horizontal line}

\item{highlight_pip}{PIP to highlight with a horizontal line}

\item{highlight_pos}{genomic positions to highlight with vertical lines}

\item{point.sizes}{size values for SNP and non-SNP data points in the scatter plots}

\item{point.alpha}{alpha values for SNP and non-SNP data points in the scatter plots}

\item{point.shapes}{shapes values for data points of different types in the scatter plots}

\item{genelabel.size}{Font size for gene text}

\item{legend.text.size}{Font size for legend text}

\item{legend.position}{position to put legends. If "none", no legends will be shown.}

\item{panel.heights}{Relative heights of the panels.}
}
\description{
Plots the cTWAS result for a single locus
}
