% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_plots.R
\name{make_locusplot}
\alias{make_locusplot}
\title{Plots the cTWAS result for a single locus}
\usage{
make_locusplot(
  finemap_res,
  region_id,
  ens_db,
  weights = NULL,
  R_snp_gene = NULL,
  R_gene = NULL,
  locus_range = NULL,
  focal_id = NULL,
  focal_gene = NULL,
  filter_protein_coding_genes = TRUE,
  filter_cs = TRUE,
  color_PIP_by_cs = TRUE,
  label_QTLs = TRUE,
  highlight_pval = NULL,
  highlight_pip = 0.8,
  highlight_pos = NULL,
  point.sizes = c(1, 3.5),
  point.alpha = c(0.4, 0.6),
  point.shapes = c(16, 15, 18, 17, 10, 12, 14, 11),
  label.text.size = 2.5,
  max.overlaps = 10,
  legend.text.size = 10,
  legend.position = "top",
  panel.heights = c(4, 4, 1, 4),
  verbose = FALSE
)
}
\arguments{
\item{finemap_res}{a data frame of cTWAS finemapping result}

\item{region_id}{region ID to be plotted}

\item{ens_db}{Ensembl database}

\item{weights}{a list of proprocessed weights, used to plot QTL track.
If NULL, will not plot the QTL track.}

\item{R_snp_gene}{SNP-gene correlation matrix of the region.
If both R_snp_gene and R_gene are available,
color data points with correlations with the focal gene.}

\item{R_gene}{gene-gene correlation matrix of the region}

\item{locus_range}{a vector of start and end positions to define the region boundary to be plotted.
If NULL, the entire region will be plotted (with 100 bp flanks on both sides).}

\item{focal_gene}{the focal gene name. By default, use the gene with the highest PIP.}

\item{filter_protein_coding_genes}{If TRUE, limits to protein coding genes only.}

\item{filter_cs}{If TRUE, limits to credible sets.}

\item{color_PIP_by_cs}{If TRUE, color PIP track by credible sets.}

\item{label_QTLs}{If TRUE, label SNP IDs in the QTL panel.}

\item{highlight_pval}{pvalue to highlight with a horizontal line}

\item{highlight_pip}{PIP to highlight with a horizontal line}

\item{highlight_pos}{genomic positions to highlight with vertical lines}

\item{point.sizes}{size values for SNP and non-SNP data points in the scatter plots}

\item{point.alpha}{alpha values for SNP and non-SNP data points in the scatter plots}

\item{point.shapes}{shapes values for data points of different types in the scatter plots}

\item{label.text.size}{Font size for gene and SNP label text}

\item{max.overlaps}{Setting for geom_text_repel() function to label texts.
Exclude text labels when they overlap too many other things.}

\item{legend.text.size}{Font size for legend text}

\item{legend.position}{position to put legends. If "none", no legends will be shown.}

\item{panel.heights}{Relative heights of the panels.}

\item{verbose}{If TRUE, print detail messages.}
}
\description{
Plots the cTWAS result for a single locus
}
