% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_plots.R
\name{make_locusplot}
\alias{make_locusplot}
\title{Plot the cTWAS result for a single locus}
\usage{
make_locusplot(
  finemap_res,
  region_id,
  weights,
  ens_db,
  save_cor = FALSE,
  cor_dir = NULL,
  region_data = NULL,
  region_info = NULL,
  snp_info = NULL,
  locus_range = NULL,
  focus_gene = NULL,
  gene_biotype = "protein_coding",
  p_thresh = NULL,
  pip_thresh = NULL,
  point.sizes = c(1, 3.5),
  point.alpha = c(0.4, 0.6),
  point.shapes = c(21:25),
  genelabel.size = 2.5,
  legend.text.size = 10,
  legend.position = "none",
  panel.heights = c(4, 4, 0.3, 4),
  highlight_pos = NULL
)
}
\arguments{
\item{finemap_res}{a data frame of cTWAS finemapping result}

\item{region_id}{region ID to be plotted}

\item{weights}{a list of weights}

\item{ens_db}{Ensembl database}

\item{save_cor}{TRUE/FALSE. If TRUE, save correlation (R) matrices to \code{cor_dir}}

\item{cor_dir}{path of correlation matrices}

\item{region_data}{a list of region_data used for finemapping}

\item{region_info}{a data frame of region definition and associated LD file names}

\item{snp_info}{a data frame, SNP info for LD reference,
with columns "chrom", "id", "pos", and "region_id".}

\item{locus_range}{a vector of start and end positions to define the locus region to be plotted

If NULL, the entire region will be plotted (with 100 bp flanks on both sides).}

\item{focus_gene}{the gene name to focus the plot. If NULL, choose the gene with the highest PIP}

\item{gene_biotype}{specified biotypes to be displayed in gene tracks or NULL to display all possible ones.
By default, only show protein coding genes.}

\item{point.sizes}{size values for SNP and non-SNP data points in the scatter plots}

\item{point.alpha}{alpha values for SNP and non-SNP data points in the scatter plots}

\item{point.shapes}{shapes values for data points of different types in the scatter plots}

\item{genelabel.size}{Font size for gene text}

\item{legend.text.size}{Font size for legend text}

\item{legend.position}{position to put legends. If "none", no legends will be shown.}

\item{panel.heights}{Relative heights of the panels.}

\item{highlight_pos}{If not NULL, highlighst the position in all panels for the genomic position.}
}
\description{
Plot the cTWAS result for a single locus
}
