% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_weights.R
\name{make_predictdb_weights_from_top_QTLs}
\alias{make_predictdb_weights_from_top_QTLs}
\title{Makes PredictDB weights from top QTLs}
\usage{
make_predictdb_weights_from_top_QTLs(
  QTL_data,
  gene_info = NULL,
  outputdir = getwd(),
  weight_name
)
}
\arguments{
\item{QTL_data}{a data frame with required columns: "gene", "varID", "weight".
for the genes, and top QTL and weight for each gene,}

\item{gene_info}{a data frame (optional) with information of the genes
("gene","genename","gene_type", etc.)}

\item{outputdir}{output directory}

\item{weight_name}{name of the output weight files}
}
\description{
Make PredictDB weights from only top QTLs.
Each gene (or molecular trait) has only one top QTL, defined by users,
based on min p-value, max abs(weight), etc.
}
