% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_map_snp_info_regions.R
\name{map_snp_info_regions}
\alias{map_snp_info_regions}
\title{Read SNP info from all regions as a data frame and map SNPs to regions.
if snp_info_file is available, read SNP info from the snp_info_file;
otherwise, read from the "SNP_info" column of region_info.}
\usage{
map_snp_info_regions(region_info, snp_info = NULL, ncore = 1)
}
\arguments{
\item{region_info}{a data frame of region definition and associated file names}

\item{snp_info}{a data frame of all variant info in the LD reference.}

\item{ncore}{The number of cores used to parallelize over regions}
}
\description{
Read SNP info from all regions as a data frame and map SNPs to regions.
if snp_info_file is available, read SNP info from the snp_info_file;
otherwise, read from the "SNP_info" column of region_info.
}
