% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_merge_regions.R
\name{merge_region_data}
\alias{merge_region_data}
\title{Merges region data for cross-boundary genes}
\usage{
merge_region_data(
  boundary_genes,
  region_data,
  region_info,
  snp_map,
  LD_map,
  z_snp,
  z_gene,
  use_LD = TRUE,
  expand = TRUE,
  maxSNP = Inf,
  trim_by = c("random", "z"),
  ncore = 1,
  seed = 99
)
}
\arguments{
\item{boundary_genes}{a data frame of boundary gene info}

\item{region_data}{a list of original region_data}

\item{region_info}{a data frame of region definition and associated LD file names}

\item{snp_map}{a list of SNP info data frames for LD reference.}

\item{LD_map}{a list of paths to LD matrices for each of the regions.}

\item{z_snp}{A data frame with columns: "id", "z", giving the z-scores for SNPs.}

\item{z_gene}{A data frame with columns: "id", "z", giving the z-scores for genes.}

\item{use_LD}{TRUE/FALSE. If TRUE, use LD for finemapping.}

\item{expand}{TRUE/FALSE. If TRUE, expand merged region_data with full SNPs}

\item{maxSNP}{Inf or integer. Maximum number of SNPs in a region. Default is
Inf, no limit. This can be useful if there are many SNPs in a region and you don't
have enough memory to run the program.}

\item{trim_by}{remove SNPs if the total number of SNPs exceeds limit, options: "random",
or "z" (trim SNPs with lower |z|) See parameter `maxSNP` for more information.}

\item{ncore}{The number of cores used to parallelize susie over regions}

\item{seed}{seed for random sampling}
}
\value{
a list of merged region data, merged region info, LD info, snp info,
and merged region IDs.
}
\description{
Merges region data for cross-boundary genes
}
