% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_merge_regions.R
\name{merge_region_data}
\alias{merge_region_data}
\title{Merges region data for cross-boundary genes}
\usage{
merge_region_data(
  boundary_genes,
  region_data,
  region_info,
  LD_map,
  snp_map,
  weights,
  z_snp,
  z_gene,
  expand = TRUE,
  maxSNP = Inf,
  ncore = 1,
  verbose = FALSE,
  logfile = NULL,
  ...
)
}
\arguments{
\item{boundary_genes}{a data frame of boundary gene info}

\item{region_data}{a list of original region_data}

\item{region_info}{a data frame of region definitions}

\item{LD_map}{a data frame with filenames of LD matrices and SNP information for the regions.}

\item{snp_map}{a list of data frames with SNP-to-region map for the reference.}

\item{weights}{a list of preprocessed weights.}

\item{z_snp}{A data frame with columns: "id", "z", giving the z-scores for SNPs.}

\item{z_gene}{A data frame with columns: "id", "z", giving the z-scores for genes.}

\item{expand}{If TRUE, expand merged region_data with full SNPs when "thin" < 1.}

\item{maxSNP}{Inf or integer. Maximum number of SNPs in a region. Default is
Inf, no limit. This can be useful if there are many SNPs in a region and you don't
have enough memory to run the program.}

\item{ncore}{The number of cores used to parallelize susie over regions}

\item{verbose}{If TRUE, print detail messages}

\item{logfile}{The log filename. If NULL, will print log info on screen.}

\item{...}{Additional arguments of \code{susie_rss}.}
}
\value{
a list of merged region data, merged region info, LD_map, snp_map,
and merged region IDs.
}
\description{
Merges region data for cross-boundary genes
}
