% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_preprocess_weights.R
\name{preprocess_weights}
\alias{preprocess_weights}
\title{Preprocess PredictDB/FUSION weights and harmonize with LD reference}
\usage{
preprocess_weights(
  weight_path,
  region_info,
  gwas_snp_ids,
  snp_info,
  LD_info = NULL,
  weight_format = c("PredictDB", "FUSION"),
  drop_strand_ambig = TRUE,
  scale_predictdb_weights = TRUE,
  filter_protein_coding_genes = TRUE,
  load_predictdb_LD = TRUE,
  fusion_method = c("lasso", "enet", "top1", "blup", "bslmm", "best.cv"),
  fusion_genome_version = c("b38", "b37"),
  fusion_top_n_snps,
  LD_format = c("rds", "rdata", "csv", "txt", "custom"),
  LD_loader_fun = NULL,
  ncore = 1,
  logfile = NULL
)
}
\arguments{
\item{weight_path}{path to the '.db' file for PredictDB weights;
or the directory containing '.wgt.RDat' files for FUSION weights.}

\item{region_info}{a data frame of region definition and associated file names.}

\item{gwas_snp_ids}{a vector of SNP IDs in GWAS summary statistics (z_snp$id).}

\item{snp_info}{a list of SNP info data frames for LD reference,
with columns "chrom", "id", "pos", "alt", "ref", and "region_id".}

\item{LD_info}{a list of paths to LD matrices for each of the regions. Required when \code{load_predictdb_LD = FALSE}.}

\item{weight_format}{a string, specifying format of each weight file, e.g. PredictDB, FUSION.}

\item{drop_strand_ambig}{TRUE/FALSE, if TRUE remove strand ambiguous variants (A/T, G/C).}

\item{scale_predictdb_weights}{TRUE/FALSE, if TRUE scale PredictDB weights by the variance.
This is because PredictDB weights assume that variant genotypes are not
standardized, but our implementation assumes standardized variant genotypes.}

\item{filter_protein_coding_genes}{TRUE/FALSE. If TRUE, keep protein coding genes only. This option is only for PredictDB weights}

\item{load_predictdb_LD}{TRUE/FALSE. If TRUE, load pre-computed LD among weight SNPs. This option is only for PredictDB weights}

\item{fusion_method}{a string, specifying the method to choose in FUSION models.
"best.cv" option will use the best model (smallest p-value) under cross-validation.}

\item{fusion_genome_version}{a string, specifying the genome version of FUSION models}

\item{fusion_top_n_snps}{a number, specifying the top n weight SNPs included in FUSION models.
By default, use all weight SNPs.}

\item{LD_format}{file format for LD matrix. If "custom", use a user defined
\code{LD_loader_fun()} function to load LD matrix.}

\item{LD_loader_fun}{a user defined function to load LD matrix when \code{LD_format = "custom"}.}

\item{ncore}{The number of cores used to parallelize computation.}

\item{logfile}{the log file, if NULL will print log info on screen.}
}
\value{
a list of processed weights
}
\description{
Preprocess PredictDB/FUSION weights and harmonize with LD reference
}
