% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_preprocess_weights.R
\name{preprocess_weights}
\alias{preprocess_weights}
\title{Preprocess PredictDB/FUSION weights and harmonize with LD reference}
\usage{
preprocess_weights(
  weight_file,
  region_info,
  gwas_snp_ids,
  type,
  context,
  snp_info,
  LD_info = NULL,
  weight_format = c("PredictDB", "FUSION"),
  drop_strand_ambig = TRUE,
  scale_by_ld_variance = FALSE,
  filter_protein_coding_genes = FALSE,
  load_predictdb_LD = FALSE,
  fusion_method = c("lasso", "enet", "top1", "blup"),
  fusion_genome_version = c("b38", "b37"),
  fusion_top_n_snps = NULL,
  LD_format = c("rds", "rdata", "csv", "txt", "custom"),
  LD_loader = NULL,
  ncore = 1,
  logfile = NULL
)
}
\arguments{
\item{weight_file}{a string or a vector, pointing path to one or multiple sets of weights in PredictDB or FUSION format.}

\item{region_info}{a data frame of region definition and associated file names.}

\item{gwas_snp_ids}{a vector of SNP IDs in GWAS summary statistics (z_snp$id).}

\item{type}{a string, specifying QTL type of each weight file, e.g. expression, splicing, protein.}

\item{context}{a string, specifying tissue/cell type/condition of each weight file, e.g. Liver, Lung, Brain.}

\item{snp_info}{a list of SNP info data frames for LD reference,
with columns "chrom", "id", "pos", "alt", "ref", and "region_id".}

\item{LD_info}{a list of paths to LD matrices for each of the regions. Required when \code{load_predictdb_LD = FALSE}.}

\item{weight_format}{a string, specifying format of each weight file, e.g. PredictDB, FUSION.}

\item{drop_strand_ambig}{TRUE/FALSE, if TRUE remove strand ambiguous variants (A/T, G/C).}

\item{scale_by_ld_variance}{TRUE/FALSE, if TRUE scale PredictDB weights by the variance.
This is because PredictDB weights assume that variant genotypes are not
standardized, but our implementation assumes standardized variant genotypes.}

\item{filter_protein_coding_genes}{TRUE/FALSE. If TRUE, keep protein coding genes only. This option is only for PredictDB weights}

\item{load_predictdb_LD}{TRUE/FALSE. If TRUE, load pre-computed LD among weight SNPs. This option is only for PredictDB weights}

\item{fusion_method}{a string, specifying the method to choose in FUSION models}

\item{fusion_genome_version}{a string, specifying the genome version of FUSION models}

\item{fusion_top_n_snps}{a number, specifying the top n weight SNPs included in FUSION models. If NULL, using all weight SNPs}

\item{LD_format}{file format for LD matrix. If "custom", use a user defined
\code{LD_loader()} function to load LD matrix.}

\item{LD_loader}{a user defined function to load LD matrix when \code{LD_format = "custom"}.}

\item{ncore}{The number of cores used to parallelize computation.}

\item{logfile}{the log file, if NULL will print log info on screen.}
}
\value{
a list of processed weights
}
\description{
Preprocess PredictDB/FUSION weights and harmonize with LD reference
}
