% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_preprocess_data.R
\name{preprocess_z_ld}
\alias{preprocess_z_ld}
\title{Preprocess GWAS summary statistics, harmonize GWAS z-scores and LD reference, and filter LD mismatches using susie_rss}
\usage{
preprocess_z_ld(
  z_snp,
  ld_R_dir,
  chr = 1:22,
  ld_regions = c("EUR", "ASN", "AFR"),
  ld_regions_version = c("b37", "b38"),
  ld_regions_custom = NULL,
  filestem,
  gwas_n,
  outputdir = getwd(),
  outname = NULL,
  logfile = NULL,
  strand_ambig_action = c("none", "drop", "recover"),
  drop_multiallelic = T,
  filter_ld_mismatch_susie = T,
  ncore = 1
)
}
\arguments{
\item{z_snp}{A data frame with two columns: "id", "A1", "A2", "z". giving the z scores for
snps. "A1" is effect allele. "A2" is the other allele. If `harmonize= False`, A1 and A2 are not required.}

\item{chr}{a vector containing the chromosome numbers to process.}

\item{ld_regions}{a string, population of LD reference, "EUR", "ASN", or "AFR".}

\item{ld_regions_version}{a string, version LD reference, "b37", "b38".}

\item{ld_regions_custom}{a string, path of LD regions file.}

\item{filestem}{a string, filestem of reference LD matrix.}

\item{gwas_n}{integer, GWAS sample size}

\item{outputdir}{a string, the directory to store output.}

\item{outname}{a string, the output name.}

\item{logfile}{the log file, if NULL will print log info on screen.}

\item{strand_ambig_action}{the action to take to harmonize strand ambiguous variants (A/T, G/C) between
the z scores and LD reference. "drop" removes the ambiguous variant from the z scores. "none" treats the variant
as unambiguous, flipping the z score to match the LD reference and then taking no additional action. "recover"
imputes the sign of ambiguous z scores using unambiguous z scores and the LD reference and flips the z scores
if there is a mismatch between the imputed sign and the observed sign of the z score. This option is computationally intensive.}

\item{drop_multiallelic}{TRUE/FALSE. If TRUE, multiallelic variants will be dropped from the summary statistics.}

\item{filter_ld_mismatch_susie}{TRUE/FALSE. If TRUE, detect LD mismatches using susie_rss,
flip the sign for variants detected with allele flipping, and filter problematic variants with p-value < 5e-8.}

\item{ncore}{integer, number of cores for parallel computing.}

\item{LD_R_dir}{a string, pointing to a directory containing all LD matrix files and variant information. Expects .RDS files which contain LD correlation matrices for a region/block.
For each RDS file, a file with same base name but ended with .Rvar needs to be present in the same folder. the .Rvar file has 5 required columns: "chrom", "id", "pos", "alt", "ref".
If using PredictDB format weights and \code{scale_by_ld_variance=T}, a 6th column is also required: "variance", which is the variance of the each SNP.
The order of rows needs to match the order of rows in .RDS file.}
}
\description{
Preprocess GWAS summary statistics, harmonize GWAS z-scores and LD reference, and filter LD mismatches using susie_rss
}
