% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_expr}
\alias{read_expr}
\title{Read .expr file into R}
\usage{
read_expr(exprf, variantidx = NULL)
}
\arguments{
\item{variantidx}{variant index. If NULL, all variants will be extracted.}
}
\value{
A matrix, columns are imputed expression for each gene, rows are
 for each sample.
}
\description{
Read .expr file into R
}
