% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_preprocess_weight.R
\name{read_weights}
\alias{read_weights}
\title{read (preprocessed) weights in PredictDB format}
\usage{
read_weights(
  weight_files,
  chrom,
  ld_snpinfo,
  z_snp = NULL,
  scale_by_ld_variance = TRUE,
  ncore = 1
)
}
\description{
read (preprocessed) weights in PredictDB format
}
