% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_screen_regions.R
\name{screen_regions}
\alias{screen_regions}
\title{Screens regions with strong non-SNP PIPs to be used in finemapping step}
\usage{
screen_regions(
  region_data,
  use_LD = TRUE,
  LD_info = NULL,
  snp_info = NULL,
  weights = NULL,
  group_prior = NULL,
  group_prior_var = NULL,
  group_names = NULL,
  L = 5,
  minvar = 2,
  mingene = 1,
  screen_method = c("nonSNP_PIP", "cs"),
  min_nonSNP_PIP = 0.5,
  ncore = 1,
  logfile = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{region_data}{a list object indexing regions, variants and genes.}

\item{use_LD}{TRUE/FALSE. If TRUE, use LD for finemapping. Otherwise, use "no-LD" version.}

\item{LD_info}{a list of paths to LD matrices for each of the regions. Required when \code{use_LD = TRUE}.}

\item{snp_info}{a list of SNP info data frames for LD reference. Required when \code{use_LD = TRUE}.}

\item{weights}{a list of weights for each gene}

\item{group_prior}{a vector of two prior inclusion probabilities for SNPs and genes.}

\item{group_prior_var}{a vector of two prior variances for SNPs and gene effects.}

\item{group_names}{a vector of group names for group_prior and group_prior_var.}

\item{L}{the number of effects for susie}

\item{minvar}{minimum number of variables (snps and genes) in a region}

\item{mingene}{minimum number of genes in a region}

\item{screen_method}{screening regions by: "nonSNP_PIP" or
"cs" (only available for "LD" version)}

\item{min_nonSNP_PIP}{If screening by "nonSNP_PIP",
regions with non-SNP PIP >= \code{min_nonSNP_PIP}
will be selected to run finemapping using full SNPs.}

\item{ncore}{The number of cores used to parallelize susie over regions}

\item{logfile}{the log file, if NULL will print log info on screen}

\item{verbose}{TRUE/FALSE. If TRUE, print detail messages}

\item{...}{Additional arguments of \code{susie_rss}.}
}
\value{
a list of screened_region_data, and
non-SNP PIPs or estimated L for each region
}
\description{
Screens regions with strong non-SNP PIPs to be used in finemapping step
}
