% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_screen_regions.R
\name{screen_regions}
\alias{screen_regions}
\title{Screen regions with strong non-SNP PIPs to be used in finemapping step}
\usage{
screen_regions(
  region_data,
  region_info,
  weights,
  snp_info,
  group_prior = NULL,
  group_prior_var = NULL,
  L = 5,
  minvar = 2,
  mingene = 1,
  min_nonSNP_PIP = 0.5,
  ncore = 1,
  logfile = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{region_data}{a list object indexing regions, variants and genes.}

\item{region_info}{a data frame of region definition and associated LD file names}

\item{weights}{a list of weights for each gene}

\item{snp_info}{a data frame, SNP info for LD reference,
with columns "chrom", "id", "pos", and "region_id".}

\item{group_prior}{a vector of two prior inclusion probabilities for SNPs and genes.}

\item{group_prior_var}{a vector of two prior variances for SNPs and gene effects.}

\item{L}{the number of effects for susie}

\item{minvar}{minimum number of variables (snps and genes) in a region}

\item{mingene}{minimum number of genes in a region}

\item{min_nonSNP_PIP}{Regions with non-SNP PIP >= \code{min_nonSNP_PIP}
will be selected to run finemapping using full SNPs.}

\item{ncore}{The number of cores used to parallelize susie over regions}

\item{logfile}{the log file, if NULL will print log info on screen}

\item{verbose}{TRUE/FALSE. If TRUE, print detail messages}

\item{...}{Additional arguments of \code{susie_rss}.}
}
\value{
a list of screened_region_data, and a data frame of region ids and non-SNP PIPs
}
\description{
Screen regions with strong non-SNP PIPs to be used in finemapping step
}
