% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_screen_regions.R
\name{screen_regions}
\alias{screen_regions}
\title{Screens regions with strong signals}
\usage{
screen_regions(
  region_data,
  group_prior = NULL,
  group_prior_var = NULL,
  min_var = 2,
  min_gene = 1,
  min_nonSNP_PIP = 0.5,
  min_snp_pval = 5e-08,
  min_gene_pval = min_snp_pval,
  use_twas_min_pval = FALSE,
  null_method = c("ctwas", "susie", "none"),
  ncore = 1,
  logfile = NULL,
  verbose = FALSE
)
}
\arguments{
\item{region_data}{a list object indexing regions, variants and genes.}

\item{group_prior}{a vector of prior inclusion probabilities for different groups.}

\item{group_prior_var}{a vector of prior variances for different groups.}

\item{min_var}{minimum number of variables (SNPs and genes) in a region.}

\item{min_gene}{minimum number of genes in a region.}

\item{min_nonSNP_PIP}{If screening by non-SNP PIPs,
regions with total non-SNP PIP >= \code{min_nonSNP_PIP}
will be selected to run finemapping using full SNPs.}

\item{min_snp_pval}{Select regions with minimum SNP p-values < \code{min_snp_pval}.}

\item{min_gene_pval}{Select regions with minimum gene p-values < \code{min_gene_pval}.
By default, it is set to the same value as \code{min_snp_pval}.}

\item{null_method}{Method to compute null model, options: "ctwas", "susie" or "none".}

\item{ncore}{The number of cores used to parallelize susie over regions.}

\item{logfile}{The log filename. If NULL, print log info on screen.}

\item{verbose}{If TRUE, print detail messages.}
}
\value{
a list, containing a data frame of selected region data,
and a data frame of screening summary for all regions.
}
\description{
Screens regions with strong signals
}
