% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_screen_regions.R
\name{screen_regions}
\alias{screen_regions}
\title{Screen regions}
\usage{
screen_regions(
  region_data,
  region_info,
  weights,
  group_prior = NULL,
  group_prior_var = NULL,
  L = 5,
  minvar = 2,
  mingene = 1,
  min_nonSNP_PIP = 0.5,
  max_iter = 100,
  ncore = 1,
  logfile = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{region_data}{a list object indexing regions, variants and genes.}

\item{region_info}{a data frame of region definition and associated LD file names}

\item{weights}{a list of weights}

\item{group_prior}{a vector of two prior inclusion probabilities for SNPs and genes.}

\item{group_prior_var}{a vector of two prior variances for SNPs and gene effects.}

\item{L}{the number of effects for susie}

\item{mingene}{minimum number of genes in a region}

\item{min_nonSNP_PIP}{Regions with non-SNP PIP >= \code{min_nonSNP_PIP}
will be selected to run finemapping using full SNPs.}

\item{max_iter}{Maximum number of IBSS iterations to perform.}

\item{ncore}{The number of cores used to parallelize susie over regions}

\item{logfile}{the log file, if NULL will print log info on screen}
}
\value{
a data frame of screened region ids and non-SNP PIPs
}
\description{
Screen regions
}
