% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_screen_regions.R
\name{screen_regions}
\alias{screen_regions}
\title{Screens regions with strong non-SNP PIPs to be used in finemapping step}
\usage{
screen_regions(
  region_data,
  use_LD = TRUE,
  LD_info = NULL,
  snp_info = NULL,
  weights = NULL,
  group_prior = NULL,
  group_prior_var = NULL,
  L = 5,
  minvar = 2,
  mingene = 1,
  filter_L = TRUE,
  filter_nonSNP_PIP = FALSE,
  min_nonSNP_PIP = 0.5,
  LD_format = c("rds", "rdata", "mtx", "csv", "txt", "custom"),
  LD_loader_fun,
  ncore = 1,
  logfile = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{region_data}{a list object indexing regions, variants and genes.}

\item{use_LD}{TRUE/FALSE. If TRUE, use LD for finemapping. Otherwise, use "no-LD" version.}

\item{LD_info}{a list of paths to LD matrices for each of the regions. Required when \code{use_LD = TRUE}.}

\item{snp_info}{a list of SNP info data frames for LD reference. Required when \code{use_LD = TRUE}.}

\item{weights}{a list of weights for each gene. Required when \code{use_LD = TRUE}.}

\item{group_prior}{a vector of two prior inclusion probabilities for SNPs and genes.}

\item{group_prior_var}{a vector of two prior variances for SNPs and gene effects.}

\item{L}{the number of effects for susie}

\item{minvar}{minimum number of variables (snps and genes) in a region}

\item{mingene}{minimum number of genes in a region}

\item{filter_L}{If TRUE, screening regions with L >= 1}

\item{filter_nonSNP_PIP}{If TRUE, screening regions with total non-SNP PIP >= \code{min_nonSNP_PIP}}

\item{min_nonSNP_PIP}{If screening by non-SNP PIPs,
regions with total non-SNP PIP >= \code{min_nonSNP_PIP}
will be selected to run finemapping using full SNPs.}

\item{LD_format}{file format for LD matrix. If "custom", use a user defined
\code{LD_loader_fun()} function to load LD matrix.}

\item{LD_loader_fun}{a user defined function to load LD matrix when \code{LD_format = "custom"}.}

\item{ncore}{The number of cores used to parallelize susie over regions}

\item{logfile}{the log file, if NULL will print log info on screen}

\item{verbose}{TRUE/FALSE. If TRUE, print detail messages}

\item{...}{Additional arguments of \code{susie_rss}.}
}
\value{
a list of screened_region_data, and
non-SNP PIPs or estimated L for each region
}
\description{
Screens regions with strong non-SNP PIPs to be used in finemapping step
}
