% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_process_regions.R
\name{select_single_effect_regions}
\alias{select_single_effect_regions}
\title{select single effect regions}
\usage{
select_single_effect_regions(
  regionlist,
  z_snp,
  z_gene,
  group_prior,
  p_single_effect = 0.8
)
}
\description{
select single effect regions
}
