% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_summarize_parameters.R
\name{summarize_param}
\alias{summarize_param}
\title{Summarize and plot cTWAS parameter estimates}
\usage{
summarize_param(param, gwas_n = NA, thin = 1, plot = TRUE)
}
\arguments{
\item{param}{a list of cTWAS parameter estimation result}

\item{gwas_n}{the sample size of the GWAS summary statistics}

\item{thin}{the proportion of SNPs used for parameter estimation, as supplied to \code{est_param}}

\item{plot}{if TRUE, return a plot of the estimated parameters}
}
\description{
Summarize and plot cTWAS parameter estimates
}
