% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_summarize_parameters.R
\name{summarize_param}
\alias{summarize_param}
\title{Summarizes estimated parameters}
\usage{
summarize_param(
  param,
  gwas_n,
  enrichment_test = c("fisher", "G", "z"),
  alternative = c("greater", "two.sided", "less")
)
}
\arguments{
\item{param}{a list of parameter estimation result from \code{est_param}}

\item{gwas_n}{the sample size of the GWAS summary statistics.}

\item{enrichment_test}{Method to test enrichment,
"G": G-test, "fisher": Fisher's exact test,
"z": p-value computed from z-score of enrichment.}

\item{alternative}{indicates the alternative hypothesis and
must be one of "greater", "two.sided", or "less".
Only used when \code{enrichment_test} is "fisher" or "z".}
}
\value{
a list of summarized parameters
}
\description{
Summarizes estimated parameters
}
