% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_summarize_parameters.R
\name{summarize_param}
\alias{summarize_param}
\title{Summarizes estimated parameters}
\usage{
summarize_param(param, gwas_n)
}
\arguments{
\item{param}{a list of parameter estimation result from \code{est_param}}

\item{gwas_n}{the sample size of the GWAS summary statistics}
}
\value{
a list of summarized parameters
}
\description{
Summarizes estimated parameters
}
