% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_diagnose_LD_mismatch.R
\name{update_finemap_res}
\alias{update_finemap_res}
\title{Update finemapping result}
\usage{
update_finemap_res(finemap_res, new_finemap_res)
}
\arguments{
\item{finemap_res}{a data frame of original finemapping result}

\item{new_finemap_res}{a data frame of new finemapping result}
}
\description{
Update finemapping result
}
