% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctwas_merge_regions.R
\name{update_merged_regions_finemap_res}
\alias{update_merged_regions_finemap_res}
\title{Update finemapping result for merged regions}
\usage{
update_merged_regions_finemap_res(
  finemap_res,
  finemap_merged_regions_res,
  merge_region_id_list
)
}
\arguments{
\item{finemap_res}{a data frame of finemapping result}

\item{finemap_merged_regions_res}{a data frame of finemapping result for merged regions}

\item{merge_region_id_list}{a list with new region IDs and merged region IDs}
}
\description{
Update finemapping result for merged regions
}
